+++ docs/articles/joss/paper.md
@@ -10,7 +10,7 @@ tags:
    
 authors:
 - name: Matthieu Gousseff
-  orcid: 0000-0000-0000-0000
+  orcid: 0000-0002-7106-2677
   equal-contrib: true
   affiliation: 1
 - name: Erwan Bocher
