+++ docs/articles/joss/paper.bib
@@ -1,5 +1,6 @@
 @article{lowe2016energy,
   title={An energy and mortality impact assessment of the urban heat island in the US},
+  doi = {10.1016/j.eiar.2015.10.004},
   author={Lowe, Scott A},
   journal={Environmental impact assessment review},
   volume={56},
@@ -10,6 +11,7 @@
 
 @article{clarke1972some,
   title={Some effects of the urban structure on heat mortality},
+  url = {https://doi.org/10.1016/0013-9351(72)90023-0},
   author={Clarke, John F},
   journal={Environmental research},
   volume={5},
@@ -21,6 +23,7 @@
 
 @inproceedings{malys2012microclimate,
   title={Microclimate and buildings energy consumption: sensitivity analysis of coupling methods},
+  doi = {10.1080/17512549.2015.1043643},
   author={Malys, L and Musy, M and Inard, C},
   booktitle={Eighth International Conference on Urban Climate and 10th Symposium on the Urban Environment},
   year={2012}
@@ -28,6 +31,7 @@
 
 @article{stewart2012local,
   title={Local climate zones for urban temperature studies},
+  doi = {10.1175/bams-d-11-00019.1},
   author={Stewart, Ian D and Oke, Tim R},
   journal={Bulletin of the American Meteorological Society},
   volume={93},
@@ -39,6 +43,7 @@
 
 @article{quan2021systematic,
   title={A systematic review of GIS-based local climate zone mapping studies},
+  doi = {10.1016/j.buildenv.2021.107791},
   author={Quan, Steven Jige and Bansal, Parth},
   journal={Building and Environment},
   volume={196},
@@ -50,6 +55,7 @@
 
 @article{bocher2021geoclimate,
   title={GeoClimate: a Geospatial processing toolbox for environmental and climate studies},
+  url = {https://doi.org/10.21105/joss.03541},
   author={Bocher, Erwan and Bernard, J{\'e}r{\'e}my and Wiederhold, Elisabeth Le Saux and Leconte, Fran{\c{c}}ois and Petit, Gwendall and Palominos, Sylvain and No{\^u}s, Camille},
   journal={Journal of Open Source Software},
   volume={6},
@@ -60,6 +66,7 @@
 
 @article{muhammad2022inference,
   title={Inference of Local Climate Zones from GIS Data, and Comparison to WUDAPT Classification and Custom-Fit Clusters},
+  doi = {10.3390/land11050747},
   author={Muhammad, Fadel and Xie, Changkun and Vogel, Julian and Afshari, Afshin},
   journal={Land},
   volume={11},
@@ -71,6 +78,7 @@
 
 @article{kotharkar2018evaluating,
   title={Evaluating urban heat island in the critical local climate zones of an Indian city},
+    url = {https://doi.org/10.1016/j.landurbplan.2017.08.009},
   author={Kotharkar, Rajashree and Bagade, Anurag},
   journal={Landscape and Urban Planning},
   volume={169},
@@ -81,6 +89,7 @@
 
 @article{grigg1965logic,
   title={The logic of regional systems},
+  doi = {10.1111/j.1467-8306.1965.tb00529.x},
   author={Grigg, David},
   journal={Annals of the Association of American Geographers},
   volume={55},
