### 2023-04-05 #5331 @MGousseff
I think it is fixed. By the way, I may have missed something about licencing. The licence.md of the package specifies LGPL-3, which seems OK according to OSI : https://opensource.org/licenses/, but the article says CC by 4.0. I'm still looking for where I need to specify this.
