### 2023-04-13 #5331 @MGousseff
I think Matthias Demuzere matthias.demuzere@ruhr-uni-bochum.de could be interested : he is from the WUDAPT team, and GeoClimate and Wudapt are two very complementary workflows : he might even suggest I better document the raster part ! Do you need me to suggest more than one ? Maybe Mike Spencer (mikerspencer) could be interested too ?
