### 2023-05-05 #5445 @editorialbot
[REVIEW]: lczexplore : an R package to explore Local Climate Zone classifications
**Submitting author:** @MGousseff (<a href="http://orcid.org/0000-0002-7106-2677">Matthieu Gousseff</a>)
**Repository:** https://github.com/orbisgis/lczexplore
**Branch with paper.md** (empty if default branch): master
**Version:** 0.0.1.0003
