### 2023-05-05 6276a9f
Fixed DOIs that were urls
