+++ docs/articles/joss/paper.bib
@@ -11,7 +11,7 @@
 
 @article{clarke1972some,
   title={Some effects of the urban structure on heat mortality},
-  url = {https://doi.org/10.1016/0013-9351(72)90023-0},
+  doi = {10.1016/0013-9351(72)90023-0},
   author={Clarke, John F},
   journal={Environmental research},
   volume={5},
@@ -55,7 +55,7 @@
 
 @article{bocher2021geoclimate,
   title={GeoClimate: a Geospatial processing toolbox for environmental and climate studies},
-  url = {https://doi.org/10.21105/joss.03541},
+  doi = {10.21105/joss.03541},
   author={Bocher, Erwan and Bernard, J{\'e}r{\'e}my and Wiederhold, Elisabeth Le Saux and Leconte, Fran{\c{c}}ois and Petit, Gwendall and Palominos, Sylvain and No{\^u}s, Camille},
   journal={Journal of Open Source Software},
   volume={6},
@@ -78,7 +78,7 @@
 
 @article{kotharkar2018evaluating,
   title={Evaluating urban heat island in the critical local climate zones of an Indian city},
-    url = {https://doi.org/10.1016/j.landurbplan.2017.08.009},
+  doi = {10.1016/j.landurbplan.2017.08.009},
   author={Kotharkar, Rajashree and Bagade, Anurag},
   journal={Landscape and Urban Planning},
   volume={169},
