### 2023-05-23 #5445 @MGousseff
Hello @matthiasdemuzere , thank you for getting involved. I followed the recommandations of Hadley Wickham in  https://r-pkgs.org/ but for common licence, the licence file is even ignored (specified in the .Rbuildignor) as CRAN considers it as redudant with the licence specified in the DESCRIPTION file. So if a copy in full text format is need, I think I can add it and add it's path in the .Rbuildignore, just let me know if it is needed.
