### 2023-05-24 #5445 @matthiasdemuzere
> * **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

The lczexplore R package presents a tool “`to compare different LCZ classifications`”, or, more general, “`any type of classifications on geographical units`”.

Personally, I am not convinced that the current state of the package represents a substantial scholarly effort. I assume the code has been developed in the context of [Bernard et al. (2023)](https://egusphere.copernicus.org/preprints/2023/egusphere-2023-371/egusphere-2023-371.pdf), a paper that is currently under review? Serving a specific purpose within this paper? Data from this paper is therefore also used as a sample dataset within this package.

But, I am not convinced this package will be useful for a broader audience, who might have very different use cases or types of spatial classifications? Other reasons for this concern::

**Need**: I’ve been working with LCZ maps for many years now. Yet I am not convinced the community is really interested in a tool that can compare two LCZ maps? Once you know about the difference / agreements between two maps, what then? How will you use this information? Without a proper reference (ground truth), I don’t really see what are the next steps one can do to put this to use? Bottom line: I am not convinced by the statement of need.

**State of the field**: I am a bit surprised by the shallow description here? The interest in LCZs is immense, yet this is not reflected here? Eg. the [LCZ generator](https://lcz-generator.rub.de/) ([Demuzere et al., 2021](https://doi.org/10.3389/fenvs.2021.637455)) that has received 5000+ LCZ submission in the past two years, the global LCZ map ([Demuzere et al., 2022](https://essd.copernicus.org/articles/14/3835/2022/essd-14-3835-2022.html)), or the many LCZ review papers, with [Huang et al. (2023)](https://www.sciencedirect.com/science/article/abs/pii/S0034425723001244?via%3Dihub) likely being the most recent and comprehensive one?

**Representation**: From a LCZ content point of view, I wonder whether it really makes sense to compare maps that are developed from earth observation information and VGI/GIS layers? Typically their spatial units are very different, one being more representative for the coarser neighborhood-scale (as intended by [Stewart and Oke (2012)](https://journals.ametsoc.org/view/journals/bams/93/12/bams-d-11-00019.1.xml)), and one more on the block level? Please note that this comment might be more about semantics, and not a key reason for me to doubt the applicability of the package as such.

**Coding standards / tests**: the code seems relatively young, with most commits in the past 2 months only. As far as I can see, it has also not been tested outside the scope of the Bernard et al. (2023) paper, using e.g. different types of data? Even testing with raster LCZ maps from the most widely used sources such as the LCZ Generator or other continental or global maps seems limited?
In any case, it does not seem very straightforward to me to do so, as exemplified by the Main Functions description in the README.md: _The following functions are the core of this package : importLCZgen : imports the LCZ layer from a GIS (tested with geojson and shapefile files) importLCZwudapt : imports LCZ from the wudapt Europe Tiff. You'll have to use importLCZgen first to create the Bounding box of your zone of interest showLCZ : plots the map of your LCZ compareLCZ : compares two LCZ classifications of the same areas, output plots and data if this comparison confidSensib : explores how the agreement between two LCZ varies according to a confidence indicator associated for the LCZ value of each geom (sensibility analysis)_. Somehow the functionality seems to be there, but rather inaccessible, and definitely not in a shape I would expect from a package? To add, I eg. also do not see any tests to verify the functionality of this software?

In summary, I support the fact that the authors want to make this code publicly available. Yet in its current state, it feels more like a personal R library serving a specific goal than a tool that will be sufficiently useful for a broader community and more general applications? As such, I don’t think it is ready to be published in JOSS @martinfleis
