### 2023-06-01 #5445 @matthiasdemuzere
@MGousseff Thanks for providing additional information regarding my concerns.

Unfortunately I am still not convinced by the statement of need.

There are nowadays hundreds of papers using LCZs for a certain purpose. Yet, typically, people are not interested in comparing one LCZ map to another (typically even not in a proper accuracy assessment of one map). They just use the LCZ map they (or someone else) created (using whatever algorithm / input data) in any application of interest.

There are some that do an intercomparison, like the [Muhammad et al (2022)](https://www.mdpi.com/2073-445X/11/5/747) paper you mention in your manuscript. In which they conclude that _the GIS-LCZ based method shows a strong improvement over the previous WUDAPT L0 result_, based on accuracy metrics (confusion matrix). We've checked this map in detail with local experts, and believe it is missing important features, eg. large LCZ 8 zones. So even though accuracy metrics can be higher, that does not necessarily mean that map is better.

Continuing with this example, I could use your `lczexplore` tool to visualize differences between the Muhammad et al. (2022) LCZ map and e.g. the LCZ map from [Fenner et al. (2017)](http://www.schweizerbart.de/papers/metz/detail/prepub/88058/Intra_and_inter_local_climate_zone_variability_of_?af=crossref). But then what? You mention the identification of how different input datasets treat vegetation, building heights, ... differently. Ok. Good. But again, what then?

So basically I am still looking for a more clear statement of need that outlines the potential of this tool to a broad community. How can it contribute to an improved understanding of the uncertainties, strengths and weakness of the LCZ framework / method(s)? Identifying differences between maps is one thing, but the crucial step is what comes after that: what to do with this information?

I hope I am not sounding to cruel here. I just genuinly would like to understand how this will benefit the community, in a broader sense and a context of continuesly advancing the field of LCZ-related work.
