### 2023-06-01 #5445 @matthiasdemuzere
> plus also a new one describing how to obtain the comparison of European WUDAPT tiff LCZ map to a GeoClimate OSM output.

This would be much appreciated. I started using the tool, but I stopped as it was not 100% clear to me how to do so with a raster file. There was some very high level info there, but in my opinion that required too much time investment. Given the selling position this is a **tool**, it should in my opinion be much more automated and self-explanatory?

On the side: how does the tool deal with different labels? LCZ labels can very widely, example for the natural classes 11-17 or A - G or 101-107, ...
More general: a large proportion of the LCZ Generator and [W2W code](https://github.com/matthiasdemuzere/w2w/tree/main) is dedicated to checking inputs, to make sure they are in line with expected formats. Does lczexplore has something similar in place?
