### 2023-06-05 21c1b11
Added detection of standard LCZ levels, and setting standard colors to these standard levels Added percentages of area by LCZ in the showLCZ function.
