+++ NAMESPACE
@@ -13,6 +13,7 @@ export(matConfLCZ)
 export(matConfLCZGlob)
 export(produceAnalysis)
 export(showLCZ)
+export(standLevCol)
 import(RColorBrewer)
 import(cowplot)
 import(dplyr)

[…]
