### 2023-06-06 c4b7dfd
Wudapt / GeoClimate OSM vignette
