+++ man/showLCZ.Rd
@@ -39,7 +39,7 @@ random colors will be chosen.}
 \item{LCZlevels}{allows you to set the grouped LCZ types.
 The values must at least cover the values of the column in the dataset, or it will be ignored.}
 
-\item{useStandCol, }{when set to TRUE, any levels detected as a standard LCZ level will receive the standard associated color}
+\item{useStandCol}{is set to TRUE implies that any levels detected as a standard LCZ level will receive the standard associated color}
 }
 \value{
 no object is returned, but plots of the LCZ levels are produced
