+++ vignettes/lczexplore_raster_vector.Rmd
@@ -1,12 +1,12 @@
 ---
-title: "Comparing a raster map and a vector GIS layer"
+title: "Raster_vector_comparison"
 author: "Matthieu Gousseff"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 resource_files:
   - vignettes/Redon.png
 vignette: >
- %\VignetteIndexEntry{lczexplore_english}
+ %\VignetteIndexEntry{Raster_vector_comparison}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
 ---
@@ -46,7 +46,7 @@ As this method is one of the most used we included the map tif in the embedded d
 
 ## Load a territory from a raster map and a bounding box
 
-The first step is to define a bounding box of the territory of interest. The lczexplore comes with vector data for the Redon Territory, a rural town of Britanny, France.
+The first step is to define a bounding box of the territory of interest. The lczexplore comes with vector data for the Redon Territory, a rural town of Brittany, France.
 The importLCZgen is a generic import function for vector data, and when the `ouput` argument is set to `bBox`, it returns the bounding box of the input layer.
 Any bounding box of the bounding box class of the package `sf` can be used to crop the raster map.
 
@@ -65,6 +65,7 @@ fileName="redonWudapt.tif",bBox=redonBbox)
 The resulting object is an sf file, the pixels of the geotiff have been transformed into geometries. The input file must be a geotiff, whose CRS will be read by the function.
 
 ```{R}
+library(sf)
 st_crs(redonWudapt)
 ```
 
@@ -115,7 +116,7 @@ summary(redonOSM$LCZ_PRIMARY)
 
 ```
 ## Visualize the resulting LCZ types with the function showLCZ
-The resulting object is an sf object and can therefore also be visualize with the showLCZ function.
+The resulting object is an sf object and can therefore also be visualized with the showLCZ function.
 ```{R}
 showLCZ(sf = redonOSM, repr="standard")
 ```
@@ -129,11 +130,45 @@ If the two datasets are not in the same coordinate reference system (CRS), the f
 
 ```{R}
 compareLCZ(sf1 = redonWudapt, column1 = "EU_LCZ_map", wf1="WUDAPT",
-           sf2 = redonOSM, column2 = "LCZ_PRIMARY", wf2 = "GeoClimate OSM",
-           ref=1 )
+           sf2 = redonOSM, column2 = "LCZ_PRIMARY", wf2 = "GeoClimate_OSM",
+           ref=1 ,exwrite=FALSE)
 ```
 
 
-# Grouping some levels and comparing the grouped
+# Grouping some levels and comparing the grouped categories
+The intent of this section is not to discuss the quality of the algorithms used for example, but to illustrate how the package
+allow to compare grouped categories.
+For instance, 51% of the areas of Redon that wudapt classifies as compact low are seen as compact mid by Geoclimate
+on OpenStreet map input data. And 34% of what Wudapt classifies as open low are seen as sparsely built by GeoClimate.
+
+But how do Wudapt and GeoClimate agree on broader categories, as built LCZ versus vegetation LCZ ? One could group these LCZs
+to explore this, using the `LCZgroup2` function.
+```{R}
+# Regroup levels of the WUDAPT sf file.
+redonWudaptGrouped<-LCZgroup2(redonWudapt,column="EU_LCZ_map",
+                              urban=c("1","2","3","4","5","6","7","8","9"),
+                              industry="10", vegetation=c("101","102","103","104"),
+                              impervious="105",pervious="106",water="107", outCol = "Regrouped")
+
+# Regroup levels of the GeoClimate sf file
+redonOSMgrouped<-LCZgroup2(redonOSM, column="LCZ_PRIMARY",
[…]
