+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(LCZareas)
 export(LCZgroup2)
 export(areColors)
 export(compareLCZ)
