+++ man/standLevCol.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/standLevCol.R
+\name{standLevCol}
+\alias{standLevCol}
+\title{Detects standard LCZ levels and associate standard colors to them}
+\usage{
+standLevCol(levels, colors = "", useStandCol = TRUE)
+}
+\arguments{
+\item{levels}{is the vector of the LCZ levels present in a dataset}
+
+\item{colors}{is the vector of colors one intends to use to visualize the LCZ of the dataset}
+
+\item{useStandCol}{allows to say if standard colors should replace user specified colors
+when standard levels are detected.}
+}
+\value{
+output is a list containing levelColors, a named vector, which names are the levels
+present in the data and which values are the associated colors,
+and case, a string spcifying what case was encountered when producing the levels and colors.
+}
+\description{
+Detects standard LCZ levels and associate standard colors to them
+}
+\examples{
+#standLevCol is not to be used directly by user. It deals with levels and colors provided by the user
+}
