### 2023-06-07 bf6197d
Small change in compareLCZ : you can plot and save the plot at the same time (maybe)
