### 2023-06-07 orbisgis/lczexplore#12 @wcjochem
JOSS Review: scholarly effort
Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

"software should be feature-complete (i.e., no half-baked solutions)" - the README for this package states that it contains experimental which are not generic or robust. Please remove these functions or keep them on a separate, "development" branch and instruct users who wish to use the experimental functions to install the dev version of the package.

[joss-reviews/issues/#5445](https://github.com/openjournals/joss-reviews/issues/5445)
