### 2023-06-07 orbisgis/lczexplore#13 @wcjochem
JOSS Review: community guidelines
Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?

Please edit the README to provide this information.

[openjournals/joss-reviews/issues#5445](https://github.com/openjournals/joss-reviews/issues/5445)
