### 2023-06-07 orbisgis/lczexplore#14 @wcjochem
JOSS Review: quality of writing
Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

Before publication, please do a final review of grammar and writing clarity.

[openjournals/joss-reviews](https://github.com/openjournals/joss-reviews/issues/5445)
