### 2023-06-07 orbisgis/lczexplore#15 @wcjochem
JOSS Review: functionality and example usage
I have not been able to verify the claim made in the paper that the package can be used to compare any two categorical spatial datasets. Please provide a workable example in the paper and software documentation demonstrating the package's more general capabilities.

https://github.com/openjournals/joss-reviews/issues/5445
