### 2023-06-07 orbisgis/lczexplore#16 @wcjochem
JOSS Review: scholarly effort coding standards
1. The package source code is mixing pipe operators. Some statements use `%>%` while others use the native R pipe (`|>`). I don't believe the native R pipe is supported in versions of R before 4.1; however, the package DESCRIPTION states that it only depends on R >2.2. This may be an incompatibility. Has the package been tested on older versions of R?

2. Please refer to: https://ggplot2.tidyverse.org/articles/ggplot2-in-packages.html#using-aes-and-vars-in-a-package-function for guidance on how to use `ggplot2` within a package.

3.  As a new user, I found the interface to the functions difficult to understand with so many different arguments, and complex return objects. The examples could show how to extract and work with the returned values a bit more.

https://github.com/openjournals/joss-reviews/issues/5445
