### 2023-06-07 #5445 @wcjochem
@MGousseff  and @martinfleis, I've raised several issues in the repo (and linked in this thread) for your consideration and have completed my review at this stage. I can see that the package does support its primary aim, although it seems to be a very narrow application. I can't verify the claim in the paper that the package's functionality would be useful more generally in other types of spatial comparisons. I'm happy to continue my review if these comments are addressed.
