### 2023-06-14 a0020e2
Supress zero area intersected geometries
