### 2023-06-14 #5445 @MGousseff
@wcjochem Thank you very much for all  your issues.

1. About the use of **different pipe** operators : you are perfectly right, using R-base pipe is less robust than the tidyverse pipe as it is a hard constraint on R version : **I will fix this.**
2. About the use of ggplot, as you noticed, some of the notes I still have are linked to this, and I will correct this as soon as possible. I think some of these are due to non-standard evaluation in othe packages of the tidyverse, I will find them or use the  globals.R roundabout.
3. About the need to show how to extract and work with the returned values a bit more : **do you think this should be done in the example section or is it better to enrich the vignettes ?**
4. About the need to illustrate how  the package can be used to compare any two categorical spatial datasets, I wrote a vignette illustrating a case. But I now feel it is better to create a dedicated function to import categorical variable other than LCZ (in order to create explicit error message, as, for instance, the need for the geometries to be polygons or multi polygons), it will be ready shortly.

I really thank you and @matthiasdemuzere  and I will deliver as soon as possible a new version with all these issues fixed.
