### 2023-06-14 617446e
Supress zero area intersected geometries in compare LCZ
