+++ R/compareLCZ.R
@@ -62,7 +62,7 @@ compareLCZ<-function(sf1,geomID1="",column1,confid1="",wf1="bdtopo_2_2",
 
   # store the column names in a way that can be injected in functions A SUPPRIMER ?
 
-  column2Init<-column2 # in case of column1==column2, this will be used to call levCol
+  # column2Init<-column2 in case of column1==column2, this will be used to call levCol
   namesf1<-deparse(substitute(sf1))
   namesf2<-deparse(substitute(sf2))
   message(paste(" The column ",column1, " of the dataset", namesf1,
@@ -236,7 +236,7 @@ compareLCZ<-function(sf1,geomID1="",column1,confid1="",wf1="bdtopo_2_2",
 
   if(repr=="grouped")
   { ############### This is a temporary feature. Grouping LCZ, showing and comparing grouped LCZ will be re-written in a  cleaner way in a later version
-    args<-list(...)
+    # args<-list(...)
 
     # Call levCol to deal with levels and colors
     levCol1<-levCol(sf1,column1,...)
