+++ inst/tinytest/test_compareLCZ.R
@@ -10,7 +10,7 @@
 expect_message(compareRedonBDTOSM<-
                  compareLCZ(sf1=redonBDT, column1="LCZ_PRIMARY", geomID1 = "ID_RSU", confid1="LCZ_UNIQUENESS_VALUE", wf1="bdtopo_2_2",
                             sf2=redonOSM, column2="LCZ_PRIMARY", geomID2 = "ID_RSU", confid2="LCZ_UNIQUENESS_VALUE", wf2="osm",
-                            repr="standard", saveG="", exwrite=TRUE, location="Redon", plot=TRUE),
+                            repr="standard", saveG="", exwrite=FALSE, location="Redon", plot=TRUE),
           "Both sf datasets need to live in the same crs projection \\(srid / epsg\\),")
 file.remove("bdtopo_2_2_osm.csv")
 
@@ -164,4 +164,3 @@ expect_warning(compareRedonBDTOSMgrouped<-
                             impervious="105",pervious="106",water="107",
                             cols=c("red","black","green","grey","burlywood","blue"),tryGroup = TRUE),
                "attribute variables are assumed to be spatially constant throughout all geometries")
-
