+++ inst/tinytest/test_importSurfQualVar.R
@@ -1,0 +1,103 @@
+#This tests the function importSurfQualVar
+# library(tinytest)
+#
+library(sf)
+# test<-st_read(
+#   "/home/gousseff/Documents/2_CodesSources/R/lczexplore/lczexplore/inst/extdata/bdtopo_2_2/Redon/rsu_lcz.geojson")
+# colonnes<-c("LCZ_PRIMARY","ID_RSU","LCZ_UNIQUENESS_VALU")
+# tetest<-try(test[colonnes]) %>% class
+# "try-error"%in%tetest
+
+
+
+
+expect_silent(
+  utrfRedonBDT<-importSurfQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"), 
+                                                         file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
+)
+# 
+# expect_silent(importLCZgen(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon")
+#                            ,file="rsu_lcz.geojson",
+#                            column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"))
+# 
+# 
+# # Tests if the imported version of Redon test data matches the Redon test data in the package
+# redonBDT2<-importLCZgen(dirPath=paste0(
+#   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),
+#   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE")
+# st_crs(redonBDT2)$wkt<-gsub("é","e",st_crs(redonBDT2)$wkt)
+# 
+# expect_identical(redonBDT,redonBDT2)
+# 
+# # Error if file doesn't exist in location
+# 
+# expect_error(importLCZgen(dirPath=paste0(
+#   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="chaussure.geojson",
+#   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),"The file doesn't seem to exist.")
+# 
+# ## Error if directory doesn't exist
+# 
+# 
+# expect_error(importLCZgen(dirPath=paste0(
+#   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redono"),file="rsu_lcz.geojson",
+#   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),
+#   "The directory set in dirPath doesn't seem to exist")
+# 
+# # test if one column doesn't exist
+# 
+# expect_error(
+# importLCZgen(dirPath=paste0(
+#   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="rsu_lcz.geojson",
+#   column="LCZ_PRIMAR",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),
+#   "It seems that some of the columns you try to import do not exist in the source file")
+# 
+# expect_error(
+#   importLCZgen(dirPath=paste0(
+#     system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="rsu_lcz.geojson",
+#     column="LCZ_PRIMARY",geomID="ID_RSY",confid="LCZ_UNIQUENESS_VALUE"),
+#   "It seems that some of the columns you try to import do not exist in the source file")
+# 
+# expect_error(
+#   importLCZgen(dirPath=paste0(
[…]
