### 2023-06-23 cee6365
importSurfQualVar works in all package : repr="grouped" replaced with repr="alter", but column="grouped" is still the default ouput for LCZgroup2 Vignettes are installed
