+++ R/compareLCZ.R
@@ -23,8 +23,8 @@
 #' use "bdtopo_2_2". When GeoClimate was used with Open Street Map data as input, use "osm".
 #' When the LCZ come from the wudapt Europe tiff, use "wudapt".
 #' @param ref : If the coordinate reference system (CRS) of sf1 and sf2 differ, ref indicates which CRS to choose for both files (1 or 2)
-#' @param repr "standard" means that original values of LCZ are used,
-#'  "grouped" means the user has groupe some of the LCZ levels under a new label.
+#' @param repr "standard" means that standard values of LCZ are expected,
+#'  "alter" means other values are expected, like grouped values of LCZ or other qualitative variable.
 #'  In the latter case, the ... arguments must contain the groups and a color vector.
 #' @param plot : when FALSE non of the graphics are plotted or saved
 #' @param saveG : when an empty character string, "", the plots are not saved. Else, the saveG string is used to produce the name of the saved png file.
@@ -36,9 +36,9 @@
 #' @param tryGroup : when TRUE, if the specified level names don't match the data, but the specified levels do,
 #' a call to the LCZgroup2 function will be tried, and if it works, the resulting grouping column will be named
 #' "grouped" and the comparison will be done using it.
-#' @param ... allow to pass arguments if representation is grouped.
-#' The expected arguments are the name of each grouped label,
-#' the levels of LCZ they contain, and last a vector of the colors to use to plot them.
+#' @param ... allow to pass arguments if repr is set to alter.
+#' The expected arguments are the name of each level of the variables contained 
+#' in column1 and column2, and last a vector cols of the colors to use to plot them.
 #' @importFrom ggplot2 geom_sf guides ggtitle aes
 #' @import sf dplyr cowplot forcats units tidyr RColorBrewer utils grDevices rlang
 #' @return returns an object called matConfOut which contains
@@ -50,14 +50,27 @@
 #' If saveG is not an empty string, graphics are saved under "saveG.png"
 #' @export
 #' @examples
-#' compareLCZ(sf1=redonBDT, column1="LCZ_PRIMARY", geomID1 = "ID_RSU",
+#' comparisonBDT_OSM<-compareLCZ(sf1=redonBDT, column1="LCZ_PRIMARY", geomID1 = "ID_RSU",
 #' confid1="LCZ_UNIQUENESS_VALUE", wf1="bdtopo_2_2",
 #' sf2=redonOSM, column2="LCZ_PRIMARY", geomID2 = "ID_RSU",
 #' confid2="LCZ_UNIQUENESS_VALUE", wf2="osm",
 #' repr="standard", saveG="", exwrite=TRUE, location="Redon", plot=TRUE)
+#' # To get the summed area of each LCZ levels for both dataset : 
+#' comparisonBDT_OSM$areas
+#' # The plots of each dataset can be produced with the /`showLCZ/` function.
+#' # To get the value of the general percentage of agreement call :
+#' comparisonBDT_OSM$percAgg
+#' # The values of that confusion matrix is available in a wide form:
+#' comparisonBDT_OSM$matConfLarge
+#' # or in a long form :
+#' comparisonBDT_OSM$matConf
+#' 
+#' # Examples for non LCZ variables are available in the importSurfQualVar function examples.
+#' 
 compareLCZ<-function(sf1,geomID1="",column1,confid1="",wf1="bdtopo_2_2",
                      sf2,column2,geomID2="",confid2="",wf2="osm",ref=1,
[…]
