+++ inst/tinytest/test_importQualVar.R
@@ -1,4 +1,4 @@
-#This tests the function importSurfQualVar
+#This tests the function importQualVar
 # library(tinytest)
 #
 library(sf)
@@ -9,12 +9,12 @@ library(dplyr)
 # tetest<-try(test[colonnes]) %>% class
 # "try-error"%in%tetest
 
-utrfRedonBDT<-importSurfQualVar(dirPath=paste0(
+utrfRedonBDT<-importQualVar(dirPath=paste0(
   system.file("extdata", package = "lczexplore"), "/bdtopo_2_2/Redon"),
   file="rsu_utrf_area.geojson", column="TYPO_MAJ")
   
 showLCZ(sf=utrfRedonBDT, column="TYPO_MAJ",repr="alter")
-  utrfRedonOSM<-importSurfQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
+  utrfRedonOSM<-importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
   file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
   
   utrfComparison<-compareLCZ(sf1=utrfRedonBDT, column1="TYPO_MAJ", sf2=utrfRedonOSM, column2="TYPO_MAJ",wf1=" UTRF BDT", wf2 = " UTRF OSM",
@@ -26,14 +26,14 @@ showLCZ(sf=utrfRedonBDT, column="TYPO_MAJ",repr="alter")
 
 
 expect_silent(
-  utrfRedonBDT<-importSurfQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"), 
+  utrfRedonBDT<-importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"), 
                                                          file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
 )
 
 showLCZ(utrfRedonBDT,column = "TYPO_MAJ",repr="alter")
 
 expect_silent(
-  utrfRedonOSM<-importSurfQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
+  utrfRedonOSM<-importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
                                   file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
 )
 
