+++ NAMESPACE
@@ -9,7 +9,7 @@ export(fetchLCZ)
 export(importLCZgc)
 export(importLCZgen)
 export(importLCZraster)
-export(importSurfQualVar)
+export(importQualVar)
 export(levCol)
 export(matConfLCZ)
 export(matConfLCZGlob)
