### 2023-06-27 52ec2da
Levels and colors fixed in show LCZ. To do : improve vignettes, rewrite article
