+++ inst/tinytest/test_showLCZ.R
@@ -5,15 +5,20 @@
 #showLCZ<-function(sf, title="", wf="",column="LCZ_PRIMARY", repr="standard", typeLevels="", cols="")
 
 expect_silent(
-  showLCZ(redonBDT, title="Zones climatiques locales à Redon",repr="standard",drop=TRUE)
+  showLCZ(redonOSM, title="Zones climatiques locales à Redon",
+          repr="standard", drop=FALSE, cols="", LCZlevels="")
 )
-expect_silent(showLCZ(redonBDT))
+expect_silent(showLCZ(redonBDT, drop=TRUE))
 
 testCol <- palette.colors(n=17, palette="Polychrome 36")
 showLCZ(redonBDT, title="Zones climatiques locales à Redon",repr="alter",
         useStandCol=FALSE,
         cols = testCol )
-levCol(sf=redonBDT, column="LCZ_PRIMARY",cols = testCol)
+
+showLCZ(redonBDT, title="Zones climatiques locales à Redon",repr="alter",
+        useStandCol=TRUE,
+        cols = testCol )
+#levCol(sf=redonBDT, column="LCZ_PRIMARY",cols = testCol)
 
 redonBDTgrouped<-LCZgroup2(redonBDT,column="LCZ_PRIMARY", urban=c("1","2","3","4","5","6","7","8","9"),
                            industry="10",
@@ -45,6 +50,10 @@ showLCZ(redonBDTgrouped,column="grouped",repr="alter", levels=c(
   "pervious" = "burlywood","water"="blue")),
 "4:"
 )
+
+levCol(sf=redonBDTgrouped,column = "grouped",levels=c(
+  "urban" = "red", "industry" = "black","vegetation" = "green","impervious" = "grey",
+  "pervious" = "burlywood","water"="blue"))
 
 expect_message(
   showLCZ(redonBDTgrouped,column="grouped",repr="alter", levels=c(
