### 2023-06-27 43fa803
Levels and colors fixed in show LCZ. english and french general vignettes OK Raster vignette simple but OK
To do : check alter vignette and rewrite JOSS article
