+++ vignettes/lczexplore_raster_vector.Rmd
@@ -94,9 +94,6 @@ showLCZ(sf = redonWudapt,column = 'EU_LCZ_map', repr= "standard")
 ```
 
 
-
-
-
 The following code shows an example of color choice.
 
 # Import and visualize a vector layer map of LCZ
@@ -122,6 +119,7 @@ showLCZ(sf = redonOSM, repr="standard")
 ```
 
 # Comparison of the two LCZ classification
+
 **The purpose of this package is to easily compare local climate zone classifications produced by different algorithms on the same study area.**
 
 The `compareLCZ` function intersects the geometries of both input sf files. The resulting geometries then either agree or disagree.
@@ -136,9 +134,11 @@ compareLCZ(sf1 = redonWudapt, column1 = "EU_LCZ_map", wf1="WUDAPT",
 
 
 # Grouping some levels and comparing the grouped categories
+
 The intent of this section is not to discuss the quality of the algorithms used for example, but to illustrate how the package
 allow to compare grouped categories.
-For instance, 51% of the areas of Redon that wudapt classifies as compact low are seen as compact mid by Geoclimate
+
+For instance, 51% of the areas of Redon that Wudapt classifies as compact low are seen as compact mid by Geoclimate
 on OpenStreet map input data. And 34% of what Wudapt classifies as open low are seen as sparsely built by GeoClimate.
 
 But how do Wudapt and GeoClimate agree on broader categories, as built LCZ versus vegetation LCZ ? One could group these LCZs
@@ -169,3 +169,5 @@ One can notice that the two explored algorithms agree rather well about vegetati
 On urban LCZ, while 68% of the surface WUDAPT classifies as urban is also classified as urban by GeoClimate, 19% is classified as vegetation and 12% as impervious by GeoClimate.
 
 The confusion matrix compare percentage of areas, so if the global area concerned is small, one has to be careful about the interpretation. For instance, the very strong agreement about pervious soils or the relative agreement about water should not be over interpreted as they only represent less than 1% of the areas of the WUDAPT generated map.
+
+For more details on the package, browse other vignettes. 

[…]
