### 2023-06-30 a6bd7cf
PseudoKappa added To do : enrich alter vignette and update JOSS article
