+++ docs/articles/joss/paper.md
@@ -9,6 +9,7 @@ tags:
 - cities
    
 authors:
+  
 - name: Matthieu Gousseff
   orcid: 0000-0002-7106-2677
   equal-contrib: true
@@ -25,8 +26,8 @@ authors:
   orcid:
   corresponding: true
   affiliation: 3   
-  
 affiliations:
+  
 - name: CNRS, Lab-STICC, UMR 6285, Vannes, France
   index: 1
 - name: University of Savoie Mont-Blanc, LOCIE, UMR 5271, France
@@ -35,55 +36,55 @@ affiliations:
   index: 3
 - name : University of Gothenburg, Department of Earth Sciences, Sweden
   index: 4
-  
 bibliography: paper.bib
 date: 01 april 2023
+editor_options: 
+  markdown: 
+    wrap: sentence
 ---
-
 
 # Summary
 
-Climate change is a growing concern for city planners as Urban
-Heat Islands has an impact on mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects. A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, such as Local Climate Zones (LCZ) defined by [@stewart2012local].
+Climate change is a growing concern for city planners as Urban Heat Islands has an impact on mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
+A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, such as Local Climate Zones (LCZ) defined by [@stewart2012local].
 
 The lczexplore package aims at comparing different LCZ classifications, but could be used to compare any type of classifications on geographical units.
 
-A spatial comparison is performed by producing agreement maps between classifications. A confusion matrix also helps to better qualify and quantify the misclassifications.
+A spatial comparison is performed by producing agreement maps between classifications.
+A confusion matrix also helps to better qualify and quantify the misclassifications.
 
 This software is available as a free and opensource R package.
 
 # Statement of need
 
-Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons. The terms rural and urban lack of a clear definition and several classification of urban and rural landscapes were proposed. 
-Stewart and Oke  [@stewart2012local] proposed an approach that complies with [@grigg1965logic] criteria of geographical classification. 
-Their Local Climate Zones are based on a logical segmentation of land-use and urban canopy parameters and define 10 urban types and 7 land cover types. 
+Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
+The terms rural and urban lack of a clear definition and several classification of urban and rural landscapes were proposed.
+Stewart and Oke [@stewart2012local] proposed an approach that complies with [@grigg1965logic] criteria of geographical classification.
+Their Local Climate Zones are based on a logical segmentation of land-use and urban canopy parameters and define 10 urban types and 7 land cover types.
 It has gained popularity in the past decade as it sums up relevant information and can help, for instance, apprehend the intensity of the Urban Heat Island ([@kotharkar2018evaluating]).
 
 Several papers describe methods to produce LCZ classifications, but there are fewer workflows that automatically produce them from available data.
-[@quan2021systematic] distinguishes two main streams of production of these LCZ:  
-- the raster stream processes remotely sensed information, and applies machine learning algorithms trained using local experts' knowledge. 
-For instance, the WUDAPT platform produced LCZ maps of Europe and North-America this way.  
--  the vector stream uses Geographic Information System (GIS) layers that represent the main topographic features, defines spatial units, 
-computes urban canopy parameters and uses them to classify spatial units into LCZ. 
[…]
