+++ R/compareLCZ.R
@@ -40,6 +40,7 @@
 #' The expected arguments are the name of each level of the variables contained 
 #' in column1 and column2, and last a vector cols of the colors to use to plot them.
 #' @importFrom ggplot2 geom_sf guides ggtitle aes
+#' @importFrom DescTools CohenKappa
 #' @import sf dplyr cowplot forcats units tidyr RColorBrewer utils grDevices rlang
 #' @return returns an object called matConfOut which contains
 #' matConfLong, a confusion matrix in a longer form, which can be written in a file by the compareLCZ function
@@ -381,11 +382,27 @@ matConfOut<-matConfLCZ(sf1=sf1, column1=column1, sf2=sf2, column2=column2,
 matConfOut$data<-echIntExpo
 matConfLong<-matConfOut$matConf
 matConfLarge<-pivot_wider(matConfLong,names_from = column2,values_from = agree)
-matConfOut$matConfLarge<-matConfLarge  
+matConfLarge<-compareRedonBDTOSM$matConfLarge %>% as.data.frame()
+row.names(matConfLarge)<-matConfLarge[,1] %>% as.character
+matConfLarge<-matConfLarge[,-1]
+matConfLarge<-as.matrix(matConfLarge)  
+
+  
+# Add pseudo Kappa Statistic to output to   
+PseudoWeightedCross<-matConfLarge*100
+pseudoK<-DescTools::CohenKappa(x=matConfLarge)  
+matConfOut$pseudoK<-pseudoK
   
 areas<-matConfOut$areas
 percAgg<-matConfOut$percAgg
 
+
+
+
+  
+  
+  
+  
 ################################################
 #  GRAPHICS
 ################################################
