+++ NAMESPACE
@@ -28,6 +28,7 @@ import(sf)
 import(tidyr)
 import(units)
 import(utils)
+importFrom(DescTools,CohenKappa)
 importFrom(forcats,fct_recode)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,geom_sf)
