### 2023-06-30 1757075
No ERROR No WARNING The 2 remaining notes are OK (size and type of example files). First draft of the new version of the JOSS aarticle completed.
TO DO : finish the lczexplore_alter vignette
