### 2023-07-03 c5707ee
New version of JOSS article. TO DO : finish lcz_explore_alter vignette.
