+++ docs/articles/joss/paper.bib
@@ -1,3 +1,45 @@
+@article{visser2006map,
+  title={The map comparison kit},
+  author={Visser, Hans and De Nijs, T},
+  journal={Environmental Modelling \& Software},
+  volume={21},
+  number={3},
+  pages={346--358},
+  year={2006},
+  publisher={Elsevier}
+}
+
+@article{monserudComparingGlobalVegetation1992,
+  title = {Comparing Global Vegetation Maps with the {{Kappa}} Statistic},
+  author = {Monserud, Robert A. and Leemans, Rik},
+  year = {1992},
+  month = aug,
+  journal = {Ecological Modelling},
+  volume = {62},
+  number = {4},
+  pages = {275--293},
+  issn = {03043800},
+  doi = {10.1016/0304-3800(92)90003-W},
+  urldate = {2023-06-30},
+  langid = {english}
+}
+
+@article{hammerbergImplicationsEmployingDetailed2018,
+  title = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling: A Comparison between {{WUDAPT}} and {{GIS}} Databases over {{Vienna}}, {{Austria}}},
+  shorttitle = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling},
+  author = {Hammerberg, Kris and Brousse, Oscar and Martilli, Alberto and Mahdavi, Ardeshir},
+  year = {2018},
+  month = apr,
+  journal = {International Journal of Climatology},
+  volume = {38},
+  pages = {e1241-e1257},
+  issn = {08998418},
+  doi = {10.1002/joc.5447},
+  urldate = {2023-06-30},
+  langid = {english},
+}
+
+
 @article{lowe2016energy,
   title={An energy and mortality impact assessment of the urban heat island in the US},
   doi = {10.1016/j.eiar.2015.10.004},
@@ -98,3 +140,47 @@
   year={1965},
   publisher={Taylor \& Francis}
 }
+
+@article{chingWUDAPTUrbanWeather2018a,
+  title = {{{WUDAPT}}: {{An Urban Weather}}, {{Climate}}, and {{Environmental Modeling Infrastructure}} for the {{Anthropocene}}},
+  shorttitle = {{{WUDAPT}}},
+  author = {Ching, J. and Mills, G. and Bechtel, B. and See, L. and Feddema, J. and Wang, X. and Ren, C. and Brousse, O. and Martilli, A. and Neophytou, M. and Mouzourides, P. and Stewart, I. and Hanna, A. and Ng, E. and Foley, M. and Alexander, P. and Aliaga, D. and Niyogi, D. and Shreevastava, A. and Bhalachandran, P. and Masson, V. and Hidalgo, J. and Fung, J. and Andrade, M. and Baklanov, A. and Dai, W. and Milcinski, G. and Demuzere, M. and Brunsell, N. and Pesaresi, M. and Miao, S. and Mu, Q. and Chen, F. and Theeuwes, N.},
+  year = {2018},
+  month = sep,
+  journal = {Bulletin of the American Meteorological Society},
+  volume = {99},
+  number = {9},
+  pages = {1907--1924},
+  issn = {0003-0007, 1520-0477},
+  doi = {10.1175/BAMS-D-16-0236.1},
+  urldate = {2023-06-30},
+  }
+  
+  @article{cohenCoefficientAgreementNominal1960,
+    title = {A {{Coefficient}} of {{Agreement}} for {{Nominal Scales}}},
+    author = {Cohen, Jacob},
+    year = {1960},
+    month = apr,
+    journal = {Educational and Psychological Measurement},
+    volume = {20},
+    number = {1},
[…]
