+++ docs/articles/joss/paper.md
@@ -45,53 +45,117 @@ editor_options:
 
 # Summary
 
-Climate change is a growing concern for city planners as Urban Heat Islands has an impact on mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
-A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, such as Local Climate Zones (LCZ) defined by [@stewart2012local].
+Climate change is a growing concern for city planners as Urban Heat Islands has an impact on mortality [@clarke1972some], 
+health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
+A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, 
+such as Local Climate Zones (LCZ) defined by [@stewart2012local].
 
-The lczexplore package aims at comparing different LCZ classifications, but could be used to compare any type of classifications on geographical units.
+The lczexplore package aims at comparing different LCZ classifications, 
+but can be used to compare any pair of classifications on geographical units.
 
-A spatial comparison is performed by producing agreement maps between classifications.
-A confusion matrix also helps to better qualify and quantify the misclassifications.
+A spatial comparison is performed by producing agreement maps between classifications, agreement statistics and 
+a confusion matrix to help qualify and quantify the misclassifications.
 
 This software is available as a free and opensource R package.
 
 # Statement of need
 
+## Comparing maps
+As stated in [@visser2006map] comparing map is an important issue in environmental research. 
+The four main reasons to compare categorical variable son geographical units are :     
+- To compare maps generated by models under different scenarios and assumptions  
+- To detect temporal changes  
+- To calibrate or validate models   
+- To perform uncertainty and sensitivity analysis  
+
+
+## Comparing specifically LCZ maps 
+
 Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
-The terms rural and urban lack of a clear definition and several classification of urban and rural landscapes were proposed.
-Stewart and Oke [@stewart2012local] proposed an approach that complies with [@grigg1965logic] criteria of geographical classification.
-Their Local Climate Zones are based on a logical segmentation of land-use and urban canopy parameters and define 10 urban types and 7 land cover types.
-It has gained popularity in the past decade as it sums up relevant information and can help, for instance, apprehend the intensity of the Urban Heat Island ([@kotharkar2018evaluating]).
+The terms rural and urban lack of a clear definition and several different classifications 
+of urban and rural landscapes were proposed.
+Stewart and Oke [@stewart2012local] defined an approach that complies with [@grigg1965logic] criteria of geographical classification.
+Their Local Climate Zones (LCZ) are based on a logical segmentation of land-use and urban canopy parameters 
+and define 10 urban types and 7 land cover types.
 
-Several papers describe methods to produce LCZ classifications, but there are fewer workflows that automatically produce them from available data.
-[@quan2021systematic] distinguishes two main streams of production of these LCZ:\
-- the raster stream processes remotely sensed information, and applies machine learning algorithms trained using local experts' knowledge.
-For instance, the WUDAPT platform produced LCZ maps of Europe and North-America this way.\
-- the vector stream uses Geographic Information System (GIS) layers that represent the main topographic features, defines spatial units, computes urban canopy parameters and uses them to classify spatial units into LCZ.
-For instance, the GeoClimate geospatial toolbox produces LCZ classifications from OpenStreetMap or french BDTopo data [@bocher2021geoclimate].
+LCZ have gained popularity in the past decade as they sum up relevant information and can help, for instance, 
+apprehend the intensity of the Urban Heat Island ([@kotharkar2018evaluating]).
 
[…]
