### 2023-07-03 e2672a8
A better way to use colors and types : no argument pre-set in function. Simpler for the user. TO DO : finish lcz_explore_alter vignette.
