### 2023-07-04 cce6e88
Cols argument was abbreviated (=bad) and ambiguous (= worse), so was replaced by colors.
