+++ inst/tinytest/test_LCZgroup2.R
@@ -13,7 +13,7 @@ redonBDTgrouped<-LCZgroup2(redonBDT,column="LCZ_PRIMARY",urban=c("1","2","3","4"
                            industry="10",
                            vegetation=c("101","102","103","104"),
                            impervious="105",pervious="106",water="107",
-                           cols=c("red","black","green","grey","burlywood","blue")
+                           colors=c("red","black","green","grey","burlywood","blue")
 )
 
 expect_silent(
@@ -21,7 +21,7 @@ expect_silent(
                                           industry="10",
                                           vegetation=c("101","102","103","104"),
                                           impervious="105",pervious="106",water="107",
-                                         cols=c("red","black","green","grey","burlywood","blue"))
+                                         colors=c("red","black","green","grey","burlywood","blue"))
 )
 
 
@@ -29,7 +29,7 @@ expect_silent(redonOSMgrouped<-LCZgroup2(redonOSM,column="LCZ_PRIMARY",urban=c("
                            industry="10",
                            vegetation=c("101","102","103","104"),
                            impervious="105",pervious="106",water="107",
-                           cols=c("red","black","green","grey","burlywood","blue"))
+                           colors=c("red","black","green","grey","burlywood","blue"))
 )
 
 
@@ -38,20 +38,20 @@ expect_silent(redonBDTgrouped2<-LCZgroup2(redonBDT,column="LCZ_PRIMARY", outCol=
                            industry="10",
                            vegetation=c("101","102","103","104"),
                            impervious="105",pervious="106",water="107",
-                           cols=c("red","black","green","grey","burlywood","blue")))
+                           colors=c("red","black","green","grey","burlywood","blue")))
 
 expect_silent(redonOSMgrouped2<-LCZgroup2(redonOSM,column="LCZ_PRIMARY",outCol="otherName",
                            urban=c("1","2","3","4","5","6","7","8","9"),
                            industry="10",
                            vegetation=c("101","102","103","104"),
                            impervious="105",pervious="106",water="107",
-                           cols=c("red","black","green","grey","burlywood","blue"))
+                           colors=c("red","black","green","grey","burlywood","blue"))
 )
 
 expect_message(
   showLCZ(redonBDTgrouped2,column="groupedLCZ",repr="alter",
         LCZlevels=c("urban","industry","vegetation","impervious","pervious","water"),
-        cols=c("red","black","green","grey","burlywood","blue")),
+        colors=c("red","black","green","grey","burlywood","blue")),
                "9: Levels specified in one vector"
 
 )
@@ -59,7 +59,7 @@ expect_message(
 # compareLCZ(sf1=redonBDTgrouped2, column1="groupedLCZ", wf1="BDT",
 #            sf2=redonOSMgrouped2,column2="otherName",wf2="osm", exwrite=FALSE, repr="alter", plot=T,
 #            urban="urban",industry="industry",vegetation="vegetation",impervious="impervious",pervious="pervious",water="water",
-#            cols=c("red","black","green","grey","burlywood","blue") )
+#            colors=c("red","black","green","grey","burlywood","blue") )
 
 #test the non-coverage of levels
 
