### 2023-07-04 80b35a4
Cols argument was abbreviated (=bad) and ambiguous (= worse), so was replaced by colors.
