+++ docs/articles/joss/paper.bib
@@ -184,3 +184,16 @@
     doi = {10.3389/fenvs.2021.637455},
     urldate = {2023-06-30},
     }
+
+@techreport{bernardGenericAlgorithmAutomatically2023,
+  type = {Preprint},
+  title = {A Generic Algorithm to Automatically Classify Urban Fabric According to the {{Local Climate Zone}} System: Implementation in {{GeoClimate}} 0.0.1 and Application to {{French}} Cities},
+  shorttitle = {A Generic Algorithm to Automatically Classify Urban Fabric According to the {{Local Climate Zone}} System},
+  author = {Bernard, J{\'e}r{\'e}my and Bocher, Erwan and Gousseff, Matthieu and Leconte, Fran{\c c}ois and Le Saux Wiederhold, Elisabeth},
+  year = {2023},
+  month = mar,
+  institution = {{Climate and Earth system modeling}},
+  doi = {10.5194/egusphere-2023-371},
+  urldate = {2023-07-04},
+
+}
