+++ docs/articles/joss/paper.md
@@ -62,7 +62,7 @@ This software is available as a free and opensource R package.
 
 ## Comparing maps
 As stated in [@visser2006map] comparing map is an important issue in environmental research. 
-The four main reasons to compare categorical variable son geographical units are :     
+The four main reasons to compare categorical variables on geographical units are :     
 - To compare maps generated by models under different scenarios and assumptions  
 - To detect temporal changes  
 - To calibrate or validate models   
@@ -72,17 +72,16 @@ The four main reasons to compare categorical variable son geographical units are
 ## Comparing specifically LCZ maps 
 
 Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
-The terms rural and urban lack of a clear definition and several different classifications 
+The terms rural and urban lack of a clear definition and different classifications 
 of urban and rural landscapes were proposed.
 Stewart and Oke [@stewart2012local] defined an approach that complies with [@grigg1965logic] criteria of geographical classification.
 Their Local Climate Zones (LCZ) are based on a logical segmentation of land-use and urban canopy parameters 
 and define 10 urban types and 7 land cover types.
 
 LCZ have gained popularity in the past decade as they sum up relevant information and can help, for instance, 
-apprehend the intensity of the Urban Heat Island ([@kotharkar2018evaluating]).
+apprehend the intensity of the Urban Heat Island [@kotharkar2018evaluating].
 
-Several papers describe methods to produce LCZ classifications, but there are fewer workflows that automatically 
-produce them from available data.
+Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification for any given area.  
 [@quan2021systematic] distinguishes two main streams of production of these LCZ:   
 
 - the raster stream processes remotely sensed information, and applies machine learning 
@@ -94,7 +93,7 @@ For instance, the GeoClimate geospatial toolbox produces LCZ classifications
 from OpenStreetMap or french BDTopo data [@bocher2021geoclimate].   
 
 The existence of several methods to produce LCZ classifications, or the use of a method with different input data,
-raise the need to quickly get :
+raises the need for a tool to quickly get :
 
 1.  Statistics measuring the general agreement between two classifications
 2.  A spatial representation of this agreement, to allow a fast visual assessment of the zones 
@@ -138,7 +137,7 @@ specific tools are needed to allow easy automatisation of LCZ maps. As a matter 
 define in matters of LCZ, and one may want to compare LCZ produced by experts and by different algorithms. 
 
 The very definition of LCZ by [@stewart2012local] rely on ranges of values for several indicators, and  sometimes
-several LCZ types could be assign to the corresponding geometries. Therefore, some algorithms associate a confidence 
+several LCZ types could be assigned to the corresponding geometries. Therefore, some algorithms associate a confidence 
 value to the LCZ type and one may want to check if filtering pixels or geometries accordingly to a confidence threshold
 has an impact on the way maps agree or disagree.   
 
@@ -160,24 +159,25 @@ the same spatial units to classify the area.
 ![Workflow and main functions of the package\label{fig:Workflow and main functions of the package}](workflow.png){width="100%"}
 
 This figure describes the general workflow of the package.
-Main functions are presented in plain lines, the dashed box represent optionnal steps.
+Main functions are presented in plain lines, the dashed box represent optionnal steps. These functions are presented
+ in next section.
 
 1.  First, the LCZ classifications are imported, from a geojson file or shapefiles.
 2.  Each LCZ classification can then be plotted
 3.  The two LCZ classifications are then compared :
-    1.  A map of the agreement/disagreement is produced
-    2.  The general agreement and a pseudo-kappa indicator of agreement are computed. 
[…]
