+++ R/showLCZ.R
@@ -19,7 +19,7 @@
 #' @examples # On original LCZ levels, use the \'standard\' value for the \'repr\' argument.
 #' showLCZ(redonBDT,column="LCZ_PRIMARY", repr="standard")
 #' # On grouped data, use the alter value for the repr argument.
-#' redonBDTgrouped<-LCZgroup2(redonBDT,column="LCZ_PRIMARY",
+#' redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",
 #' urban=c("1","2","3","4","5","6","7","8","9"),
 #' industry="10", vegetation=c("101","102","103","104"),
 #' impervious="105",pervious="106",water="107")
@@ -35,7 +35,7 @@ showLCZ<-function(sf, title="", wf="",column="LCZ_PRIMARY",
 
   try(class(sf)[1]=="sf", stop("Input data must be sf object"))
 
- if(wf!=""){nomLegende<-paste0("LCZ from ",wf," workflow")} else{nomLegende<-"LCZ"}
+ if(wf!=""){nomLegende<-paste0("LCZ from ",wf," workflow")} else{nomLegende<-"Levels"}
 
   if (repr=='standard'){
   LCZlevels<-as.character(c(1:10,101:107))
