+++ docs/articles/joss/paper.md
@@ -45,7 +45,7 @@ editor_options:
 
 # Summary
 
-Climate change is a growing concern for city planners as Urban Heat Islands has an impact on mortality [@clarke1972some], 
+Climate change is a growing concern for city planners as Urban Heat Islands have an impact on mortality [@clarke1972some], 
 health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
 A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, 
 such as Local Climate Zones (LCZ) defined by [@stewart2012local].
@@ -62,17 +62,17 @@ This software is available as a free and opensource R package.
 
 ## Comparing maps
 As stated in [@visser2006map] comparing map is an important issue in environmental research. 
-The four main reasons to compare categorical variables on geographical units are :     
-- To compare maps generated by models under different scenarios and assumptions  
-- To detect temporal changes  
-- To calibrate or validate models   
-- To perform uncertainty and sensitivity analysis  
+The four main reasons to compare categorical variables on geographical units are:     
+- to compare maps generated by models under different scenarios and assumptions,  
+- to detect temporal changes,  
+- to calibrate or validate models,   
+- to perform uncertainty and sensitivity analysis.  
 
 
 ## Comparing specifically LCZ maps 
 
 Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
-The terms rural and urban lack of a clear definition and different classifications 
+The terms *rural* and *urban* lack of a clear definition and different classifications 
 of urban and rural landscapes were proposed.
 Stewart and Oke [@stewart2012local] defined an approach that complies with [@grigg1965logic] criteria of geographical classification.
 Their Local Climate Zones (LCZ) are based on a logical segmentation of land-use and urban canopy parameters 
@@ -96,106 +96,115 @@ The existence of several methods to produce LCZ classifications, or the use of a
 raises the need for a tool to quickly get :
 
 1.  Statistics measuring the general agreement between two classifications
-2.  A spatial representation of this agreement, to allow a fast visual assessment of the zones 
-where the classifications differ.
+2.  A spatial representation to allow a fast visual assessment of the zones 
+where the classifications agree or disagree.
 3.  A confusion matrix, to explore how the levels of one classification break up 
 in levels of the other classification, shown in a graphical way to help visualize the 
 main differences between the classifications.
 
 # State of the field and feature comparison
 
-## State on the field on spatial classification comparison
+## State of the field on spatial classification comparison
 
 The question of going beyond the simple visual observation of differences between maps is not a new one. 
 A first approach would be, on raster maps, to simply compute the agreement between two maps as the proportion 
 of pixels for which the two maps have the same value of the variable of interest. This approach is often enough 
-to help specialists to compare pairs of maps, but it presents two main drawbacks : 
- - two totally random maps won't have a value of agreement of zero as some pixels may agree per chance,
- - only raster maps where the pixel perfectly coincide (same size, not translated) can be treated.
+to help specialists compare pairs of maps, but it presents two main drawbacks :  
+
+ - two totally random maps won't have a value of agreement of zero as some pixel values may agree by chance,  
+ - only raster maps where pixels match perfectly (same size, not translated) can be treated, or 
+some pre-treatment are needed (like nearest neighbour interpolation for instance).  
 
 To prevent the first drawback, [@monserudComparingGlobalVegetation1992] 
 proposed the use of Cohen's kappa coefficient of agreement for nominal scales [@cohenCoefficientAgreementNominal1960].
-Each pixel is seen as an individual to which each map assigns a value of a categorical variable (each map is seen as a "rater").  
[…]
