+++ docs/articles/joss/paper.bib
@@ -208,3 +208,31 @@
   doi = {10.5194/gmd-2021-428},
   urldate = {2022-09-05}
 }
+
+@article{bocherGeoprocessingFrameworkCompute2018,
+  title = {A Geoprocessing Framework to Compute Urban Indicators: {{The MApUCE}} Tools Chain},
+  shorttitle = {A Geoprocessing Framework to Compute Urban Indicators},
+  author = {Bocher, Erwan and Petit, Gwendall and Bernard, J{\'e}r{\'e}my and Palominos, Sylvain},
+  year = {2018},
+  month = jun,
+  journal = {Urban Climate},
+  volume = {24},
+  pages = {153--174},
+  issn = {22120955},
+  doi = {10.1016/j.uclim.2018.01.008},
+  urldate = {2023-07-05},
+  langid = {english}
+}
+
+@article{RJ-2018-009,
+  author = {Edzer Pebesma},
+  title = {{Simple Features for R: Standardized Support for Spatial
+          Vector Data}},
+  year = {2018},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2018-009},
+  url = {https://doi.org/10.32614/RJ-2018-009},
+  pages = {439--446},
+  volume = {10},
+  number = {1}
+}
