+++ R/groupLCZ.R
@@ -10,12 +10,12 @@
 #' @export
 #'
 #' @examples
-#' redonBDTgrouped<-LCZgroup2(redonBDT,column="LCZ_PRIMARY",
+#' redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",
 #' urban=c("1","2","3","4","5","6","7","8","9"),
 #' industry="10", vegetation=c("101","102","103","104"),
 #' impervious="105",pervious="106",water="107",
 #' colors=c("red","black","green","grey","burlywood","blue"))
-LCZgroup2<-function(sf,column,outCol='grouped',...)
+groupLCZ<-function(sf,column,outCol='grouped',...)
 {
   #require(forcats)
   #require(dplyr)
