+++ inst/tinytest/test_importQualVar.R
@@ -52,13 +52,13 @@ library(tidyr)
 pivot_wider(utrfComparison$matConf, names_from = TYPO_MAJ.1, values_from = agree)
 expect_equal(utrfComparison$matConf[1,3],62.96)
 # 
-# expect_silent(importLCZgen(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon")
+# expect_silent(importLCZvect(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon")
 #                            ,file="rsu_lcz.geojson",
 #                            column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"))
 # 
 # 
 # # Tests if the imported version of Redon test data matches the Redon test data in the package
-# redonBDT2<-importLCZgen(dirPath=paste0(
+# redonBDT2<-importLCZvect(dirPath=paste0(
 #   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),
 #   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE")
 # st_crs(redonBDT2)$wkt<-gsub("é","e",st_crs(redonBDT2)$wkt)
@@ -67,14 +67,14 @@ expect_equal(utrfComparison$matConf[1,3],62.96)
 # 
 # # Error if file doesn't exist in location
 # 
-# expect_error(importLCZgen(dirPath=paste0(
+# expect_error(importLCZvect(dirPath=paste0(
 #   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="chaussure.geojson",
 #   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),"The file doesn't seem to exist.")
 # 
 # ## Error if directory doesn't exist
 # 
 # 
-# expect_error(importLCZgen(dirPath=paste0(
+# expect_error(importLCZvect(dirPath=paste0(
 #   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redono"),file="rsu_lcz.geojson",
 #   column="LCZ_PRIMARY",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),
 #   "The directory set in dirPath doesn't seem to exist")
@@ -82,54 +82,54 @@ expect_equal(utrfComparison$matConf[1,3],62.96)
 # # test if one column doesn't exist
 # 
 # expect_error(
-# importLCZgen(dirPath=paste0(
+# importLCZvect(dirPath=paste0(
 #   system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="rsu_lcz.geojson",
 #   column="LCZ_PRIMAR",geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE"),
 #   "It seems that some of the columns you try to import do not exist in the source file")
 # 
 # expect_error(
-#   importLCZgen(dirPath=paste0(
+#   importLCZvect(dirPath=paste0(
 #     system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"),file="rsu_lcz.geojson",
 #     column="LCZ_PRIMARY",geomID="ID_RSY",confid="LCZ_UNIQUENESS_VALUE"),
 #   "It seems that some of the columns you try to import do not exist in the source file")
 # 
 # expect_error(
-#   importLCZgen(dirPath=paste0(
+#   importLCZvect(dirPath=paste0(
[…]
