+++ docs/articles/joss/paper.md
@@ -82,8 +82,7 @@ LCZ have gained popularity in the past decade as they sum up relevant informatio
 apprehend the intensity of the Urban Heat Island [@kotharkar2018evaluating].
 
 Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification for any given area.  
-[@quan2021systematic] distinguishes two main streams of production of these LCZ:   
-
+[@quan2021systematic] distinguishes two main streams of production of these LCZ:      
 - the raster stream processes remotely sensed information, and applies machine learning 
 algorithms trained using local experts' knowledge. For instance, the WUDAPT platform produced 
 LCZ maps of Europe and North-America this way [@chingWUDAPTUrbanWeather2018a].   
@@ -93,7 +92,7 @@ For instance, the GeoClimate geospatial toolbox produces LCZ classifications
 from OpenStreetMap or french BDTopo data [@bocher2021geoclimate].   
 
 The existence of several methods to produce LCZ classifications, or the use of a method with different input data,
-raises the need for a tool to quickly get :
+raises the need for a tool to quickly get:
 
 1.  Statistics measuring the general agreement between two classifications
 2.  A spatial representation to allow a fast visual assessment of the zones 
@@ -121,8 +120,8 @@ proposed the use of Cohen's kappa coefficient of agreement for nominal scales [@
 Each pixel is seen as an individual to which each map assigns a value of a categorical variable 
 (each map is seen as a "rater").  
 
-The comprehensive Map Comparison kit which was released in 2001 by the 
-Netherlands Environmental Assessment Agency [@visser2006map]. It is an example of tool  
+The comprehensive Map Comparison kit, which was released in 2001 by the 
+Netherlands Environmental Assessment Agency [@visser2006map], is an example of tool  
 that allows multiple methods to compare raster maps. It includes a fuzzy algorithm which allows to tackle small 
 shifts of one map from another. It only works on raster maps, only on Windows OS and 
 doesn't allow automation for several pairs of maps.  
@@ -139,25 +138,25 @@ compares some model output variables using the data of the map.
 
 As Local Climate Zones emerged as a new standard for characterizing urban landscapes [@demuzereLCZGeneratorWeb2021], 
 specific tools are needed to allow easy automation of LCZ maps. As a matter of fact, "ground truth" is hard to 
-define in matters of LCZ, and one may want to compare LCZ produced by experts and by different algorithms. 
+define in matters of LCZs, and one may want to compare LCZs produced by experts and by different algorithms. 
 
-The very definition of LCZ by [@stewart2012local] relies on ranges of values for several indicators, and  sometimes
-several LCZ types could be assigned to the corresponding geometries. Therefore, some algorithms associate a confidence 
-value to the LCZ type and one may want to check if filtering pixels (for raster data) or geometries (for vector data) accordingly to a confidence threshold
-has an impact on the way maps agree or disagree.   
+The very definition of LCZs by [@stewart2012local] relies on ranges of values for several indicators, and  sometimes
+several LCZ types could be assigned to the corresponding geometries. Therefore, one may associate a confidence 
+value to the LCZ type and check if filtering pixels (for raster data) or geometries (for vector data) 
+accordingly to a confidence threshold has an impact on the way maps agree or disagree.   
 
 A comparison of a raster stream result and a GIS vector stream approach was proposed by [@muhammad2022inference].
 This comparison relies on rasterising vector data. It uses several tools: QGIS, python scripts and SAGA GIS. As far as we know, 
 the scripts and the automation of the method are not publicly available.
 
[…]
