+++ inst/tinytest/test_levCol.R
@@ -105,12 +105,12 @@ expect_equal(grep("10:",test$case),1)
 #           the other a vector of colors, whose length is shorter than the specified levels.
 #           Missing colors will be picked from a standard palette.
 
-test<-levCol(redonBDTgrouped, column="grouped",
-             levels=c("urban","industry","vegetation","impervious","pervious","water"),
-             colors=c("red","black","green","grey","blue"))
+# test<-levCol(redonBDTgrouped, column="grouped",
+#              levels=c("urban","industry","vegetation","impervious","pervious","water"),
+#              colors=c("red","black","green","grey","blue"))
 
 
-expect_equal(grep("11:",test$case),1)
+# expect_equal(grep("11:",test$case),1)
 
 # case 12: One vector seems to be a vector of levels,
 #           which covers the values of the data,
