### 2023-07-07 47e5ca6
On the way to a cleaner easier to maintain version of levCol TO DO : rename the case and write the test file.
