+++ R/levCol2.R
@@ -1,0 +1,222 @@
+#' Manages the levels and the colors of the LCZ columns
+#'
+#' @param sf is the input sf file
+#' @param column is the column that contains the data
+#' @param drop is set to TRUE if one wants to drop unused levels, in case column is a factor with unused levels
+#' @param useStandCol is set to TRUE if one wants recognized standard levels to be associated with standard colors. Default is FALSE
+#' @param ... other parameters specified, expected a vector of levels and a vector of colors
+#' whose name must begin with colors. Other cases are handled to enhance usability.
+#' @import dplyr sf
+#' @importFrom grDevices palette.colors
+#'
+#' @return output is a list containing levelColors, a named vector, which names are the levels
+#' present in the data and which values are the associated colors,
+#' and case, a string spcifying what case was encountered when producing the levels and colors.
+#' @export
+#'
+#' @examples
+#' #levCol2 is not to be used directly by user. It deals with levels and colors provided by the user
+levCol<-function(sf,column,drop=FALSE,useStandCol=FALSE,...){
+# Getting ALL the arguments in (...) and computing the values needed for conditions statements
+  args<-list(...)
+  
+# Stop if there are more than 37 arguments.   
+  if (length(args)>=37){ stop("This function can not deal with more than 36 arguments.
+  You can use the function groupLCZ to group some levels.")}
+
+# get the natural levels from the data : uniqueData
+  uniqueData<-sf[column] %>% sf::st_drop_geometry()  %>% unique() 
+  # Attention, unique outputs a list of length 1
+  
+  # Sometimes, the input data can be a factor, with levels not present in the actual data.
+  # One may want to drop these levels
+  
+  if(drop==TRUE){uniqueData<-droplevels(uniqueData)}
+  
+  uniqueData<-levels(uniqueData[,1]) %>% as.character() %>% as.vector()
+
+  # Stop if there are more than 36 levels : really impossible to read such a map
+  # or confusion matrix
+  
+  if(length(uniqueData)>36){ stop(
+    "This package is not suited for qualitative variables/classifications with more than 36 levels/types.
+  You can use the function groupLCZ to group some levels.") }
+
+# Getting the levels and separating level vector(s) and colors vector if present.   
+  
+  argNames<-names(args)
+  indCol<-grep(x=argNames, pattern="colors")
+  if (length(indCol) != 0) {
+    if (length(indCol)>1 ) {
+      stop(
+        "Only one argument can start with colors, and it must contain the colors,
+      please rename your arguments and retry.")} else {
+        argCol <- args[indCol][[1]]
+        argLev<-args[-indCol]
+        # In case the color vector is made of empty strings
+        if (prod(argCol=="")==1){
+          args<-args[-indCol]
+          argCol<-NULL
+          # In case the vector of levels is simultaneously made of empty strings
+          if(prod(unlist(argLev)=="")==1){args<-NULL}
+      }
+    }} else
+  {
+    argCol<-NULL
+    argLev<-args
+  }
+
+  if(length(argLev)>36){ stop("This package is not suited for classification with more than 36 levels or types.
+  You can use the function groupLCZ to group some levels.") }
+  
+#############################################################################################
+## Simplest Case no arguments passed : levels and colors deduced from the Data
+#############################################################################################
+   
+if (length(args) == 0 ||
+  (is.null(args)) ||
+  (prod(unlist(args)=="")==1)) {
+  case<-"1: No level vector and no color vector, less than 36 levels,
+        levels will be deduced from the data
+        and colors will be chosen from a standard palette."
+
+  typeLevels<-standLevCol( levels=uniqueData,
+                           colors=palette.colors(n=length(uniqueData), palette="Polychrome 36"),
[…]

[…]
