### 2023-07-07 055333d
levCol function covers all cases. TO DO : rewrite messages and write test_levCol2.R. Repalce levCol2.R by levCol2 and same with test files
