+++ inst/tinytest/test_levCol2.R
@@ -1,8 +1,10 @@
 # This tests the function levCol.R
 # library(tinytest)
 # library(lczexplore)
-#
+# library(dplyr)
 # library(sf)
+
+dealtWithCases<-c("1","2","3","3.1","4","5","6","7","8","13")
 
 redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",urban=c("1","2","3","4","5","6","7","8","9"),
                            industry="10",
@@ -127,7 +129,7 @@ test<-levCol(redonBDTgrouped, column="grouped",
              colors=c("red",  "black",   "green",     "grey",       "blue"))
 
 
-# expect_equal(grep("11:",test$case),1)
+expect_equal(grep("11:",test$case),1)
 
 # case 12: One vector seems to be a vector of levels,
 #           which covers the values of the data,
@@ -160,11 +162,13 @@ expect_equal(grep("14.0:",test$case),1)
 #        and the number of specified colors is zero or less than the number of levels present,
 #        levels will be deduced from the data and colors will be chosen from a standard palette.
 
-test<-levCol(sf=redonBDTgrouped,column="grouped",
+test<-levCol(sf=redonBDTgrouped, column="grouped",
        industry="10",
        vegetation=c("101","102","103","104"),
        impervious="105",pervious="106",water="107",
-       colors=c("red","black","green","grey","burlywood","blue"))
+       colors=c("red","black","green","grey","burlywood"))
+
+
 expect_equal(grep("14:",test$case),1)
 
 # case 15: The specified levels don't cover the levels in the data
@@ -172,7 +176,7 @@ expect_equal(grep("14:",test$case),1)
 # to the number of levels present in the data,
 # they are matched in the order of appearence.
 
-test<-levCol(sf=redonBDTgrouped,column="grouped",urban=c("1","2","3","4","5","6","7","8","9"),
+test<-levCol(sf=redonBDTgrouped,column="grouped",urban=c("1","2","3","4","5","6","7","8","9"), 
        vegetation=c("101","102","103","104"),
        impervious="105",pervious="106",water="107",
        colors=c("red","black","green","grey","burlywood","blue","orange"))
@@ -205,6 +209,8 @@ test<-levCol(sf=redonBDTgrouped,column="grouped",urban=c("1","2","3","4","5","6"
        vegetation=c("101","102","103","104"),
        impervious="105",pervious="106",water="107",
        colors=c("red","black","green","grey","burlywood","blue"))
+
+
 expect_equal(grep("17:",test$case),1)
 
 # case 18 : Several arguments are specified, whose names cover
@@ -217,6 +223,9 @@ test<-levCol(sf=redonBDTgrouped,column="grouped",urban=c("1","2","3","4","5","6"
        vegetation=c("101","102","103","104"),
        impervious="105",pervious="106",water="107",
        colors=c("red","black","green","grey","DJ Shadow","blue"))
+
+
+
 expect_equal(grep("18:",test$case),1)
 
 redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",urban=c("1","2","3","4","5","6","7","8","9"),

[…]
