+++ inst/tinytest/test_levCol2.R
@@ -1,248 +1,0 @@
-# This tests the function levCol.R
-# library(tinytest)
-# library(lczexplore)
-# library(dplyr)
-# library(sf)
-
-dealtWithCases<-c("1","2","3","3.1","4","5","6","7","8","13")
-
-redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",urban=c("1","2","3","4","5","6","7","8","9"),
-                           industry="10",
-                           vegetation=c("101","102","103","104"),
-                           impervious="105",pervious="106",water="107",colors=c("red","black","green","grey","burlywood","blue"))
-
-# several vector covering the levels in the data, with colors of the proper size
-
-# case 0: The number of levels must be less than 37 for the map to be readable,
-#               you may want to group some of the levels using groupLCZ function
-expect_error(
-  levCol(sf=redonBDT, column="LCZ_PRIMARY",1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18,
-       19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38,colors="black"),
-  "more than 36 arguments")
-
-# case 1: No level vector and no color vector, less than 36 levels,
-# levels will be deduced from the data
-# and colors will be chosen from a standard palette. Or there are a level vector and a color vector, but they are empty.
-
-test<-levCol(redonBDTgrouped,column="grouped",levels="",colors="")
-expect_equal(grep("1:",test$case),1)
-
-
-test<-levCol(sf=redonBDT, column="LCZ_PRIMARY")
-expect_equal(grep("1:",test$case),1)
-
-# case 2: No levels but a color vector which size covers the number of levels in the data
-
-test<-levCol(sf=redonBDTgrouped,column="grouped",
-             colors=c("red","black","green","grey","burlywood","blue"))
-expect_equal(grep("2:",test$case),1)
-
-# case 3: No levels but a color vector which size does not cover the number of levels in the data,
-# colors will be picked from a standard palette..
-
-test<-levCol(sf=redonBDTgrouped,column="grouped",
-             colors=c("red","black","green","grey","burlywood"))
-expect_equal(grep("3:",test$case),1)
-
-test<-test<-levCol(sf=redonBDTgrouped,column="grouped",
-                   colors=c("red","black","green","grey","burlywood","blue","purple"))
-
-
-
-
-
-
-# case 4:  A single vector was provided, whose names cover the levels in the data
-# and whose values are colors.
-test<-levCol(redonBDTgrouped, column="grouped", levels=c("urban"="red","industry"="black",
-                                                         "vegetation"="green",
-                                                         "impervious"="grey",
-                                                         "pervious"="burlywood","water"="blue"))
-expect_equal(grep("4:",test$case),1)
-
-# case 5: A single vector was provided, whose values are colors
-#         but whose names don't cover the levels in the data.
-#         Specified colors will be associated to levels deduced from the data,
-#          in order of appearence.
-test<-levCol(redonBDTgrouped, column="grouped", drop=TRUE, levels=c("urban"="red","industry"="black",
-                                                         "vegetation"="green",
-                                                         "impervious"="grey",
-                                                         "not in data"="white",
-                                                         "pervious"="burlywood","water"="blue"))
-expect_equal(grep("5:",test$case),1)
-
-# case 6: A single vector was provided, whose values are colors
-#         but whose names don't cover the levels in the data.
-#         Colors will be associated to unique values of the data
-#         and missing colors will be added from a standard palette.
[…]
