+++ man/levCol.Rd
@@ -1,11 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/levCol.R, R/levCol2.R
+% Please edit documentation in R/levCol.R
 \name{levCol}
 \alias{levCol}
 \title{Manages the levels and the colors of the LCZ columns}
 \usage{
-levCol(sf, column, drop = FALSE, useStandCol = FALSE, ...)
-
 levCol(sf, column, drop = FALSE, useStandCol = FALSE, ...)
 }
 \arguments{
@@ -24,17 +22,10 @@ whose name must begin with colors. Other cases are handled to enhance usability.
 output is a list containing levelColors, a named vector, which names are the levels
 present in the data and which values are the associated colors,
 and case, a string spcifying what case was encountered when producing the levels and colors.
-
-output is a list containing levelColors, a named vector, which names are the levels
-present in the data and which values are the associated colors,
-and case, a string spcifying what case was encountered when producing the levels and colors.
 }
 \description{
 Manages the levels and the colors of the LCZ columns
-
-Manages the levels and the colors of the LCZ columns
 }
 \examples{
-#levCol is not to be used directly by user. It deals with levels and colors provided by the user
 #levCol2 is not to be used directly by user. It deals with levels and colors provided by the user
 }
