+++ R/produceAnalysis.R
@@ -1,7 +1,8 @@
 #' Calls functions of the package to produce an analysis of two set of LCZ
 #' produced exclusively by GeoClimate.
-#' Typically one will compare the LCZ produced by GeoClimate using the OSM data as input
-#' to the LCZ produced by GeoClimate using the french BD_TOPO as input.
+#' Typically one will compare the LCZ produced by GeoClimate using the OpenStreetMap
+#' data as input to 
+#' the LCZ produced by GeoClimate using the french BDTopo v2.2 as input.
 #' Same analysis are reproducible using all the generic functions of the package.
 #'
 #' @param location defines the study area. The study area must have been coputed
@@ -113,7 +114,7 @@ if(repr=="standard"){
                      sf2=df2,
                      column2="LCZ_PRIMARY", geomID2="ID_RSU", confid2="LCZ_UNIQUENESS_VALUE",wf2=wf2,
                      ref=1,
-                     repr="standard", saveG=nameG, exwrite=T,outDir=outDir,location=location)
+                     repr="standard", saveG=nameG, exwrite=TRUE,outDir=outDir,location=location)
                )
 
       }
@@ -124,7 +125,7 @@ if(repr=="standard"){
                    column1="EU_LCZ_map",
                    sf2=df2,
                    column2='LCZ_PRIMARY',
-                   saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=T,outDir=outDir,location=location,...)
+                   saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=TRUE,outDir=outDir,location=location,...)
       }
 
       if(wf2=="wudapt"&& (wf1=="osm"|wf1=="bdtopo_v2")){
@@ -133,7 +134,7 @@ if(repr=="standard"){
                      column1='LCZ_PRIMARY',
                      sf2=df2,
                      column2="EU_LCZ_map",
-                     saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=T,outDir=outDir,location=location,...)
+                     saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=TRUE,outDir=outDir,location=location,...)
       }
 }
   if(repr=='alter'){
@@ -150,7 +151,7 @@ if(repr=="standard"){
                  column1='grouped',
                  sf2=df2,
                  column2='grouped',
-                 ref=1,saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=T,outDir=outDir,location=location,...)
+                 ref=1,saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,exwrite=TRUE,outDir=outDir,location=location,...)
       }
 
     if(wf1=="wudapt"&(wf2=="osm"|wf2=="bdtopo_2_2")){
@@ -160,7 +161,7 @@ if(repr=="standard"){
                  column1='grouped',
                  sf2=df2,
                  column2='grouped',
-                 saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,location=location,exwrite=T,outDir=outDir,...)
+                 saveG=nameG,repr=repr,wf1=wf1,wf2=wf2,location=location,exwrite=TRUE,outDir=outDir,...)
     }
     if(wf2=="wudapt"&(wf1=="osm"|wf1=="bdtopo_2_2")){
       df1<-groupLCZ(df1,column="LCZ_PRIMARY",...)
@@ -169,7 +170,7 @@ if(repr=="standard"){
                  column1='grouped',
                  sf2=df2,
                  column2='grouped',
[…]
