+++ R/importLCZvect.R
@@ -1,27 +1,32 @@
-#' Imports LCZ from a standard geographical file (tested : geojson, shp, more to come)
+#' Imports Local Climate Zone classifications from a standard geographical file (tested : geojson, shp, more to come)
 #'
 #' @param dirPath is the path of the directory of the file
 #' @param file is the name of the file from which the LCZ are imported
-#' @param column indicates the name of the column containing LCZ values, all other
-#' columns will be dropped. If empty string, all the colums will be kept
-#' @param geomID is the name of the column containing the ID of each geom to load. If an empty string, no column is loaded.
+#' @param column indicates the name of the column containing LCZ values. 
+#' LCZ values are expected to be of a standard LCZ format (1 to 17, or 1 to 10 and 101 to 107 or 1 to G),
+#' else, use the importQualVar function
+#' @param geomID is the name of the column containing the ID of each geom to load. 
+#' If an empty string, no column is loaded.
 #' @param confid is the name of the column containing a confidence indicator to filter geoms,
 #' for instance the uniqueness of the LCZ level of each geom
-#' @param output : if sfFile, the function returns an sfFile with the LCZ colums,
+#' @param output : if sfFile, the function returns an sfFile with the specified columns,
 #' if bBox, returns a bounding box one can use to crop a raster file or to intersect another sf file
 #' @param typeLevels the levels of the imported LCZ classification
 #' @param verbose if TRUE show the discrepancies between specified levels of LCZ and
 #' levels actually present in column
-#' @param drop : the default is TRUE, which means all the column are dropped excepted those specified in previous parameters
+#' @param drop : the default is TRUE, which means all the column are 
+#' dropped excepted those specified in previous parameters
 #' @import dplyr forcats rlang sf
 #' @importFrom terra crop
 #' @importFrom terra rast
-
-#' @return returns an sf object containing at least the geoms and the LCZ values, and if specified, columns for the IDs of the geoms and the confidence value of the LCZ levels.
+#' @return returns an sf object containing at least the geoms and the LCZ values, 
+#' and if specified, columns for the IDs of the geoms and the confidence value of the LCZ levels.
 #' @export
-#' @examples importLCZvect(dirPath=paste0(system.file("extdata", package = "lczexplore"),
+#' @examples 
+#' redonBDTex<-importLCZvect(dirPath=paste0(system.file("extdata", package = "lczexplore"),
 #' "/bdtopo_2_2/Redon"), file="rsu_lcz.geojson", column="LCZ_PRIMARY",
 #' geomID="ID_RSU",confid="LCZ_UNIQUENESS_VALUE")
+#' showLCZ(redonBDTex)
 importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column="LCZ_PRIMARY",
                        geomID="", confid="",
                        typeLevels=c("1"="1","2"="2","3"="3","4"="4","5"="5","6"="6","7"="7","8"="8",
