+++ R/areColors.R
@@ -1,4 +1,4 @@
-#' Check if a string or a vector of string define colors in R
+#' Checks if a string or a vector of strings define colors in R
 #'
 #' @param x is the input string
 #' @return a vector of booleans indicting if the elements of x define a color in R (TRUE) or don't (FALSE)
