+++ man/groupLCZ.Rd
@@ -2,7 +2,8 @@
 % Please edit documentation in R/groupLCZ.R
 \name{groupLCZ}
 \alias{groupLCZ}
-\title{Allows to group local climate zones to improve the analysis.}
+\title{Allows to group (Local Climate Zone) geographical classification levels into broader categories
+to explore classification agreements}
 \usage{
 groupLCZ(sf, column, outCol = "grouped", ...)
 }
@@ -19,7 +20,8 @@ groupLCZ(sf, column, outCol = "grouped", ...)
 a file containing the original geom and lcz, and a new grouped column
 }
 \description{
-Allows to group local climate zones to improve the analysis.
+Allows to group (Local Climate Zone) geographical classification levels into broader categories
+to explore classification agreements
 }
 \examples{
 redonBDTgrouped<-groupLCZ(redonBDT,column="LCZ_PRIMARY",
@@ -27,4 +29,9 @@ urban=c("1","2","3","4","5","6","7","8","9"),
 industry="10", vegetation=c("101","102","103","104"),
 impervious="105",pervious="106",water="107",
 colors=c("red","black","green","grey","burlywood","blue"))
+
+showLCZ(redonBDT,column="LCZ_PRIMARY", repr="standard")
+showLCZ(redonBDTgrouped,column="grouped",repr="alter",
+LCZlevels=c("urban","industry","vegetation","impervious","pervious","water"),
+colors=c("red","black","green","grey","burlywood","blue"),wf="BD TOPO")
 }

[…]
