### 2023-07-10 977c693
Update of man page fro compareLCZ
