+++ R/compareLCZ.R
@@ -1,53 +1,49 @@
-#' Compares two LCZ classification on the same zone,
-#' produces a map for each classification, a map of their agreement and a representation of a confusion matrix between them
+#' Compares two set of geographical classifications, especially Local Climate Zones classifications. It
+#' produces a map for each classification, a map of their agreement (and a pseudo Kappa coefficent), 
+#' and a confusion matrix between them. All are stored in a list, easily reusable. 
 #'
-#' @param sf1 is the sf object that contains the first LCZ classification
-#' @param geomID1 is the name of the column storing the ID of the geoms in sf1
+#' @param sf1 is the sf object that contains the first (LCZ) classification
 #' @param column1 is the column of sf1 that contains the LCZ classification for each geom of sf1.
+#' @param geomID1 is the name of the optionnal column storing the ID of the geoms in sf1
 #' By defautl it is set to an empty string and no ID is loaded.
-#' @param confid1 is a column that contains an indicator of confidence
-#' of the level of the LCZ in column 1, e.g. a uniqueness value, or a probability of belonging to the class...
+#' @param confid1 is an optionnal column that contains an indicator of confidence
+#' of the values in column1, e.g. a uniqueness value, or a probability of belonging to the class...
 #' By defautl it is set to an empty string and no confidence indicator is loaded.
-#' @param wf1 is the workflow used to produce the first LCZ classification.
+#' @param wf1 is the workflow used to produce the first LCZ classification. 
+#' It is used ti create titles or legends. 
 #' When GeoClimate was used with BD_TOPO V2 data as input,
 #' use "bdtopo_2_2". When GeoClimate was used with Open Street Map data as input, use "osm".
 #' When the LCZ come from the wudapt Europe tiff, use "wudapt".
-#' @param sf2 is the sf object that contains the second LCZ classification
+#' @param sf2 is the sf object that contains the second (LCZ) classification
 #' @param geomID2 is the name of the column storing the ID of the geoms in sf2
 #' @param column2 is the column of sf2 that contains the LCZ classification for each geom of sf2
 #' @param confid2 is a column that contains an indicator of confidence
-#' of the level of the LCZ in column 2, e.g. a uniqueness value, or a probability of belonging to the class...
-#' By defautl it is set to an empty string and no confidence indicator is loaded.
 #' @param wf2 is the workflow used to produce the second LCZ classification.
-#' When GeoClimate was used with BD_TOPO V2 data as input,
-#' use "bdtopo_2_2". When GeoClimate was used with Open Street Map data as input, use "osm".
-#' When the LCZ come from the wudapt Europe tiff, use "wudapt".
-#' @param ref : If the coordinate reference system (CRS) of sf1 and sf2 differ, ref indicates which CRS to choose for both files (1 or 2)
+#' @param ref : If the coordinate reference systems (CRS) of sf1 and sf2 differ, ref indicates which CRS to choose for both files (1 or 2)
 #' @param repr "standard" means that standard values of LCZ are expected,
 #'  "alter" means other values are expected, like grouped values of LCZ or other qualitative variable.
-#'  In the latter case, the ... arguments must contain the groups and a color vector.
-#' @param plot : when FALSE non of the graphics are plotted or saved
+#'  In the latter case, the ... arguments can contain the expected levels and a color vector.
+#' @param plot : when FALSE none of the graphics are plotted or saved
 #' @param saveG : when an empty character string, "", the plots are not saved. Else, the saveG string is used to produce the name of the saved png file.
[…]
