### 2023-07-10 #5445 @MGousseff
@wcjochem , @matthiasdemuzere: Thank you for all the improvements you suggested. It took a little time but we think we addressed all of those improvements in this new version of the paper. Here are the points you raised and what we did about them.

- About the statement of need: we added references to some articles of map comparisons for environmental purpose. It led us to add a pseudo Cohen's Kappa coefficient that takes into account the part of agreement that could occur by chance.

- The question of what to do once differences have been highlighted is beyond the scope of this package/article, but lczexplore helps identify which steps of an algorithm need further investigation (in our example, one is the estimation of building height by an algorithm, and another, is the quality of the vegetation data).

- About the LCZ labels variability : the ones we met are taken into account by the `standard` representation :
1 to 10 then 11 to 17
1 to 10 then 101 to 107
1 to 10 then A to G
For other standards, see next point.

- A specific function was needed to import easily any  qualitative variable:  we added the importQualVar function.
It also allows to import LCZ classifications whose levels would not be expected in importLCZvect and importLCZraster functions.

- About coding standard :
The only pipe left is the tidyr one, as it is easier for retro-compatibility
The devtools::check_installed function only outputs 0 ERROR, 0 WARNING and 2 NOTES.
The notes are related to the example data: we chose to keep in the package full areas in order to have illustrative examples. No note is related to code anymore.
We used the Renv package to help users know which versions of packages we used.
Package folder has a Renv.lock. Using the renv::status() gives the version of packages when the sources were built.

- The vignettes are now properly built.
"lczexplore_raster_vector" vignette is dedicated to the comparison of a WUDAPT like LCZ map and a GeoClimate Like Vector map.

We recommand to start with `vignette("lczexplore_en")`

- The article describes the use of functions and the different arguments and ouputs in a more comprehensive way, and so do the man pages.

We thank you as your comments really helped to improve lczexplore package and we hope we answered your main concerns about this paper.
