### 2023-07-21 #5445 @matthiasdemuzere
@MGousseff, thanks for revising this submission.

For starters, I tried to install the package with vignetting:

```R
library(devtools)
devtools::install_github("orbisgis/lczexplore", build_vignettes = TRUE)
library(lczexplore)
```

Unfortunately this fails on my system, with the following complaints:

```R
✔  checking for file ‘/tmp/RtmpE8qrsq/remotes13db89663a5a/orbisgis-lczexplore-fa35bd9/DESCRIPTION’ ...
─  preparing ‘lczexplore’:
E  creating vignettes (19.2s)
   --- re-building ‘lczexplore_alter.Rmd’ using knitr

   Quitting from lines 15-24 [unnamed-chunk-1] (lczexplore_alter.Rmd)
   Error: processing vignette 'lczexplore_alter.Rmd' failed with diagnostics:
   trying to use CRAN without setting a mirror
   --- failed re-building ‘lczexplore_alter.Rmd’

   --- re-building ‘lczexplore_en.Rmd’ using knitr

   Quitting from lines 15-24 [unnamed-chunk-1] (lczexplore_en.Rmd)
   Error: processing vignette 'lczexplore_en.Rmd' failed with diagnostics:
   trying to use CRAN without setting a mirror
   --- failed re-building ‘lczexplore_en.Rmd’

   --- re-building ‘lczexplore_fr.Rmd’ using knitr

   Quitting from lines 17-27 [unnamed-chunk-1] (lczexplore_fr.Rmd)
   Error: processing vignette 'lczexplore_fr.Rmd' failed with diagnostics:
   trying to use CRAN without setting a mirror
   --- failed re-building ‘lczexplore_fr.Rmd’

   --- re-building ‘lczexplore_raster_vector.Rmd’ using knitr

   Quitting from lines 15-24 [unnamed-chunk-1] (lczexplore_raster_vector.Rmd)
   Error: processing vignette 'lczexplore_raster_vector.Rmd' failed with diagnostics:
   trying to use CRAN without setting a mirror
   --- failed re-building ‘lczexplore_raster_vector.Rmd’

   SUMMARY: processing the following files failed:
     ‘lczexplore_alter.Rmd’ ‘lczexplore_en.Rmd’ ‘lczexplore_fr.Rmd’
     ‘lczexplore_raster_vector.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error: Failed to install 'lczexplore' from GitHub:
  ! System command 'R' failed
```

I am not sure why this is.
To be honest, I have not used R for a very long time. So I did a clean install on my ubuntu 20.04 64bit system, with this version in place:

```R
> version
               _
platform       x86_64-pc-linux-gnu
arch           x86_64
os             linux-gnu
system         x86_64, linux-gnu
status
major          4
minor          3.1
year           2023
month          06
day            16
svn rev        84548
language       R
version.string R version 4.3.1 (2023-06-16)
nickname       Beagle Scouts
```

FYI: I am about to take a summer break, and I will only be able to look into this again from September onwards. My apologies for the delay this may cause.
