### 2023-07-22 #5445 @MGousseff
@matthiasdemuzere Thank you for trying to install this version.
I tried to install the package with these instruction on my home computer without this error.

The vignettes use a package called png, and the instruction at line 15-24 of the vignette is to check if it is installed and if not, to try and download it. Maybe there is a problem in the way CRAN package are downloaded ? I'll try from another computer which doesn't have any R version on it, and I'll install everything from scratch to see if I can reproduce the problem.

Maybe you can try to install knitr and png from your session before you install/load lczexplore ?

No problem for the delay, I am actually on holidays too, I'll just take the time to check from a test computer I keep to this end at home.

I hope we manage to fix this soon to see if the changes make the package more useful for you and your team.
Enjoy your summer.
