### 2023-08-22 c6afca7
produceAnalysis now keeps the created objects for further use fetchLCZ now checks if geojson file present before zip files, this prevents downloads when zip file deleted after extraction
