+++ R/fetchLCZ.R
@@ -46,21 +46,13 @@ fetchLCZ<-function(location,outDir,wf="bdtopo_2_2",refYear="2022",
       folderToken<-1
     }
 
-    if(file.exists(destFile)){
-      message('The zip file already exists')
-      if (file.exists(destLCZfile)){
-        message(' and the unzipped rsu_lcz.geojson file already exists
-            for this location and workflow.')}
-      else{message('and will be unzipped now.')
+  if (file.exists(destLCZfile)){message(' but an rsu_lcz.geojson already exists in this directory. \n If you think it is not the proper file
+              try deleting it and re-running the fetchLCZ function.')} else
+  { if(file.exists(destFile)) {
+      message('The zip file already exists and will be unzipped now.')
         unzip(destFile,exdir=folder)}
-    }
-    else {
-      message('The zip file doesn\'t exist in the directory')
-      if (file.exists(destLCZfile)){
-        message(' but an rsu_lcz.geojson already exists in this directory. \n If you think it is not the proper file
-              try deleting it and re-running the fetchLCZ function.')}
-      else{
-        message('\n it will be downloaded and unzipped.')
+     else {
+      message('The zip file doesn\'t exist in the directory \n it will be downloaded and unzipped.')
         try_fetch<-try(download.file(url=url,method="auto",destfile=destFile))
         if(is(try_fetch,"try-error")){
          warning("The file couldn't be downloaded, maybe the location wasn't proceeded by the GeoClimate team ?")
@@ -69,8 +61,9 @@ fetchLCZ<-function(location,outDir,wf="bdtopo_2_2",refYear="2022",
         download.file(url=url,destfile=destFile)
         unzip(destFile,exdir=folder)}
       }
-      }
   }
+}
+
 
 
 
