### 2023-08-23 7f5d3b0
More readable graphics for sensitivity analysis (and related tests)
