### 2023-08-24 838613f
Added a CRAN mirror for png package installation for vignette compilation.
