+++ vignettes/lczexplore_en.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
         comment = "#>"
 )
 if (!require("png")) {
-  install.packages("png")
+  install.packages("png",repos = "http://cran.us.r-project.org")
   library("png")
 }
 
