### 2023-08-24 #5445 @MGousseff
Dear @matthiasdemuzere, the error comes from the fact you haven't specified a CRAN mirror in your session. RStudio does it automatically in interactive mode, but not when a vignette is built, and I integrated the installation of the png package to allow better visualizations in the vignette. I have fixed this by specifying a CRAN mirror in the vignette code. If you try the install with this code

`library(devtools)`
`devtools::install_github("orbisgis/lczexplore")`
`library(lczexplore)`

it should install without problems. If you use RStudio, `vignette("lczexplore_en")` should offer a quick overview. But the code in the article must work as easily.

I also modified the graphics of the confidence analysis, for readability reasons.

I hope I understood your concerns and improved the package/article (for instance with other qualitative variable import and exploration) and you find some use of the actual version of the package.
