### 2023-09-13 orbisgis/lczexplore#12 @MGousseff
Done
