### 2023-09-13 orbisgis/lczexplore#15 @MGousseff
Done
