### 2023-09-15 #5445 @MGousseff
I'm sorry for closing the issue too fast. The different threads (reviews and Issues) was a bit confusing to me, I seeked advice and followed a colleague point of view of closing the issue after flagging them done. If I'm not mistaken, the last Issue still open is about 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?
What would be the best way to improve this in your opinion ? Would it be to add my contact info in the readme ?
As per your review, I thought it would appear here for me to take into account the improvements you may have asked ? (Thanks for your patience, still not used to which thread to follow).
