### 2023-09-18 #5445 @wcjochem
> What would be the best way to improve this in your opinion ? Would it be to add my contact info in the readme ?

This item is specifically listed in the JOSS checklist (https://joss.readthedocs.io/en/latest/review_checklist.html). It's up to you to decide how you want users to interact with your project. I suggest you check other JOSS published packages to see how they handle it. Some people add a section to their README, others add `.github/CONTRIBUTING.md`, others have other options.
