### 2023-09-18 ecacd6c
added CONTRIBUTING.md and CODE_OF_CONDUCT.md
