### 2023-09-18 3e9c2cd
added CONTRIBUTING.md and CODE_OF_CONDUCT.md
