### 2023-09-18 orbisgis/lczexplore#13 @MGousseff
Thank you very much for the improvement. I added to the .github a CONTRIBUTING.md and linked it to a CODE_OF_CONDUCT.md (the classical one inspired from Mozilla). I'm very grateful for all the improvements you suggested, it really helps me update my practices. Let me know if it is OK before I close the issue.
