### 2023-09-18 #5445 @MGousseff
Thank you for helping me improve my practice. I added contributing and code of conduct files. I let the issue opened so that you can confirm it's okay. Thank you for helping me improve my practices.
