### 2023-09-19 orbisgis/lczexplore#13 @wcjochem
Thanks for the addition. I'll make this complete now.
