### 2023-10-02 #5445 @MGousseff
@wcjochem Can I close the issue about contributing ?
@matthiasdemuzere, I have been self employed a decade ago, I wish you good luck for the redtape. Let me know if some modifications are needed.
