+++ R/importLCZvect.R
@@ -51,7 +51,9 @@ importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column
   colErr<-c("It seems that some of the columns you try to import do not exist in the source file,
             are you sure you meant ",
                  paste(badCol)," ?")
-  if (prod(inCol)==0){ stop(colErr) } else { sfFile<-sf::st_read(dsn=fileName,quiet=!verbose)[,colonnes] }
+  if (prod(inCol)==0){ stop(colErr) } else { 
+    if (drop== TRUE) {sfFile<-sf::st_read(dsn=fileName,quiet=!verbose)[,colonnes] } else {sfFile<-sf::st_read(dsn=fileName,quiet=!verbose)[,]}
+  }
 
   # if typeLevels is empty
   if (length(typeLevels)==1){
@@ -60,20 +62,16 @@ importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column
   }
 
   # if typeLevels is not specified it will be set to default and we need to capture this later
-  typeLevelsDefault<-c("1"="1","2"="2","3"="3","4"="4","5"="5","6"="6","7"="7","8"="8",
-                    "9"="9","10"="10","101"="101","102"="102","103"="103","104"="104",
-                    "105"="105","106"="106","107"="107","101"="11","102"="12","103"="13","104"="14",
-                    "105"="15", "106"="16","107"="17")
+  # typeLevelsDefault<-c("1"="1","2"="2","3"="3","4"="4","5"="5","6"="6","7"="7","8"="8",
+  #                   "9"="9","10"="10","101"="101","102"="102","103"="103","104"="104",
+  #                   "105"="105","106"="106","107"="107","101"="11","102"="12","103"="13","104"="14",
+  #                   "105"="15", "106"="16","107"="17")
 # Select columns from original file
-  if (column!=""){
-    if(drop==T){sfFile<-subset(sfFile,select=colonnes)}
-
-
- prov<-as.character(unique((st_drop_geometry(subset(sfFile,select=column,drop=T))))) %>% as.character
- names(prov)<-prov
-
-    if( prod(prov%in%typeLevels)==0 ){
-      if (verbose==TRUE){
+if (column!=""){
+   prov<-as.character(unique((st_drop_geometry(subset(sfFile,select=column,drop=T))))) %>% as.character
+  names(prov)<-prov
+  if( prod(prov%in%typeLevels)==0 ) {
+    if (verbose==TRUE){
         print("levels in typeLevels are : ")
         print(typeLevels)
         print("levels in original data set are ")
@@ -83,7 +81,7 @@ importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column
               Some geoms have been dropped,this could seriously alter your analysis, please check the levels or enter an empty string as typeLevels")
 
     }
-    if( sum(prov%in%typeLevels)==0 ){
+ if( sum(prov%in%typeLevels)==0 ){
       stop(
         paste0("none of the levels present in ",column,
                " is covered by the levels you specified.",
@@ -91,7 +89,6 @@ importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column
                " If you let typeLevels set by default, ", column,
                " must contain LCZ types in a standard format"))
     }
-
 
     sfFile <-
     sfFile%>%
@@ -104,8 +101,9 @@ importLCZvect<-function(dirPath, file="rsu_lcz.geojson", output="sfFile", column
[…]
