### 2023-10-06 39548b9
geoClimateConfigFile creates a json config file for geoclimate with the user agruments geoClimateCall feeds the system with a command launching the specified jar file and feeding it the  specified config file and workflow
