### 2023-10-06 a7116d1
Just draft of the needed structure to code a shiny interface in order to call GeoClimate with OSM Data.
