### 2023-10-09 #5445 @wcjochem
> @wcjochem Can I close the issue about contributing ?

@MGousseff, I believe I closed that issue earlier after you made the change. Because I included a link between that issue and this thread you can see in this thread (above) that the issue is closed in your repository. Thanks for the update and your work to resolve these issues.
