+++ NAMESPACE
@@ -5,6 +5,8 @@ export(areColors)
 export(compareLCZ)
 export(confidSensib)
 export(fetchLCZ)
+export(geoClimateCall)
+export(geoClimateConfigFile)
 export(groupLCZ)
 export(importLCZgc)
 export(importLCZraster)
@@ -35,6 +37,8 @@ importFrom(ggplot2,geom_sf)
 importFrom(ggplot2,ggtitle)
 importFrom(ggplot2,guides)
 importFrom(grDevices,palette.colors)
+importFrom(jsonlite,toJSON)
+importFrom(jsonlite,unbox)
 importFrom(magrittr,"%>%")
 importFrom(stats,quantile)
 importFrom(terra,as.polygons)
