### 2023-10-09 b9ac649
Added to the Shiny App the choice of the folder in which GeoClimate outputs its results
