### 2023-10-13 dc0d3c9
Done : outFolder fed to geoClimateFile but the default value parameter not fixed yet
