+++ R/shinyGC/app.R
@@ -15,20 +15,25 @@ ui <- fluidPage(
     sidebarPanel(
       shinyDirButton(id = "outFolder", 
                     label = "Output  folder for GeoClimates results", 
-                    title = "Select the folder where GeoClimates will output its results", FALSE),
+                    title = "Select the folder where GeoClimates will output its results"),
+      
       selectInput(inputId="wf", label="Workflow", 
                   choices = list(OpenStreetMap="OSM","BD TOPO V2"="BDTOPO_V2.2","BD TOPO V3"="BDTOPO_V3"),
                   selected=list(OpenStreetMap="OSM")),
+      
       conditionalPanel(
         condition='input.wf!="OSM"',
-        shinyDirButton("inBDTfolder",
+        shinyDirButton("BDTinFolder",
                        label = "BD_TOPO folder",
-                       title = "Choose in which folder are the BD_TOPO files"),
+                       title = "Choose in which folder are the BD_TOPO files",FALSE),
         numericInput(inputId="inseeCode", label="Enter Insee code of your location (town)", value = "29031")
       ),
-      textInput(inputId="location",label="Enter your locations here",
-                value="Redon",placeholder="A town name or some coordinates")
       
+      conditionalPanel(
+        condition='input.wf=="OSM"',
+        textInput(inputId="location",label="Enter your locations here",
+                  value="Redon",placeholder="A town name or some coordinates")
+      )      
     ),
 
     # Main panel for displaying outputs ----
@@ -45,46 +50,77 @@ ui <- fluidPage(
 
 # Define server logic required to draw a histogram ----
 server <- function(input, output) {
-
-
-  # Choose the folder where results of geoclimate will be put
+# Choose the folder where results of geoclimate will be put
     # set the path to the button in the interface (input outfolder)
-  shinyFiles::shinyDirChoose(input, 'outFolder', roots=getVolumes()(), 
-                             defaultPath = "/" )
+  
   # set a default value
-  global<-reactiveValues(datapath = "/tmp")
+  # global<-reactiveValues(datapath = "/tmp")
   # get what the interface outputs (nothing before user action) and put it in a reactive value
-  outFolder<-reactive(input$outFolder)
+  # outFolder<-reactive(input$outFolder)
   
-  output$outFolder<-renderText({
-    global$datapath
-  })
-  # Now when the user sets the value in the interface : 
-  observeEvent(ignoreNULL = TRUE,
-               eventExpr = {
-                 input$outFolder
-               },
-               handlerExpr = {
-                 if (!"path" %in% names(dir())) return()
-                 home <- normalizePath("~")
-                 global$datapath <-
-                   file.path(home, paste(unlist(outFolder()$path[-1]), collapse = .Platform$file.sep))
-               })
+  # output$outFolder<-renderText({
+  #   global$datapath
+  # })
+
+  shinyFiles::shinyDirChoose(input, 'outFolder', roots=getVolumes()(),
+                             defaultPath = "", allowDirCreate = TRUE )
+  # observe({
+  #   cat("\ninput$outFolder value:\n\n")
+  # print(input$outFolder)})
+  # Now when the user sets the value in the interface : (I'm a script kiddie, got this on SO and do not fully understand it now, shame to be corrected
+  # observeEvent(ignoreNULL = TRUE,
+  #              eventExpr = {
+  #                input$outFolder
+  #              },
+  #              handlerExpr = {
[…]
