### 2023-10-13 cfff859
Shiny interface now allows to create your configuration file (some les frequently used parameters are still to be added), to choose where your geoclimate jar file is, and to run it on the config file you've created.
TODO :
- add unit tests for geoClimateConfigFile and geoClimateCall functions
- add a tab to visualize the output (and compare to OSM map on the same zone)
- add a tab to compare two outputs
