+++ R/geoClimateConfigFile.R
@@ -19,20 +19,21 @@
 #' "VEGETATION_FRACTION",
 #' "ROAD_FRACTION",
 #' "IMPERVIOUS_FRACTION","LCZ_PRIMARY","LCZ_FRACTION","UTRF")
-#' @param outDir is the folder were the resulting JSON file will be put, the folder where GeoClimate will read it from
+#' @param outConfigDir is the folder were the resulting JSON file will be put, the folder where GeoClimate will read it from
 #' (different from out outFolder, where GeoClimates will put its geoJSON ouputs), default is "/tmp"
-#' @param outFile is the name of your configuration file, if anb empty string, a name woill be created from location
+#' @param outConfigFile is the name of your configuration file, if anb empty string, a name woill be created from location
 #' and workflow parameters. 
+#' @param forceSRID some BD TOPO input file may not have an srid, this forces srid to be 2154 
 #' @importFrom jsonlite unbox toJSON
 #' @return returns a JSON configuration file to be fed to GeoClimate
 #' @export
 #' @examples
-#' test<-geoClimateConfigFile(outFile="", wf="osm",outFolder="/tmp",locations="Redon",
+#' test<-geoClimateConfigFile(outConfigFile="", wf="osm",outFolder="/tmp",locations="Redon",
 #' rsuIndics = c("LCZ","TEB","UTRF"),
 #' gridIndics = c("BUILDING_FRACTION","BUILDING_HEIGHT","WATER_FRACTION","VEGETATION_FRACTION","ROAD_FRACTION",
 #' "IMPERVIOUS_FRACTION","LCZ_PRIMARY","LCZ_FRACTION","UTRF"))
 #' test
-geoClimateConfigFile<-function(locations, wf,
+geoClimateConfigFile<-function(wf,locations,forceSRID=FALSE, 
                                outFolder="/tmp",
                                rsuIndics=c("LCZ","TEB","UTRF"),
                                svSimplified = TRUE,
@@ -50,36 +51,56 @@ geoClimateConfigFile<-function(locations, wf,
                                             "LCZ_PRIMARY",
                                             "LCZ_FRACTION",
                                             "UTRF"),
-                               outDir = "/tmp",
-                               outFile="",
-                               BDTinFolder="",BDTinseeCode=29301) {
-  print(wf)
-
+                               outConfigDir = "/tmp",
+                               outConfigFile="configFile",
+                               BDTinFolder="",BDTinseeCode=29301,
+                               writeNow=FALSE) {
   description<-"Test de description unique"
   
-    # if (wf=="OSM"){description<-"Processing OSM data"} else {
-    # if (wf=="BDTOPO_V2.2") {description<-"Processing BDTopo v2 data"} else {
-    # if (wf=="BDTOPO_V3") {
-    #   description<-"Processing BDTopo v2 data"}
-    #  else { description<-paste0("Processing on an unrecognized workflow: ", wf)
-    # }
-    # }
-    # }
+    if (wf=="OSM"){description<-"Processing OSM data"} else {
+    if (wf=="BDTOPO_V2.2") {description<-"Processing BDTopo v2 data"} else {
+    if (wf=="BDTOPO_V3") {
+      description<-"Processing BDTopo v3 data"}
+     else { description<-paste0("Processing on an unrecognized workflow: ", wf)
+    }
+    }
+    }
 
-
-  
-# description<-list(description = description)
-# input<-list(input=list(locations=locations))
-
-outFolder<-list(folder=unbox(outFolder))
[…]
