+++ R/shinyGC/app.R
@@ -1,130 +1,176 @@
 library(shiny)
 library(shinyFiles)
 library(lczexplore)
+library(magrittr)
 
 # Define UI for app that draws a histogram ----
 ui <- fluidPage(
-
-  # App title ----
-  titlePanel("Run Geoclimate on OSM Data"),
+  
+  tabsetPanel(
+  
+    tabPanel("Create your  GeoClimate configuration JSON file",
 
   # Sidebar layout with input and output definitions ----
-  sidebarLayout(
+      sidebarLayout(
     
     # Sidebar panel for inputs ----
-    sidebarPanel(
-      shinyDirButton(id = "outFolder", 
+        sidebarPanel(
+          shinyDirButton(id = "outFolder", 
                     label = "Output  folder for GeoClimates results", 
                     title = "Select the folder where GeoClimates will output its results"),
       
-      selectInput(inputId="wf", label="Workflow", 
-                  choices = list(OpenStreetMap="OSM","BD TOPO V2"="BDTOPO_V2.2","BD TOPO V3"="BDTOPO_V3"),
+          selectInput(inputId="wf", label="Workflow", 
+                  choices = list(OpenStreetMap="OSM","BD TOPO V2"="BDTOPO_V2","BD TOPO V3"="BDTOPO_V3"),
                   selected=list(OpenStreetMap="OSM")),
       
-      conditionalPanel(
-        condition='input.wf!="OSM"',
-        shinyDirButton("BDTinFolder",
+          conditionalPanel(
+            condition='input.wf!="OSM"',
+          shinyDirButton("BDTinFolder",
                        label = "BD_TOPO folder",
                        title = "Choose in which folder are the BD_TOPO files",FALSE),
-        numericInput(inputId="inseeCode", label="Enter Insee code of your location (town)", value = "29031")
-      ),
+            checkboxInput("forceSRID",label="Force SRID of BD TOPO inputs to 2154",value=FALSE),
+            textInput(inputId="inseeCode", label="Enter Insee code of your location (town)", value = "29031")
+          ),
       
-      conditionalPanel(
-        condition='input.wf=="OSM"',
-        textInput(inputId="location",label="Enter your locations here",
-                  value="Redon",placeholder="A town name or some coordinates")
-      )      
+          conditionalPanel(
+            condition='input.wf=="OSM"',
+            textInput(inputId="location",label="Enter your locations here",
+                  value="Allaire",placeholder="A town name or some coordinates")
+          ),
+
+        checkboxGroupInput(inputId="rsuIndics",label = "Choose the indicators to compute at RSU scale",
+             choices=c("LCZ","TEB","UTRF"),selected=c("LCZ")),
+        
+        checkboxGroupInput(inputId="gridIndics",label = "Choose the indicators to compute at grid scale",
+             choices=c("BUILDING_FRACTION",
+                       "BUILDING_HEIGHT",
+                       "WATER_FRACTION",
+                       "VEGETATION_FRACTION",
+                       "ROAD_FRACTION",
+                       "IMPERVIOUS_FRACTION",
+                       "LCZ_FRACTION"),
+             selected=c("BUILDING_FRACTION"
+                        )),
+        shinyDirButton("configDirOut",
+                         label = "Folder to export config File",
+                         title = "Choose in which folder to export the config file",FALSE),
+          textInput(inputId="configOutFile",
+                    label="Name your configuration file (without extension)",
+                    value=""),
[…]

[…]
