### 2023-10-16 12e6f54
Shows the result once GC ran the config file. TO DO : add the OSM map for the same area.
