### 2023-10-17 #5445 @matthiasdemuzere
**-Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.**

There is an extensive list of dependencies. But related to the comment above, perhaps `png` and `knitr` should be listed there as well?  Preferably, the package installs all the dependencies by default, to easen the life of a user? See also comment [right above](https://github.com/openjournals/joss-reviews/issues/5445#issuecomment-1766310610) this one.
