### 2023-10-17 #5445 @matthiasdemuzere
**- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**

I don't see any community guidelines?
