### 2023-10-17 #5445 @MGousseff
Dear @matthiasdemuzere, Thank you for finding the time to review again. I am so sorry this problem of vignettes makes you lose time. It is not the same error message as last time, and I can't reproduce it on my computer. I am currently on the same ubuntu version so I can't see what is the hiatus. I am setting a virtualbox image to check from start and see if I can reproduce the bug. I'll tell you as soon as I find a fix if I can reproduce. Sorry for this and thank you for the time spent.
