### 2023-10-17 #5445 @MGousseff
> **- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
>
> I don't see any community guidelines?

They are in the .github folder. I was told it is standard practice to put them there, but I can move them to the package root if you find it more convenient ?
