### 2023-10-17 #5445 @MGousseff
> Error: processing vignette 'lczexplore_en.Rmd' failed with diagnostics:
>    X11 font -adobe-helvetica-%s-%s-*-*-%d-*-*-*-*-*-*-*, face 1 at size 9 could not be loaded
>    --- failed re-building

I have no idea where this comes from as I don't specify the font anywhere. I wonder if it could come from the fact that there are bold fonts somewhere in the vignette ?
