### 2023-10-17 #5445 @martinfleis
@MGousseff Explicit files in the root are better in my opinion but either way is fine.

Not sure about the .github model but it is, from my perspective, wrong.
