### 2023-10-17 #5445 @matthiasdemuzere
**- Software paper: Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?**

Following my comments on `Summary` and `A statement of need`, perhaps you can critically look into the structure again? You currently try to convey two messages (LCZ and general map intercomparisons), which does not help the general story line?

I also noted some typos on places, or have other textual suggestions, eg.
- L21: As stated in REF comparing map**s** (add s)
- L30: remove "The terms rural and urban lack of a clear definition and different classifications of urban and rural landscapes were proposed"
- L31: LCZ explanation as such does not provide link to UHI in L29. LCZ provice a research framework for urban heat island studies and standardizes the worldwide exchange of urban temperature observations ...
- L35: The use of LCZs has gone well beyond UHI, see eg. section 4 of [Demuzere et al. (2022)](https://doi.org/10.5194/essd-14-3835-2022) or section 4.8 in [Huang et al. (2023)](10.1016/j.rse.2023.113573)
- L40: the most comprehensive review on LCZ mapping to date is from [Huang et al. (2023)](10.1016/j.rse.2023.113573). Might be worthwhile to integrate this as well.
- L42: I don't really agree with "For instance, the WUDAPT platform produced LCZ maps of Europe and North-America this way". As there is no such thing as "the wudapt platform". In my opinion, there is not even thé WUDAPT method. So what about: "In this way, the WUDAPT community (Ching et al., 2018) not only produced thousands of city-based LCZ maps (accessible via the LCZ Generator (Demuzere et al., 2021)), but also large-scale maps for Europe, the continental United States and the whole world (Demuzere et al., [2019](http://doi.org/10.1371/journal.pone.0214474), [2020](http://dx.doi.org/10.1038/s41597-020-00605-z), [2022](https://doi.org/10.5194/essd-14-3835-2022))".
- L62: ... two main drawbacks**:** (add:)
- L68: remobe brackets around reference, with year in ()? You seem to have this in other places as well, please check!
- L69: I do not understand the sentence "Each pixel is seen as an individual to which each map assignes a value ..."
- L73: is an example of **a** tool (add a) that **provides** (instead of allows) multiple methods ...
- L78: "... the value whose area is the most present in the
pixel ..." -> rephrase in interest of clarity?
- L81: ... compare (remove s) some model output ...
- L83: the review of [Huang et al. (2023)](10.1016/j.rse.2023.113573) also provides a comprehensive overview of accuracy assessment.
- L97: "the need for automation of vector comparison" -> yet your package also allows to vector <-> raster comparison?
- L108: allow the following steps of exploration: (add :)
- L113: a space too much between compared and :
- L149: comparing LCZ **maps**
- L177: a space too much between question and :
- L185: change to "... increases the agreement to more than 90%"
