### 2023-10-17 #5445 @matthiasdemuzere
**-Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

_(apologies for no longer following the order of the checklist, this will be the last one I address)_

There is a good minimal example, that showcases the capacity of the tool using the data that was also used / produced in [Bernard et al. (2023)](https://doi.org/10.5194/egusphere-2023-371)

Now, as an additional stress-test, I'll check whether I can use some of the modules to compare two unknow LCZ maps, just to see if all works as expected. I'll try to look into this this evening, and will share my experiences here.
