### 2023-10-17 #5445 @MGousseff
I will look in details all this modifications. I didn't manage to reproduce the font problem on a virtual machine nor on another ubuntu machine. But the vignettes are not the vital part of the package. Maybe I will wait review from  @wcjochem  to integrate all modification in a single pull request ?
