### 2023-10-17 #5445 @matthiasdemuzere
**-Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

_(apologies for no longer following the order of the checklist, this will be the last one I address)_

There is a good minimal example, that showcases the capacity of the tool using the data that was also used / produced in [Bernard et al. (2023)](https://doi.org/10.5194/egusphere-2023-371)

Now, as an additional stress-test, I'll check whether I can use some of the modules to compare two other LCZ maps, just to see if all works as expected. I'll try to look into this this evening, and will share my experiences here.

Ok, so I extracted two LCZ maps for the area of Rennes (France), from:
1) The LCZ Generator
2) The most recent global LCZ map (version 3)

For both maps, I selected the filtered LCZ band. The files are here: [lcz_maps.zip](https://github.com/openjournals/joss-reviews/files/12945789/lcz_maps.zip)

Let me try to read one map, for which I need `importLCZraster`.

There is documentation in there, though with quite some typos and honestly not very clear?
But I see that I need the following usage:

```
     importLCZraster(
       dirPath,
       zone = "europe",
       bBox,
       fileName = "EU_LCZ_map.tif",
       column = "EU_LCZ_map",
       typeLevels = c(`1` = "1", `2` = "2", `3` = "3", `4` = "4", `5` = "5", `6` = "6", `7` =
         "7", `8` = "8", `9` = "9", `10` = "10", `101` = "11", `102` = "12", `103` = "13",
         `104` = "14", `105` = "15", `106` = "16", `107` = "17")
     )
```
More or less clear, except how to get `bBox`? In the docs:

    bBox: bBox is the bounding box needed to crop the wudapt tiff file.
          It can be produced bu the importLCZvect function. It can
          either be of class bBox or of class sfc

Mmm, ok. Not very clear? After some googling (remember I don't work with R), I decided to use something roughly based on the `sanDiegobBoxCoord` example (why use the CONUS map in this part whilst all other references refer to the EU map?):

```
RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.93,47.95)), st_point(c(-1.41,48.29)),crs = 4326))
RennesBbox<-st_bbox(RennesBoxCoord)
```

So far so good.

Then I read the map:

```
LCZ1<-importLCZraster(
  dirPath,
  fileName="9bce5ae7f3e297ebad75e6577fe1030b0d3c6ae6_lczfilter.tif",
  column="lczFilter",
  bBox=RennesBbox
)
```

with and without defining `typeLevels`, as the role of this is not very clear?
Seems to work, and the labels are changed to 1-10,101-107 instead of the original 1-17 in the .tif

Then I try to plot this map:

`showLCZ(LCZ1,column="lczfilter")`

and get an error:

```
Error in `fct_recode()`:
! `.f` must be a factor or character vector, not a <NULL> object.
Run `rlang::last_trace()` to see where the error occurred.
Warning message:
In structure(x, class = setdiff(class(x), "sf")) :
  Calling 'structure(NULL, *)' is deprecated, as NULL cannot have attributes.
  Consider 'structure(list(), *)' instead.
```

Mmm ... ok.

So let's try the cropped area from the global map.

```
LCZ2<-importLCZraster(
  dirPath,
  fileName="lcz_filter_globalmap.tif",
  column="LCZ_Filter",
  bBox=RennesBbox
)
showLCZ(LCZ2,column="LCZ_Filter")
```

Now I got another error:

```
Error in grid.Call(C_textBounds, as.graphicsAnnot(x$label), x$x, x$y,  :
  invalid use of -176 < 0 in 'X11_MetricInfo'
```

That's bad luck.

So I tried again, using the EU map you refer to in the `importLCZraster`description.

```
LCZ_EU<-importLCZraster(
  dirPath,
  zone = "europe",
  bBox=RennesBbox,
  fileName = "EU_LCZ_map.tif",
  column = "EU_LCZ_map",
  typeLevels = c("1" = "1", "2" = "2", "3" = "3", "4" = "4", "5" = "5", "6" = "6", "7" =
                   "7", "8" = "8", "9" = "9", "10" = "10", "101" = "11", "102" = "12", "103" = "13",
                 "104" = "14", "105" = "15", "106" = "16", "107" = "17")
)
showLCZ(LCZ_EU,column="EU_LCZ_map")
```

That gives me the same error:

```
Error in grid.Call(C_textBounds, as.graphicsAnnot(x$label), x$x, x$y,  :
  invalid use of -176 < 0 in 'X11_MetricInfo'
```

Ok, none of this really worked for me.

But perhaps it is just local settings that does not allow me to plot attempt 2 and 3. So let's try to compare two maps:

```
comparison<-compareLCZ(sf1=LCZ2,column1="LCZ_Filter", wf1="Global map",
                       sf2=LCZ1, column2="lczFilter", wf2="Generator map",
                       ref=1, repr="standard", exwrite=F, location="Rennes")
```

I get the text `The column  LCZ_Filter  of the dataset LCZ2 is the reference against which the  lczFilter  column of the dataset  LCZ1 will be compared.` and then it keeps running. Already for over an hour now, so not sure whether this is still leading to something?
