### 2023-10-17 #5445 @matthiasdemuzere
@MGousseff @martinfleis

Long story short on the above tests using some other datasets: I did not really get anywhere.
And it does require some creativity from the user to understand what is needed.

My feeling about this work: it is clear that the routines have been developed for a very specific purpose, namely to support the research done for [this manuscript](https://egusphere.copernicus.org/preprints/2023/egusphere-2023-371/). The minimal examples with the datasets from this effort seems to work well.

Then these routines are brought together and advertised here as an R package, not only to explore the difference between LCZ classifications, but also "any pair of classifications on geographical units" (as stated in the Summary).

After some testing **I am not convinced that this bundle of scripts is mature enough to go forward with the review process in JOSS**. I believe the routines in lczexplore can only be valuable and widely used if users can easily read and assess their own maps, without fiddling with the code to get the data/labels right. Of course also key functions should work (eg. compareLCZ), which is not the case for me, even though I am using common LCZ map databases.
