### 2023-10-18 #5445 @MGousseff
For this error :
> Then I try to plot this map:
> showLCZ(LCZ1,column="lczfilter")

I think it comes from the fact that when you import the file the column is named :  `column="lczFilter"`, and when you try to show it you call `showLCZ(LCZ1,column="lczfilter")` with a lower case  f ?
