### 2023-10-18 #5445 @wcjochem
> I will look in details all this modifications. I didn't manage to reproduce the font problem on a virtual machine nor on another ubuntu machine. But the vignettes are not the vital part of the package. Maybe I will wait review from @wcjochem to integrate all modification in a single pull request ?

@MGousseff I already completed my checklist and have raised no new issues. From that perspective, my review is complete and I will not be monitoring this thread further. I think Matthias has raised valid concerns, and I leave that to an editorial decision from @martinfleis.
