### 2023-10-18 #5445 @MGousseff
I will try and reproduce the example you gave to day and see if I can reproduce bugs.
I agree than I am more used to deal with vector maps than raster maps, and the issue you raised comes from an error in the man file. There is no column to specify for a raster file, the name of the LCZ column is derived from the name of the file, it's an error I'll promptly correct. The following code works.

```
library(lczexplore)
library(sf)

pathEuWudaptMap<-"/home/gousseff/Documents/3_data/WUDAPTSources/WudaptEurope"
fileNameWudaptEurope<-"EU_LCZ_map.tif"

RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.93,47.95)), st_point(c(-1.41,48.29)),crs = 4326))
RennesBbox<-st_bbox(RennesBoxCoord)

LCZ1<-importLCZraster(
  dirPath=pathEuWudaptMap,
  fileName=fileNameWudaptEurope,
  bBox=RennesBbox
)

showLCZ(LCZ1,column="EU_LCZ_map")
```

For the bounding box, what would seem an acceptable way to define it ? I tried to avoid calls to API as we have seen many sites modify their policies towards this. I confess that my workflow starts with vector reference, so producing the bbox from a vector reference map IS an oriented way of seeing this. A very simple roundabout is to use the getbb function from osmdata, which is quite commonly used by the R-spatial community. This would give the follwing code, which works very easily :

```
library(lczexplore)
library(sf)
library(osmdata)

pathEuWudaptMap<-"/home/gousseff/Documents/3_data/WUDAPTSources/WudaptEurope"
fileNameWudaptEurope<-"EU_LCZ_map.tif"

rennesBbox<-getbb( "Rennes",format_out = "sf_polygon")

LCZ1<-importLCZraster(
  dirPath=pathEuWudaptMap,
  fileName=fileNameWudaptEurope,
  bBox=RennesBbox
)

showLCZ(LCZ1,column="EU_LCZ_map")

```
I am still reading your comment about redaction and will integrate them today.
