### 2023-10-18 #5445 @matthiasdemuzere
> The following code works.
>
> ```
> library(lczexplore)
> library(sf)
>
> pathEuWudaptMap<-"/home/gousseff/Documents/3_data/WUDAPTSources/WudaptEurope"
> fileNameWudaptEurope<-"EU_LCZ_map.tif"
>
> RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.93,47.95)), st_point(c(-1.41,48.29)),crs = 4326))
> RennesBbox<-st_bbox(RennesBoxCoord)
>
> LCZ1<-importLCZraster(
>   dirPath=pathEuWudaptMap,
>   fileName=fileNameWudaptEurope,
>   bBox=RennesBbox
> )
>
> showLCZ(LCZ1,column="EU_LCZ_map")
> ```

Unfortunately not for me. I get the following graphic:
![image](https://github.com/openjournals/joss-reviews/assets/6926916/21835959-6bd0-48bf-ad92-48a774e4bd2a)

With this warnings in the terminal:

```
In grid.Call.graphics(C_path, x$x, x$y, index, switch(x$rule,  ... :
  Path drawing not available for this device
```

In Rstudio, I get the error on `X11_MetricInfo`.

Both errors might be related to the configuration of my OS itself?
That of course not necessarily reflects the quality of your scripts.

Yet at the same time it also does not adhere to `[the guidelines of JOSS in terms of installation](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions)`?
