### 2023-10-18 #5445 @matthiasdemuzere
> I think it comes from the fact that when you import the file the column is named : `column="lczFilter"`, and when you try to show it you call `showLCZ(LCZ1,column="lczfilter")` with a lower case f ?

Indeed, my mistake.

Yet I get the same error as above when plotting, with an empty map.
