### 2023-10-18 #5445 @MGousseff
I tried with `RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.93,47.95)), st_point(c(-1.41,48.29)),crs = 4326))
RennesBbox<-st_bbox(RennesBoxCoord)` and with `library(osmdata) ; rennesBbox<-getbb( "Rennes",format_out = "sf_polygon") ` and it works in both cases. I guess you changed
```pathEuWudaptMap<-"/home/gousseff/Documents/3_data/WUDAPTSources/WudaptEurope"
fileNameWudaptEurope<-"EU_LCZ_map.tif"
```
to include your path ?

The very strange thing is that we are both on Ubuntu. Hard to believe the locale settings would have such an impact ? I'll try on yet another computer this morning.
