### 2023-10-18 #5445 @MGousseff
I tried on another PC, which runs on windows, and I expected to be problematic. To my surprise it worked with the three ways to build the bBox :  with a bBox built with the coordinates and the sf package, with the call to getbb from the package osmdata, and with a bBox built from a vector file using importLCZvect on an area a asking for bBox as an output. I fed these bBox to `importLCZraster` with the EU map available [here](https://figshare.com/articles/dataset/European_LCZ_map/13322450) as the raster source to crop, and in all three cases, the import and the plot worked seamlessly. I'm a bit at a loss as to why it doesn't work on your computer.
