### 2023-10-18 #5445 @MGousseff
I am actively integrating the corrections you proposed. I will make a pull request with the corrections of the man page, the modifications of the article. The only corrections I don't understand is the problem you stated with the reference which didn't seem to appeared bugged on my version of the paper.pdf.
