### 2023-10-18 5931713
rather functional shiny app to build a config file and rune GeoClimate.
