### 2023-10-18 #5445 @MGousseff
I think it should link to [this paper](https://www.sciencedirect.com/science/article/pii/S0034425723001244) that I''m now currently reading. Thank you for the ref.
