### 2023-10-18 #5445 @ebocher
@MGousseff

> I think it should link to [this paper](https://www.sciencedirect.com/science/article/pii/S0034425723001244) that I''m now currently reading. Thank you for the ref.

We are talking about a paper published after this one ? It's an unusual suggestion.

@matthiasdemuzere

> Long story short on the above tests using some other datasets: I did not really get anywhere.
And it does require some creativity from the user to understand what is needed.

A long story and a long waiting review in balance ;-) . Let me tell you : If there was an error when comparing with raster data, due to the documentation, LZCExplorer is used without any issues to compare vector data with urban planners, students and colleagues.

But in general, I'm more intrigued by your position on the story line and the usefulness of the tool.

> Now, I realize we already had an extensive discussion on the need for a tool to compare LCZ maps.
I don't want to go back there, and will leave it up to the user community to decide whether this is useful.

Comparing model outputs, and in our case maps, is a natural approach in GIS science if you want to be able to assess the relevance of a method. Despite WUDAPT's long experience in producing LCZs, there's nothing to suggest that this method is any better than maps produced by human or by any other tool. The opposite is true. ;-)

I hope @MGousseff  will have the energy to include all the new comments.

We are therefore aware of the limited technical contribution of this tool, but its use corresponds to a real need for evaluating different LCZ classifications.

At the end , as @wcjochem wrote  "leave that to an editorial decision from @martinfleis" will be the best option.

Best regards

@j3r3m1
