+++ docs/articles/joss/paper.md
@@ -45,9 +45,11 @@ editor_options:
 
 # Summary
 
-Climate change is a growing concern for city planners as Urban Heat Islands have an impact on mortality [@clarke1972some], 
-health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
-A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, 
+Climate change is a growing concern for city planners as Urban Heat Islands have an impact on 
+mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy 
+for building cooling [@malys2012microclimate] among other effects.
+A first step towards large scale study of urban climate is to define classes 
+based on logical division of the landscape, 
 such as Local Climate Zones (LCZ) defined by [@stewart2012local].
 
 The lczexplore package aims at comparing different LCZ classifications, 
@@ -61,7 +63,7 @@ This software is available as a free and opensource R package.
 # Statement of need
 
 ## Comparing maps
-As stated in [@visser2006map] comparing map is an important issue in environmental research. 
+As stated in [@visser2006map] comparing maps is an important issue in environmental research. 
 The four main reasons to compare categorical variables on geographical units are:     
 - to assess the differences between maps generated by models under different scenarios and assumptions,  
 - to detect temporal changes,  
@@ -84,8 +86,10 @@ apprehend the intensity of the Urban Heat Island [@kotharkar2018evaluating].
 Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification for any given area.  
 [@quan2021systematic] distinguishes two main streams of production of these LCZ:      
 - the raster stream processes remotely sensed information, and applies machine learning 
-algorithms trained using local experts' knowledge. For instance, the WUDAPT platform produced 
-LCZ maps of Europe and North-America this way [@chingWUDAPTUrbanWeather2018a].   
+algorithms trained using local experts' knowledge. In this way, the WUDAPT community [@chingWUDAPTUrbanWeather2018a] 
+produced thousands of city-based LCZ maps (accessible via the LCZ Generator ([@demuzereLCZGeneratorWeb2021])) 
+but also large-scale maps for Europe, the continental United States and the whole world ([@demuzere2019mapping],
+[@demuzere2020combining], [@demuzere2022global]).   
 - the vector stream uses Geographic Information System (GIS) layers that represent the main topographic features, 
 defines spatial units, computes urban canopy parameters and uses them to classify spatial units into LCZ.
 For instance, the GeoClimate geospatial toolbox produces LCZ classifications 
@@ -109,7 +113,7 @@ The question of going beyond the simple visual observation of differences betwee
 A first approach would be, on raster maps, to simply compute the agreement between two maps as the proportion 
 of pixels for which the two maps have the same value of the variable of interest. 
 This approach is often sufficient to help specialists compare pairs of
-maps, but it has two main drawbacks 
+maps, but it has two main drawbacks:  
 
  - two totally random maps won't have a value of agreement of zero, as some pixel values may agree by chance,  
  - only raster maps where pixels match perfectly (same size, not translated) can be treated, or 
@@ -117,22 +121,24 @@ some pre-treatment are needed (like nearest neighbour interpolation for instance
 
 To prevent the first drawback, [@monserudComparingGlobalVegetation1992] 
 proposed the use of Cohen's kappa coefficient of agreement for nominal scales [@cohenCoefficientAgreementNominal1960].
-Each pixel is seen as an individual to which each map assigns a value of a categorical variable 
-(each map is seen as a "rater").  
+Cohen's Kappa allow to assess how two raters agree or differ in the task of rating individual. 
+In our case, each pixel is seen as an individual, each map is seen as a rater and the value of the raster at each pixel 
+is seen as the rate.   
 
[…]
