+++ docs/articles/joss/paper.bib
@@ -236,3 +236,36 @@
   volume = {10},
   number = {1}
 }
+
+
+@article{demuzere2019mapping,
+  title={Mapping Europe into local climate zones},
+  author={Demuzere, Matthias and Bechtel, Benjamin and Middel, Ariane and Mills, Gerald},
+  journal={PloS one},
+  volume={14},
+  number={4},
+  pages={e0214474},
+  year={2019},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
+
+@article{demuzere2022global,
+  title={A global map of Local Climate Zones to support earth system modelling and urban scale environmental science},
+  author={Demuzere, Matthias and Kittner, Jonas and Martilli, Alberto and Mills, Gerald and Moede, Christian and Stewart, Iain D and van Vliet, Jasper and Bechtel, Benjamin},
+  journal={Earth System Science Data Discussions},
+  volume={2022},
+  pages={1--57},
+  year={2022},
+  publisher={G{\"o}ttingen, Germany}
+}
+
+@article{demuzere2020combining,
+  title={Combining expert and crowd-sourced training data to map urban form and functions for the continental US},
+  author={Demuzere, Matthias and Hankey, Steve and Mills, Gerald and Zhang, Wenwen and Lu, Tianjun and Bechtel, Benjamin},
+  journal={Scientific data},
+  volume={7},
+  number={1},
+  pages={264},
+  year={2020},
+  publisher={Nature Publishing Group UK London}
+}
