### 2023-10-18 6ce38f7
Examples corrected in inportLCZraster Typos corrected in paper.md Some typos corrected in the vignettes Several references added to paper.md
