+++ docs/articles/joss/paper.bib
@@ -154,36 +154,36 @@
   issn = {0003-0007, 1520-0477},
   doi = {10.1175/BAMS-D-16-0236.1},
   urldate = {2023-06-30},
-  }
-  
-  @article{cohenCoefficientAgreementNominal1960,
-    title = {A {{Coefficient}} of {{Agreement}} for {{Nominal Scales}}},
-    author = {Cohen, Jacob},
-    year = {1960},
-    month = apr,
-    journal = {Educational and Psychological Measurement},
-    volume = {20},
-    number = {1},
-    pages = {37--46},
-    issn = {0013-1644, 1552-3888},
-    doi = {10.1177/001316446002000104},
-    urldate = {2023-06-30},
-    langid = {english}
-  }
-  
-  @article{demuzereLCZGeneratorWeb2021,
-    title = {{{LCZ Generator}}: {{A Web Application}} to {{Create Local Climate Zone Maps}}},
-    shorttitle = {{{LCZ Generator}}},
-    author = {Demuzere, Matthias and Kittner, Jonas and Bechtel, Benjamin},
-    year = {2021},
-    month = apr,
-    journal = {Frontiers in Environmental Science},
-    volume = {9},
-    pages = {637455},
-    issn = {2296-665X},
-    doi = {10.3389/fenvs.2021.637455},
-    urldate = {2023-06-30},
-    }
+}
+
+@article{cohenCoefficientAgreementNominal1960,
+  title = {A {{Coefficient}} of {{Agreement}} for {{Nominal Scales}}},
+  author = {Cohen, Jacob},
+  year = {1960},
+  month = apr,
+  journal = {Educational and Psychological Measurement},
+  volume = {20},
+  number = {1},
+  pages = {37--46},
+  issn = {0013-1644, 1552-3888},
+  doi = {10.1177/001316446002000104},
+  urldate = {2023-06-30},
+  langid = {english}
+}
+
+@article{demuzereLCZGeneratorWeb2021,
+  title = {{{LCZ Generator}}: {{A Web Application}} to {{Create Local Climate Zone Maps}}},
+  shorttitle = {{{LCZ Generator}}},
+  author = {Demuzere, Matthias and Kittner, Jonas and Bechtel, Benjamin},
+  year = {2021},
+  month = apr,
+  journal = {Frontiers in Environmental Science},
+  volume = {9},
+  pages = {637455},
+  issn = {2296-665X},
+  doi = {10.3389/fenvs.2021.637455},
+  urldate = {2023-06-30},
+}
 
 @techreport{bernardGenericAlgorithmAutomatically2023,
   type = {Preprint},
@@ -235,4 +235,37 @@
   pages = {439--446},
   volume = {10},
   number = {1}
+}
+
+
+@article{demuzere2019mapping,
+  title={Mapping Europe into local climate zones},
+  author={Demuzere, Matthias and Bechtel, Benjamin and Middel, Ariane and Mills, Gerald},
+  journal={PloS one},
+  volume={14},
+  number={4},
+  pages={e0214474},
+  year={2019},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
+
+@article{demuzere2022global,
+  title={A global map of Local Climate Zones to support earth system modelling and urban scale environmental science},
+  author={Demuzere, Matthias and Kittner, Jonas and Martilli, Alberto and Mills, Gerald and Moede, Christian and Stewart, Iain D and van Vliet, Jasper and Bechtel, Benjamin},
+  journal={Earth System Science Data Discussions},
+  volume={2022},
+  pages={1--57},
+  year={2022},
+  publisher={G{\"o}ttingen, Germany}
+}
+
+@article{demuzere2020combining,
[…]
