### 2023-10-18 98637b1
Examples corrected in inportLCZraster Typos corrected in paper.md Some typos corrected in the vignettes Several references added to paper.md
