+++ docs/articles/joss/paper.md
@@ -1,59 +1,61 @@
 ---
-title: 'lczexplore : an R package to explore Local Climate Zone classifications'
+title: 'lczexplore: an R package to explore Local Climate Zone classifications'
 tags:
-- R SOFTWARE
-- climate
-- environment
-- GIS
-- spatial analysis
-- cities
-   
+  - R SOFTWARE
+  - climate
+  - environment
+  - GIS
+  - spatial analysis
+  - cities
+
 authors:
-  
-- name: Matthieu Gousseff
-  orcid: 0000-0002-7106-2677
-  equal-contrib: true
-  affiliation: 1
-- name: Erwan Bocher
-  orcid : 0000-0002-4936-7079
-  equal-contrib: true
-  affiliation: 1
-- name: Jérémy Bernard
-  orcid : 0000-0001-7374-5722
-  corresponding: true # (This is how to denote the corresponding author)
-  affiliation: "1, 2, 4"
-- name: Elisabeth Le Saux Wiederhold
-  orcid:
-  corresponding: true
-  affiliation: 3   
+
+  - name: Matthieu Gousseff
+    orcid: 0000-0002-7106-2677
+    equal-contrib: true
+    affiliation: 1
+  - name: Erwan Bocher
+    orcid: 0000-0002-4936-7079
+    equal-contrib: true
+    affiliation: 1
+  - name: Jérémy Bernard
+    orcid: 0000-0001-7374-5722
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: "1, 2, 4"
+  - name: Elisabeth Le Saux Wiederhold
+    orcid:
+    corresponding: true
+    affiliation: 3
 affiliations:
-  
-- name: CNRS, Lab-STICC, UMR 6285, Vannes, France
-  index: 1
-- name: University of Savoie Mont-Blanc, LOCIE, UMR 5271, France
-  index: 2
-- name: Université Bretagne Sud, Lab-STICC, UMR 6285, Vannes, France
-  index: 3
-- name : University of Gothenburg, Department of Earth Sciences, Sweden
-  index: 4
+
+  - name: CNRS, Lab-STICC, UMR 6285, Vannes, France
+    index: 1
+  - name: University of Savoie Mont-Blanc, LOCIE, UMR 5271, France
+    index: 2
+  - name: Université Bretagne Sud, Lab-STICC, UMR 6285, Vannes, France
+    index: 3
+  - name: University of Gothenburg, Department of Earth Sciences, Sweden
+    index: 4
 bibliography: paper.bib
 date: 01 april 2023
-editor_options: 
-  markdown: 
+editor_options:
+  markdown:
     wrap: sentence
 ---
 
 # Summary
 
-Climate change is a growing concern for city planners as Urban Heat Islands have an impact on mortality [@clarke1972some], 
-health in general [@lowe2016energy] and consumption of energy for building cooling [@malys2012microclimate] among other effects.
-A first step towards large scale study of urban climate is to define classes based on logical division of the landscape, 
+Climate change is a growing concern for city planners as Urban Heat Islands have an impact on
+mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy
+for building cooling [@malys2012microclimate] among other effects.
+A first step towards large scale study of urban climate is to define classes
+based on logical division of the landscape,
 such as Local Climate Zones (LCZ) defined by [@stewart2012local].
 
-The lczexplore package aims at comparing different LCZ classifications, 
+The lczexplore package aims at comparing different LCZ classifications,
 but can be used to compare any pair of classifications on geographical units.
 
-A spatial comparison is performed by producing agreement maps between classifications, agreement statistics and 
[…]
