+++ vignettes/lczexplore_en.Rmd
@@ -24,7 +24,7 @@ if (!require("png")) {
 ```
 
 
-# A package to explore Local Climate Zones : lczexplore
+# A package to explore Local Climate Zones: lczexplore
 ## Purpose and origin
 The lczexplore package was developed thanks to the PAENDORA2 (Pour la gestion du confort estival : Données, Outils et Recherche-Action, 2022 -2025, funded by ADEME, a French agency for ecological transition).
 
@@ -42,18 +42,18 @@ It has been extended to compare any pair of qualitative variable set on valid ge
 
 ## Installation
 
-The following allows the installation from the github repository : (`devtools::install_github("orbisgis/lczexplore", build_vignettes = TRUE)`).
+The following allows the installation from the github repository: (`devtools::install_github("orbisgis/lczexplore", build_vignettes = TRUE)`).
 
 On can also install from the tar.gz file and install it (see `?install.packages`).
 
 
-Load the package : `library(lczexplore)`.
+Load the package: `library(lczexplore)`.
 ```{r}
 library(lczexplore)
 ```
 
 
-# First use-case : load a file with geometries and standard LCZ
+# First use-case: load a file with geometries and standard LCZ
 
 ## Import files with importLCZvect function
 
@@ -111,7 +111,7 @@ showLCZ(sf=redonOSM,wf="OSM",column="LCZ_PRIMARY",repr="standard",drop=TRUE)
 ```
 ### Alternative visualization
 
-The non-standard representation, triggered with `repr=alter` is useful when : 
+The non-standard representation, triggered with `repr=alter` is useful when: 
 - one wants to **regroup** some LCZ types into more general categories,
 - the LCZ **don't have a standard encoding**
 - one want's to use **custom colors**
@@ -121,7 +121,7 @@ With `repr=alter`, one passes the vector of possible levels to the argument `LCZ
 
 If no vector of levels or no vector of colors is given, the levels will be deduced from the unique values present in the data, and the colors will be picked from the "Polychrome 36" palette of the `grDevices` package.
 
-For instance : one can first plot the column without knowing the LCZ levels present.
+For instance: one can first plot the column without knowing the LCZ levels present.
 ```{r, fig.dim = c(8, 6)}
 # Choice of random colors from a palette
 testCol <- palette.colors(n=17, palette="Polychrome 36")
@@ -166,10 +166,10 @@ showLCZ(redonOSMgrouped, column="grouped",repr="alter",wf="OSM",
 # Compare two LCZ classification with the function compareLCZ
 
 **The purpose of this package is to easily compare local climate zone classifications produced by different algorithms on the same study area.**
-In our example, the GeoClimate workflow was used with two different input data to produce two LCZ classifications : 
+In our example, the GeoClimate workflow was used with two different input data to produce two LCZ classifications: 
 the BD TOPO® v2.2, produced by the French National Institute of Geographic and Forest Information, and OpenStreetMap. 
 
-NOTE : The vignette "lczexplore_raster_vector" explores in detail another example with WUDAPT raster data. 
+NOTE: The vignette "lczexplore_raster_vector" explores in detail another example with WUDAPT raster data. 
 
 
 One has to call `compareLCZ` and feed it, for each dataset, 
@@ -180,12 +180,12 @@ you need to tell which dataset is the reference
 
 - A map of each LCZ classification is plotted.  
 - The function then intersects the geometries of both sf objects in order to get geometries where the **two classification either totally agree either totally disagree.** This agreement is plotted on a third graphics.  
-- At last, a confusion matrix (or agreement matrix) is computed : 
+- At last, a confusion matrix (or agreement matrix) is computed: 
[…]
