### 2023-10-18 #5445 @MGousseff
After retest : there wasn't any error in importLCZraster, not even in the examples. The arg column is the name you want to give to the output, not a column to look for in the tif input. I was worried as this would have been detected by the unit tests, but the tests are OK so far.

I added as an example a way to get the bounding box with osmdata package. It didn't seem useful to include a dedicated function as it is already present in this package.

About the example with the CONUS map, it seemed better to also put an example with data not included in the package, so that the user doesn't think only rasters included in the package could be used. Plus it showed WUDAPT maps were available worldwide !

I added the reference to american and world maps as suggested. I didn't add the Huang review, as it was published after the submission of our article  (and seems quite focused on remote sensor based methods). I checked after today's pull request, I can't see the problems you noted about the references. I hoped I'm looking at the right place.

About the structure of the paper,  I think it reflects the community which is most likely to use it : people comparing LCZ maps.  Yet, if a package allowed map comparison with ease, maybe lczexplore wouldn't be useful, hence the part about comparing maps, as a first subsection of the statement of need. By the way, I have to thank you, as your remarks made me add this section, and thus add the Cohen's Kappa as a better indicator of agreement than only the percentage of agreeing surface.

I hope I fixed all the typo you noted. The space before ":" is due to french habits ! Corrected in the paper *and* in the vignettes !

I moved the CODE_OF_CONDUCT.md and CONTRIBUTING.md to the root of the package. I understood why some R coders hid it in the `.github` : it raises a note in R Check, as they are not considered standard files at the root of an R package. I think we can accept a note (not a warning or an error) and have a far better visibility for these files.

As stated by the reviewers and our team, the scope of the package may seem narrow, but our collaborations with various  users make us think it is useful, helps gain productivity and insights. Therefore, it is probably up to JOSS and @martinfleis to decide whether it is worth being published in the journal.

Whatever the decision, I thank you @wcjochem @matthiasdemuzere @martinfleis @ebocher  for the improvements I could make thanks to all the remarks.
