### 2023-10-19 #5445 @matthiasdemuzere
> After retest : there wasn't any error in importLCZraster, not even in the examples. The arg column is the name you want to give to the output, not a column to look for in the tif input.

I see. That was not clear to me from the docs:

```
  column: indicates the name of the column containing LCZ values, all
          other
```

> I was worried as this would have been detected by the unit tests, but the tests are OK so far.

Reading the geotif is not the problem. So this piece of code works.

```
RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.97,48)), st_point(c(-1.40,48.1)),crs = 4326))
rennesBbox<-st_bbox(RennesBoxCoord)
#rennesBbox<-getbb( "Rennes",format_out = "sf_polygon") #

LCZ1<-importLCZraster(
  dirPath="/mnt/SSD2TB/data/FIGSHARE/LCZ_EU/v4",
  zone = "europe",
  bBox=rennesBbox,
  fileName = "EU_LCZ_map.tif",
  column = "EU_LCZ_map",
  typeLevels = c(`1` = "1", `2` = "2", `3` = "3", `4` = "4", `5` = "5", `6` = "6", `7` =
                   "7", `8` = "8", `9` = "9", `10` = "10", `101` = "11", `102` = "12", `103` = "13",
                 `104` = "14", `105` = "15", `106` = "16", `107` = "17")
)

```

LCZ1 is a `Simple feature collection with 47858 features and 1 field`.
But I can not plot this using `showLCZ(LCZ1,column="EU_LCZ_map")`

In Rstudio I get this error without a plot:
```
Error in grid.Call(C_textBounds, as.graphicsAnnot(x$label), x$x, x$y,  :
  invalid use of -176 < 0 in 'X11_MetricInfo'
```

In a terminal I get an empty plot with legend, and this error:

```
50: In grid.Call.graphics(C_path, x$x, x$y, index, switch(x$rule,  ... :
  Path drawing not available for this device
```

Others have identified this as well (eg. [here](https://github.com/r-spatial/sf/issues/1395)), though I don't really see a solution?

My R session info (the same in R and Rstudio):

```
> sessionInfo()
R version 4.3.0 (2023-04-21)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 22.04.3 LTS

Matrix products: default
BLAS/LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=nl_BE.UTF-8        LC_COLLATE=en_US.UTF-8     LC_MONETARY=nl_BE.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=nl_BE.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=nl_BE.UTF-8 LC_IDENTIFICATION=C

time zone: Europe/Brussels
tzcode source: system (glibc)

attached base packages:

other attached packages:
[1] osmdata_0.2.5         lczexplore_0.0.1.0002 sf_1.0-14

 [1] gld_2.6.6          gtable_0.3.4       ggplot2_3.4.4      htmlwidgets_1.6.2  devtools_2.4.5     remotes_2.4.2.1    processx_3.8.2     lattice_0.21-8     callr_3.7.3
[10] vctrs_0.6.4        tools_4.3.0        ps_1.7.5           generics_0.1.3     tibble_3.2.1       proxy_0.4-27       fansi_1.0.5        pkgconfig_2.0.3    Matrix_1.5-4
[19] KernSmooth_2.23-20 data.table_1.14.8  RColorBrewer_1.1-3 readxl_1.4.3       rootSolve_1.8.2.4  lifecycle_1.0.3    farver_2.1.1       compiler_4.3.0     stringr_1.5.0
[28] Exact_3.2          munsell_0.5.0      terra_1.7-55       codetools_0.2-19   httpuv_1.6.11      htmltools_0.5.6.1  usethis_2.2.2      DescTools_0.99.50  class_7.3-21
[37] later_1.3.1        pillar_1.9.0       crayon_1.5.2       urlchecker_1.0.1   tidyr_1.3.0        MASS_7.3-58.4      ellipsis_0.3.2     classInt_0.4-10    cachem_1.0.8
[46] sessioninfo_1.2.2  boot_1.3-28.1      mime_0.12          tidyselect_1.2.0   digest_0.6.33      mvtnorm_1.2-3      stringi_1.7.12     dplyr_1.1.3        purrr_1.0.2
[55] forcats_1.0.0      cowplot_1.1.1      fastmap_1.1.1      grid_4.3.0         expm_0.999-7       lmom_3.0           colorspace_2.1-0   cli_3.6.1          magrittr_2.0.3
[64] pkgbuild_1.4.2     utf8_1.2.3         e1071_1.7-13       withr_2.5.1        prettyunits_1.2.0  scales_1.2.1       promises_1.2.1     httr_1.4.7         cellranger_1.1.0
[73] memoise_2.0.1      shiny_1.7.5.1      miniUI_0.1.1.1     profvis_0.3.8      rlang_1.1.1        Rcpp_1.0.11        xtable_1.8-4       glue_1.6.2         DBI_1.1.3
[82] pkgload_1.3.3      rstudioapi_0.15.0  R6_2.5.1           fs_1.6.3           units_0.8-4
```

> I added as an example a way to get the bounding box with osmdata package. It didn't seem useful to include a dedicated function as it is already present in this package.

I tried to use `rennesBbox<-getbb( "Rennes",format_out = "sf_polygon")` when reading the EU map with the above script. But that process was still running after 15min, so I killed it. The `sf` approach does seem to work.

>
> About the example with the CONUS map, it seemed better to also put an example with data not included in the package, so that the user doesn't think only rasters included in the package could be used. Plus it showed WUDAPT maps were available worldwide !

I understand your reasoning. But also the EU map is not included in the package itself?
So why not:
- just add one example
- in the description just point to https://www.wudapt.org/lcz-maps/ that lists where wudapt raster maps can be downloaded?

>
> I added the reference to american and world maps as suggested. I didn't add the Huang review, as it was published after the submission of our article (and seems quite focused on remote sensor based methods). I checked after today's pull request, I can't see the problems you noted about the references. I hoped I'm looking at the right place.

I am not 100% sure what note you refer to? Probably my note on having the references that are embedded within a sentence in between (). Eg. As stated in (Visser & De Nijs, 2006) ...
This should be: As stated in Visser & De Nijs (2006). See the [JOSS citation guidelines](https://joss.readthedocs.io/en/latest/submitting.html#citations)

>
> About the structure of the paper, I think it reflects the community which is most likely to use it : people comparing LCZ maps. Yet, if a package allowed map comparison with ease, maybe lczexplore wouldn't be useful, hence the part about comparing maps, as a first subsection of the statement of need. By the way, I have to thank you, as your remarks made me add this section, and thus add the Cohen's Kappa as a better indicator of agreement than only the percentage of agreeing surface.
>

I am not sure if you now kept or removed the part about comparing non-LCZ maps?
In any case, given the packages "breaths" LCZ (package name, its title, functions names, examples etc), perhaps just focus on LCZ mapping only, to simplify the narrative?

> I hope I fixed all the typo you noted. The space before ":" is due to french habits ! Corrected in the paper _and_ in the vignettes !
>
Ok.

> I moved the CODE_OF_CONDUCT.md and CONTRIBUTING.md to the root of the package. I understood why some R coders hid it in the `.github` : it raises a note in R Check, as they are not considered standard files at the root of an R package. I think we can accept a note (not a warning or an error) and have a far better visibility for these files.
>
Ok

> As stated by the reviewers and our team, the scope of the package may seem narrow, but our collaborations with various users make us think it is useful, helps gain productivity and insights. Therefore, it is probably up to JOSS and @martinfleis to decide whether it is worth being published in the journal.
>
> Whatever the decision, I thank you @wcjochem @matthiasdemuzere @martinfleis @ebocher for the improvements I could make thanks to all the remarks.
