### 2023-10-19 2c90e70
ShinyApp saves config file in tmp if no folder is specified, but outputs the message in console and not through
