### 2023-10-19 ffb52a4
Update importLCZraster.R
typo
