### 2023-10-19 f63ab86
Update importLCZraster.R
Fixed typos in the example of importLCZraster
