### 2023-10-20 #5445 @matthiasdemuzere
> The plot bug is really very strange. I didn't manage to reproduce it, neither on Linux (with French and English locales) or on Windows. The link you provided shows that some of those who encountered it couldn't explain it either. The underlying libraries used for plotting are `sf` for the data and `ggplot2` for the graphics productions, which are quite standard libraries ! I may try to use `tmap` to produce the maps in a future version, if I do so, maybe you can see if it works better ? But I still can't explain why no other user of the package would face this bug. Does it also happen when you use `showLCZ()` after importing a vector with `importLCZvect()` ? If so, it is strange. I do not dismiss this is a bug, but as you pointed, it is seen by users not using `lczexplore` and may be a problem of either `sf` or `ggplot2` ?

Well, as I am a user of lczexplore, this was blocking me of course completely.

But ... good thing I am a persistent user when it comes to software bugs.

So, I completely removed the R environment on my system and did a fresh install. After that I also installed `lczexplore` again. And good news: this fresh install was able to build the vignettes, and I am now able to explore the package using my own data, and visualize the results. Unfortunately I have not been able to identify the problem in my previous set-up, so I can't provide feedback on this.

But since the package now works for me, I can do some tests.

### 1. Compare the [EU](https://figshare.com/articles/dataset/European_LCZ_map/13322450) and [Global LCZ](https://zenodo.org/doi/10.5281/zenodo.6364593) maps for Rennes (France)
Both in terms of terms of all LCZ classes as well as grouped classes. This exercise also tests the capacity to read a large-scale map, and the re-projection of the EU map, that is in a different coordinate system than the Global map. This works fine, as you can see for the grouped comparison below.

```
library(lczexplore)
library(sf)

#bBox Rennes
RennesBoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(-1.78,48.05)), st_point(c(-1.55,48.2)),crs = 4326))
rennesBbox<-st_bbox(RennesBoxCoord)
#rennesBbox<-getbb( "Rennes",format_out = "sf_polygon")

LCZ1<-importLCZraster(
  dirPath="/mnt/SSD2TB/data/FIGSHARE/LCZ_EU/v4",
  zone = "europe",
  bBox=rennesBbox,
  fileName = "EU_LCZ_map.tif",
  column = "EU_LCZ_map",
  typeLevels = c(`1` = "1", `2` = "2", `3` = "3", `4` = "4", `5` = "5", `6` = "6", `7` =
                   "7", `8` = "8", `9` = "9", `10` = "10", `101` = "11", `102` = "12", `103` = "13",
                 `104` = "14", `105` = "15", `106` = "16", `107` = "17")
)

showLCZ(LCZ1,column="EU_LCZ_map")

LCZ2<-importLCZraster(
  dirPath="/mnt/SSD2TB/data/ZENODO/LCZ_GLOBAL/v3",
  zone = "global",
  bBox=rennesBbox,
  fileName = "lcz_filter_v3.tif",
  column = "lcz_filter_v3",
  typeLevels = c(`1` = "1", `2` = "2", `3` = "3", `4` = "4", `5` = "5", `6` = "6", `7` =
                   "7", `8` = "8", `9` = "9", `10` = "10", `101` = "11", `102` = "12", `103` = "13",
                 `104` = "14", `105` = "15", `106` = "16", `107` = "17")
)

showLCZ(LCZ2,column="lcz_filter_v3")

## Comparison
comparison<-compareLCZ(sf1=LCZ1,column1="EU_LCZ_map", wf1="LCZ EU",
                        sf2=LCZ2,column2="lcz_filter_v3",wf2="LCZ Globe", ref=1,
                        repr="standard",exwrite=F,location="Rennes",
                        saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/compareLCZ.png")

## Comparison - grouped
LCZ1grouped<-groupLCZ(LCZ1,column="EU_LCZ_map",
                           urban=c("1","2","3","4","5","6","7","8"),
                           industry="10",
                           vegetation=c("9", "101","102","103","104"),
                           impervious="105",pervious="106",water="107",colors=c("red","black","green","grey","burlywood","blue"))
LCZ2grouped<-groupLCZ(LCZ2,column="lcz_filter_v3",
                           urban=c("1","2","3","4","5","6","7","8"),
                           industry="10",
                           vegetation=c("9", "101","102","103","104"),
                           impervious="105",pervious="106",water="107",colors=c("red","black","green","grey","burlywood","blue"))
comparisonGrouped<-compareLCZ(sf1=LCZ1grouped,column1="grouped",
           sf2=LCZ2grouped, column2="grouped",
           repr="alter",
           levels=c("urban","industry","vegetation","impervious","pervious","water"),
           colors=c("red","black","green","grey","burlywood","blue"),
           saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/compareLCZgrouped.png")

![compareLCZgrouped](https://github.com/openjournals/joss-reviews/assets/6926916/dbd053bb-099c-45e2-ad5c-b78f6a6f0e99)
```

I'll do some more testing, so stay tuned!
