### 2023-10-20 #5445 @MGousseff
@matthiasdemuzere I find it very kind of you to have taken the time to redo an install from scratch, and I am sorry the first installation didn't work.

If I understand your example, after grouping, one can see, for instance, that what was seen as water in the EU_map is seen in the world map as water for only 49% and as some vegetation for 37% and as urban for 14%. The default workflows are "BDTOPO V2.2" for the first dataset and "OSM" for the second dataset, and it reflects our main usage, but of course, you could have set them to wf1="Europe map" and wf2=" World map", and the labels would have followed.

All publication matters set aside, I'm really glad you could install the package, that it works, and that you took time to  think of a testing use case. Maybe my last comment  better reframed a usefull situation, and I should have started by this example.

I'm sure that given your conscientiousness you will now find new bugs for me to fix ! Game may be back on for a little moment, and I'm eager to fix any problem you may find.

Thank you for your persistance.
