### 2023-10-20 #5445 @matthiasdemuzere
### 2. Compare a [LCZ Generator map](https://lcz-generator.rub.de/factsheets/86bb1a1c28c6faa2c3cd109867b8ef5f87054d98/86bb1a1c28c6faa2c3cd109867b8ef5f87054d98_factsheet.html) with the [Global LCZ](https://zenodo.org/doi/10.5281/zenodo.6364593) for Sydney (Australia)

As the LCZ generator geotiff (get it [here](https://lcz-generator.rub.de/factsheets/86bb1a1c28c6faa2c3cd109867b8ef5f87054d98/86bb1a1c28c6faa2c3cd109867b8ef5f87054d98.tif)) contains 3 bands, for simplicity I first extracted the lczFilter band, to have one band only.

`gdal_translate -b 2 86bb1a1c28c6faa2c3cd109867b8ef5f87054d98.tif 86bb1a1c28c6faa2c3cd109867b8ef5f87054d98_lczfilter.tif`

Using this code provides me with an error:

```
#Define Bbox
BoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(150.5,-34.15)), st_point(c(151.3,-33.6)),crs = 4326))
Bbox<-st_bbox(BoxCoord)
#rennesBbox<-getbb( "Rennes",format_out = "sf_polygon")

LCZ1<-importLCZraster(
  dirPath="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review",
#  zone = "europe",
  bBox=Bbox,
  fileName = "86bb1a1c28c6faa2c3cd109867b8ef5f87054d98_lczfilter.tif",
  column = "LCZ1",
  typeLevels = c(`1` = "1", `2` = "2", `3` = "3", `4` = "4", `5` = "5", `6` = "6", `7` =
                   "7", `8` = "8", `9` = "9", `10` = "10", `101` = "11", `102` = "12", `103` = "13",
                 `104` = "14", `105` = "15", `106` = "16", `107` = "17")
)
```

Error:

```
Error in `stopifnot()`:
ℹ In argument: `LCZ1 = fct_recode(...)`.
Caused by error:
! `LCZ1` must be size 545302 or 1, not 0.
Run `rlang::last_trace()` to see where the error occurred.
```

But, when I change `column = lczFilter` (the actual name of the band in the geotiff), it does work. So this seems to contradict your previous statement that `column` only refers to the name of the column that is provided to the data object?

So now I am able to read the LCZ Generator map.

Reading the global map works as well, but the LCZ labels are not converted to the pre-defined types. This seems to be due to the `nodata` values in the clipped LCZ map, since the global map often has no LCZ labels over oceans. Perhaps future versions of the code should introduce a treatment of missing values?

For now I just clipped a smaller area of interest, just to complete the test.
```
BoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(151.0,-33.7)), st_point(c(151.2,-33.97)),crs = 4326))
Bbox<-st_bbox(BoxCoord)
```

Since the compareLCZ process seems to hang for a while now, I further decreased the area of interest.
```
BoxCoord<-st_sf(a=1:2, geom=st_sfc(st_point(c(151.0,-33.7)), st_point(c(151.1,-33.8)),crs = 4326))
Bbox<-st_bbox(BoxCoord)
```

The comparison sections have been running for more than half an hour now, without completing.
As there is no messaging on the command line (except The column  lczFilter  of the dataset LCZ1 is the reference against which the  lcz_filter_v3  column of the dataset  LCZ2 will be compared.`), it is hard to say whether anything breaks in the process, or whether it is just slow. So:
- some debugging might be required here
- I believe it would be useful to add some progress statements in these routines

```
comparison<-compareLCZ(sf1=LCZ1,column1="lczFilter", wf1="LCZ Generator",
                        sf2=LCZ2,column2="lcz_filter_v3",wf2="LCZ Globe", ref=1,
                        repr="standard",exwrite=F,location="Sydney",
                        saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/compareLCZ_Sydney.png")

## Comparison - grouped
LCZ1grouped<-groupLCZ(LCZ1,column="lczFilter",
                         urban=c("1","2","3","4","5","6","7","8"),
                         industry="10",
                         vegetation=c("9", "101","102","103","104"),
                         impervious="105",pervious="106",water="107",
                         colors=c("red","black","green","grey","burlywood","blue"))
LCZ2grouped<-groupLCZ(LCZ2,column="lcz_filter_v3",
                           urban=c("1","2","3","4","5","6","7","8"),
                           industry="10",
                           vegetation=c("9", "101","102","103","104"),
                           impervious="105",pervious="106",water="107",colors=c("red","black","green","grey","burlywood","blue"))
comparisonGrouped<-compareLCZ(sf1=LCZ1grouped,column1="grouped",
           sf2=LCZ2grouped, column2="grouped",
           repr="alter",
           levels=c("urban","industry","vegetation","impervious","pervious","water"),
           colors=c("red","black","green","grey","burlywood","blue"),
           saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/compareLCZgrouped_Sydney.png")
```
