### 2023-10-20 #5445 @matthiasdemuzere
3. Reproducing vignette lczexplore_raster_vector

Yet instead of using the WUDAPT .tif as part of the package, I clip it from the global LCZ map v3.

All seems to work fine, as you can see from the grouped comparison results below.

```
library(lczexplore)
library(osmdata)
library(sf)

redonBbox<-importLCZvect(dirPath=paste0(system.file("extdata", package = "lczexplore"),
"/bdtopo_2_2/Redon"), file="rsu_lcz.geojson",column="LCZ_PRIMARY",output="bBox")

redonWudapt<-importLCZraster(
  dirPath="/mnt/SSD2TB/data/ZENODO/LCZ_GLOBAL/v3",
  zone = "europe",
  bBox=redonBbox,
  fileName = "lcz_filter_v3.tif",
  column = "lcz_filter_v3",
  typeLevels = c("1" = "1", "2" = "2", "3" = "3", "4" = "4", "5" = "5", "6" = "6", "7" =
                   "7", "8" = "8", "9" = "9", "10" = "10", "101" = "11", "102" = "12", "103" = "13",
                 "104" = "14", "105" = "15", "106" = "16", "107" = "17")
)
st_crs(redonWudapt)

# Path to embedded Data
dirPath<-paste0(system.file("extdata",package="lczexplore"),"/")
# Path to OSM data specifically
dirPathOSM<-paste0(dirPath,"osm/2022/Redon")
# loading Redon Data produced with GeoClimate on OSM input Data
redonOSM<-importLCZvect(dirPath=dirPathOSM,file="rsu_lcz.geojson",column = "LCZ_PRIMARY", drop=TRUE)
summary(redonOSM$LCZ_PRIMARY)

## Comparison
compareLCZ(sf1 = redonWudapt, column1 = "lcz_filter_v3", wf1="WUDAPT",
           sf2 = redonOSM, column2 = "LCZ_PRIMARY", wf2 = "GeoClimate_OSM",
           ref=1 ,exwrite=FALSE,
	   saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/lczexplore_raster_vector_redon.png")

## Comparison - grouped
# Regroup levels of the WUDAPT sf file.
redonWudaptGrouped<-groupLCZ(redonWudapt,column="EU_LCZ_map",
                              urban=c("1","2","3","4","5","6","7","8","9"),
                              industry="10", vegetation=c("101","102","103","104"),
                              impervious="105",pervious="106",water="107", outCol = "Regrouped")

# Regroup levels of the GeoClimate sf file
redonOSMgrouped<-groupLCZ(redonOSM, column="LCZ_PRIMARY",
urban=c("1","2","3","4","5","6","7","8","9"),
industry="10", vegetation=c("101","102","103","104"),
impervious="105",pervious="106",water="107", outCol = "Regrouped")

# Pass the files, the columsn and a vector of colors associated to the broader categories
compareLCZ(sf1 = redonWudaptGrouped, column1 = "Regrouped", wf1 = "Wudapt",
           sf2 = redonOSMgrouped, column2 = "Regrouped", wf2 = " GeoClimate on OSM data",
           exwrite = FALSE, plot=TRUE, repr="alter", levels=c("urban","industry","vegetation","impervious","pervious","water"),
           colors=c("red","black","green","grey","burlywood","blue"),
           saveG="/home/matthias/Documents/reviews/Gousseff_lcz_explore_review/lczexplore_raster_vector_redon_grouped.png")
#>  The column  Regrouped  of the dataset redonWudaptGrouped is the reference against which the  Regrouped  column of the dataset  redonOSMgrouped will be compared.
#> Both sf datasets need to live in the same crs projection (srid / epsg),
#> they will be coerced to the specified reference (redonWudaptGrouped)
#> Warning: attribute variables are assumed to be spatially constant throughout all
#> geometries

#> Warning: attribute variables are assumed to be spatially constant throughout all
#> geometries
```

![lczexplore_raster_vector_redon png](https://github.com/openjournals/joss-reviews/assets/6926916/18c58e41-46e2-496b-835f-8d187ea9be38)
