### 2023-10-20 #5445 @matthiasdemuzere
### 4. Work with the confidence levels - vignette lczexplore_en

I like this idea, so I wanted to test it out ,using the probability layer that is embedded within the WUDAPT-based LCZ maps. This probability reflects the number of types the modal LCZ is assigned to a pixel during 25 bootstraps.

But .... I realized this procedure is only available for the vector files read by `importLCZvect`, and not with the raster files read by `importLCZraster`.

Too bad, perhaps this should become a new feature in upcoming releases.
