### 2023-10-20 #5445 @matthiasdemuzere
> Even weirder is your stalling with the getbb function. Below is the time it took on my R install :
>
> ```
> system.time(rennesBbox<-getbb( "Rennes",format_out = "sf_polygon"))
>    user  system elapsed
>   0.501   0.000   1.307
> ```

The problem is not in getting the bounding box. It exists when using this (`rennesBbox<-getbb( "Rennes",format_out = "sf_polygon")`) bounding box in clipping the map. For the EU map it does not seem to work, it does however when clipping from the global map? Or perhaps it takes long for the EU map because a reprojection is required?

I am not sure who you documented the definition of creating a bounding box in the updated version (via `sf` or `osmdata`, but perhaps it makes sense to double-check if it really works?

> About the reference, they are produced by the github action from the paper.md and paper.bib, I think it should be okay with JOSS guidelines ? If not, I have to investigate why the github action would produce something different, maybe my bib file is not standard ? I redid a pull request in case the action would not have worked as intended.
>
Instead of using [@REF], you need to use @REF for references as part of sentences. See eg. [my W2W JOSS paper](https://github.com/matthiasdemuzere/w2w/tree/main/JOSS) and references in there.

> About Focusing on LCZ and not including the possibility to compare other categorical maps. It was our first approach. Then, working with specialists of urban development and urban climate, it appeared that they use other types of classifications quite frequently, like utrf, or even urban planning zoning, and they saw the `importQualVar()` function as quite useful.
>
That I understand. Yet all scripts are really tailored towards LCZs (and have LCZs in their name). Also, to me the JOSS text does not work trying to have both audiences targeted at the same time. It just makes things confusing? I see two possible pathways:
1. Target mainly LCZ users, and add a small note somewhere that these scripts can be used on other maps as well. No need to change the codes
2. Revise the whole manuscript and the codes to make it generic, and then use LCZs as just one example, given the interest of this scheme in the urban climate community.

If I where you, I would opt for 1.

Perhaps analogues: the [LCZ Generator tool ](https://lcz-generator.rub.de/) could be used to map any type of land cover or landscape feature, as long as you provide the expected training label names. But for simplicity, we did not include this in the paper.

> I don't think our difference of points of view is that controversial nor problematic.
>
> > I don't see the added value of idenfying the (dis)similarity between map A and B without knowing the best available reference.
>
> In our own practice, when we compared some maps produced from OpenStreetMap data or from BD_TOPO data, highlighting the disagreements wasn't a way to blame one or the other source of data, but to understand what the differences could teach us. I find very hard to grasp how one could not find interesting to see how two methods differ on the same area.
>
> Let's imagine a theoretical situation : you produce a map from satellite images, but you have no prior expert knowledge on this area. Then, you work with ground experts, who can qualify some areas, and you can now use this information to train a machine learning algorithm on this area and produce a refined map. Wouldn't it be useful to see how the use of expert knowledge modifies the classification, where, and how the old values are broken-down into the new values ? There is no "absolute ground truth" either, but a useful insight as to how expert knowledge impacted the new vision of the territory ? Does this example seem somehow relevant to you ? I hope so.

The part that starts with ground experts ... you do know you just described "the" WUDAPT approach no?? Irony?

That said. No, unfortunately I do not share the same interest. I've been making LCZ maps for years now, and my main concern is to make sure to reach a certain level of quality. This is something we typically do by bootstrapping and accuracy assessment against the expert labels, or via independent products that serve as proxies for the underlying paramater values. Not perfect no, as it also has its limitations (always described transparently in papers or [here](https://lcz-generator.rub.de/faq#how-accuracy-is-calculated)).

I don't see how comparing two maps can serve this. Two maps could be similar, but because of false positives. Or dissimilar because of true negatives. Or anything in between. But that does not tell me really what I should focus to improve the accuray, nor what features or procedures I should use to make one of the two maps better.

So we do see things differently. But as said, I leave it up to the community to assess the value of this package. So I will not get back to this anymore.

>
> I think I took into account most of the improvements you suggested, to the package and the article, and they helped improve both.
>
> @martinfleis Please let me know if you feel I need to do something else to help the editor about the decision regarding this paper.
