### 2023-10-20 #5445 @matthiasdemuzere
@martinfleis: it has taken me a long time to delve into this. And now I finally did. Trying to be critical yet constructive at the same time, which has been rather time-consuming.

I have updated [my reviewers check list](https://github.com/openjournals/joss-reviews/issues/5445#issuecomment-1556645157).

As you might have seen from the discussions above, the authors and myself not always agree on things. But this disagreement does not outweigh the work that has been put in these scripts. I was able to test most of them, also with unseen data, and they perform as expected.

I do think that the paper itself needs some work still. In terms of structure, text and clarity. The same might be true for some module docs and vignettes. I am willing to read JOSS paper one more time when revised, but other than that I can not put more time in this review ....
