### 2023-10-20 #5445 @MGousseff
@matthiasdemuzere
1) I didn't think of the possibility to have a confidence value as a band of a raster, and I didn't know you had one layer of probability included in the map rasters : I am coding it right now. The `terra` package, on which `lczexplore` relies for raster allows this, and it would allow the use of the sensitivity analysis when comparing raster maps: very interesting.

2) There was absolutely no irony in my use case example. I'm sorry if sometimes text  written in a a language which is not my first doesn't convey the proper tone. It was, indeed, inspired from the workflow of wudapt, as I tried to find a case which could be useful in your frame of work.

3) About the vignettes. The graphic library I use is ggplot2. When one executes the chunks of code it renders quite well, as one can zoom at will, but it seems to render poorly when a vignette is built. I'll try to tweak graphics size, and if it doesn't behave properly, I'll export the graphs to png and upload them in the vignette (I did it for one of the figure and is why you saw it as duplicate, but it makes the package heavier). Or if you consider this being a problem, I'll get rid of the vignettes until I have fixed the issue (vignettes are not a core part of the package).

4) About the article. I'm a bit surprised about your point of view. I added the part about map utility as a response to your concern of may 24 :

> Once you know about the difference / agreements between two maps, what then? How will you use this information?

It seems to me that *comparing map* section addresses the need to compare two maps, and once this general problem is cleared, the article focuses on the core aim of the package : comparing LCZ maps.

5) In response to
> I see two possible pathways:
> 1. Target mainly LCZ users, and add a small note somewhere that these scripts can be used on other maps as well. No need to change the codes
> 2. Revise the whole manuscript and the codes to make it generic, and then use LCZs as just one example, given the interest of this scheme in the urban climate community.

I feel that the article is already following 1.
I describe all the process for LCZ maps and then just show a function exists to do the same with any other qualitative variable. I'll let that sink during the weekend because I have a hard time seeing how it confuses the reading flow, but I may be too "acquainted" to my own paper to see this. I'll have a better point of view next week, after I finish coding multilayer rasters, probability layer, raster sensitivity analysis.

We have a core difference of points of view about the use of comparing maps. Don't you find it interesting to see how your global map and your Europe map differ on the Rennes area, as you showed with your example ? I find it interesting. It would have been a different message if the disagreement was not about water but urban areas, for instance. Even if none of the map is ground truth, it's still a lead on where to focus when trying to get more information on ground truth ?

Anyway, I'd like to say again : there is no irony nor controversial feeling on my side. I see the work you put for this review, I am grateful for it. I'm very glad I can improve the package and hopefully make it more useful for a part of the community who uses more raster sources than I do.

Next Pull Request (and probably last for this review) will occur before the end of last week.
