### 2023-10-23 30601f9
Possibility to chose the raster band and include a confidence band when importing a raster.
