+++ R/importLCZraster.R
@@ -1,18 +1,24 @@
-#' Imports the LCZ classifications produced on raster maps, mainly by the WUDAPT algorithm.
-#' For now, the function  import from the european tiff
-#' produced by WUDAPT. Users can DOWNLOAD the WUDAPT tiff file
-#' at the following url : https://figshare.com/articles/dataset/European_LCZ_map/13322450
-#' A future version may include the world data once a strategy is defined to deal with CRS.
+#' Imports the LCZ classifications produced on raster maps.  
+#' For now, the european tif map produced by WUDAP is used as a reference example. 
+#' Users can DOWNLOAD the WUDAPT tiff file at the following url : 
+#' https://figshare.com/articles/dataset/European_LCZ_map/13322450
+#' A future version may include examples with the world data once a strategy is defined to deal with CRS.
 #'
 #' @param dirPath is the path to the directory where the
-#' @param fileName is the name of the raster file (tif or geotif), by default \'EU_LCZ_map.tif\' . 
+#' @param fileName is the name of the raster file (tif or geotif), by default \'EU_LCZ_map.tif\'. 
 #' Will be useful when other zones will be added
-#' @param column indicates the name of the column which will contain the LCZ in the output file
-#' @param typeLevels indicates a named vector of the unique values contained in column,
+#' @param LCZband the band of the raster files contains the LCZ types. Can be a number or a band name (default = 1)
+#' @param LCZcolumn the name of the column of the output sf File in which the LCZ types will be stored
+#' (default = "LCZ")
+#' @param confidenceBand the band of the raster files contains a measure of confidence. 
+#' Can be a number or a band name (default = "", that is no measure of confidence provided)
+#' @param confidenceColumn the name of the column of the output sf File in which the confidences values will be stored
+#' (default = "", no confidence value specified)
+#' @param typeLevels indicates a named vector of the unique values contained in LCZcolumn,
 #' @param zone set to europe by default, may include world once a strategy is defined
-#' @param bBox bBox is the bounding box needed to crop the wudapt tiff file.
+#' @param bBox bBox is the bounding box needed to crop the raster file.
 #' It can be produced bu the importLCZvect function. It can either be of class bBox or of class sfc
-#' @return an sf file containing the geom and LCZ levels from the WUDAPT Europe tiff within the bBox bounding box
+#' @return an sf file containing the geom and LCZ levels from theraster tiff within the bBox bounding box
 #' @import sf dplyr forcats
 #' @importFrom terra crop
 #' @importFrom terra rast
@@ -47,27 +53,38 @@
 #' #sanDiegoWudapt<-importLCZraster(
 #' #dirPath="path_of_the_tiff",
 #' #fileName="CONUS_LCZ_map_NLCD_v1.0_epsg4326.tif",
-#' #column="CONUS_LCZ_map_NLCD_v1.0_epsg4326"
+#' #LCZcolumn="CONUS_LCZ_map_NLCD_v1.0_epsg4326"
 #' #  ,bBox=sanDiegoBbox)
-#' #showLCZ(sanDiegoWudapt, column="CONUS_LCZ_map_NLCD_v1.0_epsg4326")
-importLCZraster<-function(dirPath,zone="europe",bBox,fileName="EU_LCZ_map.tif", column='EU_LCZ_map',
+#' #showLCZ(sanDiegoWudapt, LCZcolumn="CONUS_LCZ_map_NLCD_v1.0_epsg4326")
[…]
