+++ R/importLCZraster.R
@@ -1,21 +1,19 @@
 #' Imports the LCZ classifications produced on raster maps.  
 #' For now, the european tif map produced by WUDAP is used as a reference example. 
-#' Users can DOWNLOAD the WUDAPT tiff file at the following url : 
+#' Users can DOWNLOAD the WUDAPT tif file at the following url : 
 #' https://figshare.com/articles/dataset/European_LCZ_map/13322450
-#' A future version may include examples with the world data once a strategy is defined to deal with CRS.
 #'
 #' @param dirPath is the path to the directory where the
-#' @param fileName is the name of the raster file (tif or geotif), by default \'EU_LCZ_map.tif\'. 
-#' Will be useful when other zones will be added
-#' @param LCZband the band of the raster files contains the LCZ types. Can be a number or a band name (default = 1)
+#' @param fileName is the name of the raster file (tif or geotif), by default \'EU_LCZ_map.tif\'.
+#' @param LCZband defines the band of the raster file which contains the LCZ types. 
+#' It can be a number or a band name (default = 1)
 #' @param LCZcolumn the name of the column of the output sf File in which the LCZ types will be stored
 #' (default = "LCZ")
-#' @param confidenceBand the band of the raster files contains a measure of confidence. 
+#' @param confidenceBand the band of the raster files which contains a measure of confidence. 
 #' Can be a number or a band name (default = "", that is no measure of confidence provided)
-#' @param confidenceColumn the name of the column of the output sf File in which the confidences values will be stored
+#' @param confidenceColumn the name of the column of the output sf File in which the confidence values will be stored
 #' (default = "", no confidence value specified)
 #' @param typeLevels indicates a named vector of the unique values contained in LCZcolumn,
-#' @param zone set to europe by default, may include world once a strategy is defined
 #' @param bBox bBox is the bounding box needed to crop the raster file.
 #' It can be produced bu the importLCZvect function. It can either be of class bBox or of class sfc
 #' @return an sf file containing the geom and LCZ levels from theraster tiff within the bBox bounding box
@@ -56,33 +54,31 @@
 #' #LCZcolumn="CONUS_LCZ_map_NLCD_v1.0_epsg4326"
 #' #  ,bBox=sanDiegoBbox)
 #' #showLCZ(sanDiegoWudapt, LCZcolumn="CONUS_LCZ_map_NLCD_v1.0_epsg4326")
-importLCZraster<-function(dirPath,zone="europe",bBox,fileName="EU_LCZ_map.tif", LCZband=1, LCZcolumn='EU_LCZ_map',
-                          confidenceBand="",confidenceColumn="confidence",
+importLCZraster<-function(dirPath, bBox, fileName="EU_LCZ_map.tif", LCZband=1, LCZcolumn='EU_LCZ_map',
+                          confidenceBand = "", confidenceColumn = "confidence",
                           typeLevels=c("1"="1","2"="2","3"="3","4"="4","5"="5","6"="6","7"="7","8"="8",
                            "9"="9","10"="10","101"="11","102"="12","103"="13","104"="14",
                             "105"="15", "106"="16","107"="17")){
 # internal import function used in two loops
   effectiveImport<-function(fileName,bBox,LCZband,confidenceBand,LCZcolumn, confidenceColumn){
-  if(confidenceBand=="") {
-    argList<-list(LCZband)
-    names(argList)<-LCZcolumn
-    sfFile<-rast(fileName, lyrs = argList)
+
[…]
