### 2023-10-24 #5445 @matthiasdemuzere
@martinfleis I checked all the boxes in [my checklist](https://github.com/openjournals/joss-reviews/issues/5445#issuecomment-1556645157) so I consider my work done.
