+++ vignettes/lczexplore_en.Rmd
@@ -49,9 +49,9 @@ On can also install from the tar.gz file and install it (see `?install.packages`
 
 Load the package: `library(lczexplore)`.
 ```{r}
+library(ggplot2)
 library(lczexplore)
 ```
-
 
 # First use-case: load a file with geometries and standard LCZ
 
@@ -193,17 +193,17 @@ the compareLCZ function will concatenate a ".1" string to the column names of th
 ```{r, fig.dim = c(12,9 ), echo = TRUE, fig.show="hide"}
 comparison<-compareLCZ(sf1=redonBDT,column1="LCZ_PRIMARY", wf1="BD TOPO v2.2",
                         sf2=redonOSM,column2="LCZ_PRIMARY",wf2="Open Street Map", ref=1,
-                        repr="standard",exwrite=F,location="Redon",saveG="")
+                        repr="standard",exwrite=F, location="Redon", saveG="comparison")
 ```
 
 ```{r, out.width = "680px"}
-path <- system.file('extdata/RedonCompare.png',package="lczexplore")
-knitr::include_graphics(path, error=FALSE)
+
+knitr::include_graphics("comparison.png", error=FALSE)
 
 ```
 
 The comparison also works on grouped data. 
-```{r, fig.dim = c(8, 6), echo = TRUE}
+```{r, echo = TRUE, fig.show="hide" }
 redonOSMgrouped<-groupLCZ(redonOSM,column="LCZ_PRIMARY",
                            urban=c("1","2","3","4","5","6","7","8","9"),
                            industry="10",
@@ -218,14 +218,18 @@ comparisonGrouped<-compareLCZ(sf1=redonOSMgrouped,column1="grouped",
            sf2=redonBDTgrouped, column2="grouped",
            repr="alter", 
            levels=c("urban","industry","vegetation","impervious","pervious","water"),
-           colors=c("red","black","green","grey","burlywood","blue"))
+           colors=c("red","black","green","grey","burlywood","blue"), saveG="comparisonGrouped") 
+```
 
-                           
+```{r, out.width = "680px"}
+
+knitr::include_graphics("comparisonGrouped.png", error=FALSE)
+
 ```
 
 A little cheat is possible ! You can try to group and compare on-the-fly with `tryGroup=TRUE`. The syntax is then as shown in the nex code chunk. 
 
-```{r, fig.dim = c(8, 6), echo = TRUE}
+```{r, out.width="70%", echo = TRUE, fig.show="hide"}
 comparisonGrouped<-compareLCZ(sf1=redonBDT,column1="LCZ_PRIMARY", wf1="groupedBDT",
     sf2=redonOSM, column2="LCZ_PRIMARY", wf2="groupedOSM",
     repr="alter", ref=2, saveG="", exwrite=FALSE, location="Redon", plot=TRUE,
@@ -234,6 +238,8 @@ comparisonGrouped<-compareLCZ(sf1=redonBDT,column1="LCZ_PRIMARY", wf1="groupedBD
     vegetation=c("101","102","103","104"),
     impervious="105",pervious="106",water="107",
     colors=c("red","black","green","grey","burlywood","blue"),tryGroup = TRUE)
+
+
 ```
 
 
@@ -298,18 +304,28 @@ The span of the confidence values is divided on as many points as the user choos
 
 For each of these thresholds, the function only keeps the reference spatial units (RSU, i.e. the geometries) for which the confidence value is higher than the threshold. Then it computes the agreement between the classifications based on these geoms.
 
-```{r, fig.dim = c(12,9)}
+```{r, fig.show="hide"}
 names(redonCompare$data)
[…]
