### 2023-10-24 00dccda
Check tests : some files still created and to be deleted after testing. Check references Check author references.
