+++ docs/articles/joss/paper.bib
@@ -256,7 +256,8 @@
   volume={2022},
   pages={1--57},
   year={2022},
-  publisher={G{\"o}ttingen, Germany}
+  publisher={G{\"o}ttingen, Germany},
+  doi = {10.5194/ems2022-83}
 }
 
 @article{demuzere2020combining,
@@ -267,5 +268,6 @@
   number={1},
   pages={264},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s41597-020-00605-z}
 }
