+++ inst/tinytest/test_importLCZraster.R
@@ -28,14 +28,66 @@ upCorner<-st_point(c(-117.227554,32.864593))
 outBbox<-st_sfc(lowCorner,upCorner,crs=4326)
 #importLCZraster("/home/gousseff/Documents/2_CodesSources/Wudapt/WudaptEurope/",bBox=outBbox)
 
-
+# test bounding box not intersecting with rastet
 expect_error(current = test<-importLCZraster(dirPath = system.file("extdata", package = "lczexplore"),fileName="redonWudapt.tif",bBox=outBbox),
              "The bounding box doesn\'t intersect ")
 
 
+# test chosing wich band of the raster is imported
 redonWudapt2<-importLCZraster(  
   system.file("extdata", package = "lczexplore"),
   fileName="redonWudapt.tif",bBox=redonBbox,
 LCZband=1, LCZcolumn="LCZ")
  str(redonWudapt2)
 
+sidneyOSM<-importLCZvect(
+  dirPath = system.file("extdata/osm/2022/Sidney", package = "lczexplore"), file="sidney_rsu_lcz.geojson",
+  confid="LCZ_UNIQUENESS_VALUE",
+  geomID="ID_RSU")
+
+sidneyBbox<-importLCZvect(
+  system.file("extdata/osm/2022/Sidney", package = "lczexplore"), file="sidney_rsu_lcz.geojson",
+  ,output = "bBox")
+
+# test default import
+
+expect_warning(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                              fileName="rasterMD.tif",bBox=sidneyBbox),
+               "attribute variables are assumed to be spatially constant throughout all geometries")
+
+# test default specifying band number
+
+expect_warning(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                             fileName="redonWudapt.tif",bBox=redonBbox, LCZband=1, LCZcolumn='EU_LCZ_map'),
+               "attribute variables are assumed to be spatially constant throughout all geometries")
+
+
+expect_error(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                              fileName="redonWudapt.tif",bBox=redonBbox, LCZband=4, LCZcolumn='EU_LCZ_map'),
+             "invalid name")
+
+# test importing a band that doesn't exist (wrong name)
+expect_error(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                                           fileName="redonWudapt.tif",bBox=redonBbox, LCZband="Ski shoe", LCZcolumn='EU_LCZ_map'),
+             "invalid name")
+
+# test importing several bands of the raster is imported, one being the confidence, one being numeric and the other the name of a band
+
+expect_warning(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                fileName="rasterMD.tif", LCZband=1,  LCZcolumn = "LCZraster",
+                confidenceBand = 3, confidenceColumn = "confidence",
+                bBox = sidneyBbox ),
+               "attribute variables are assumed to be spatially constant")
+
+expect_warning(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
+                                             fileName="rasterMD.tif", LCZband="lcz",  LCZcolumn = "LCZraster",
+                                             confidenceBand = 3, confidenceColumn = "confidence",
+                                             bBox = sidneyBbox ),
+               "attribute variables are assumed to be spatially constant")
+
+# test importing several bands, one being wrong
+
+expect_error(sidneyWudapt<-importLCZraster(system.file("extdata", package = "lczexplore"),
[…]
