+++ docs/articles/joss/paper.md
@@ -23,7 +23,7 @@ authors:
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: "1, 2, 4"
   - name: Elisabeth Le Saux Wiederhold
-    orcid:
+    orcid: 0000-0002-2079-8633
     corresponding: true
     affiliation: 3
 affiliations:
@@ -46,7 +46,7 @@ editor_options:
 # Summary
 
 Climate change is a growing concern for city planners as Urban Heat Islands have an impact on
-mortality [@clarke1972some], health in general [@lowe2016energy] and consumption of energy
+mortality @clarke1972some, health in general [@lowe2016energy] and consumption of energy
 for building cooling [@malys2012microclimate] among other effects.
 A first step towards large scale study of urban climate is to define classes
 based on logical division of the landscape,
@@ -76,15 +76,16 @@ The four main reasons to compare categorical variables on geographical units are
 Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
 The terms *rural* and *urban* lack of a clear definition and different classifications
 of urban and rural landscapes were proposed.
-Stewart and Oke [@stewart2012local] defined an approach that complies with [@grigg1965logic] criteria of geographical classification.
+@stewart2012local defined an approach that complies with criteria of geographical classification defined in [@grigg1965logic].
 Their Local Climate Zones (LCZ) are based on a logical segmentation of land-use and urban canopy parameters
 and define 10 urban types and 7 land cover types.
 
 LCZ have gained popularity in the past decade as they sum up relevant information and can help, for instance,
 apprehend the intensity of the Urban Heat Island [@kotharkar2018evaluating].
 
-Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification for any given area.  
-[@quan2021systematic] distinguishes two main streams of production of these LCZ:
+Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification 
+for any given area.  
+@quan2021systematic distinguishes two main streams of production of these LCZ:
 - the raster stream processes remotely sensed information, and applies machine learning
   algorithms trained using local experts' knowledge. In this way, the WUDAPT community [@chingWUDAPTUrbanWeather2018a]
   produced thousands of city-based LCZ maps (accessible via the LCZ Generator ([@demuzereLCZGeneratorWeb2021]))
@@ -119,7 +120,7 @@ maps, but it has two main drawbacks:
 - only raster maps where pixels match perfectly (same size, not translated) can be treated, or
   some pre-treatment are needed (like nearest neighbour interpolation for instance).
 
-To prevent the first drawback, [@monserudComparingGlobalVegetation1992]
+To prevent the first drawback, @monserudComparingGlobalVegetation1992
 proposed the use of Cohen's kappa coefficient of agreement for nominal scales [@cohenCoefficientAgreementNominal1960].
 Cohen's Kappa allow to assess how two raters agree or differ in the task of rating individual.
 In our case, each pixel is seen as an individual, each map is seen as a rater and the value of the raster at each pixel
@@ -146,12 +147,12 @@ As Local Climate Zones emerged as a new standard for characterizing urban landsc
 specific tools are needed to allow easy automation of LCZ maps. As a matter of fact, "ground truth" is hard to
 define in matters of LCZs, and one may want to compare LCZs produced by experts and by different algorithms.
 
-The very definition of LCZs by [@stewart2012local] relies on ranges of values for several indicators, and  sometimes
+The very definition of LCZs by @stewart2012local relies on ranges of values for several indicators, and  sometimes
[…]
