+++ inst/tinytest/test_produceAnalysis.R
@@ -12,6 +12,10 @@ expect_warning(produceAnalysis(location="Redon",
                           wf2="osm",refYear1="2022",refYear2="2022",repr="standard",saveG=location),
                "attribute variables are assumed to be spatially constant")
 
+file.remove("Redon_bdtopo_2_2_osm_standard.png")
+file.remove("bdtopo_2_2_osm.csv")
+
+
 expect_message(produceAnalysis(location="Redon",
                 outDir=paste0(
                   system.file(package="lczexplore"),"/tinytest"),
@@ -20,6 +24,8 @@ expect_message(produceAnalysis(location="Redon",
  " rsu_lcz.geojson already exists in this directory"
 )
 
+file.remove("Redon_bdtopo_2_2_osm_standard.png")
+file.remove("bdtopo_2_2_osm.csv")
 file.remove(paste0(
   system.file(package="lczexplore"),"/tinytest/bdtopo_2_2/Redon/rsu_lcz.geojson"))
 
@@ -48,13 +54,9 @@ expect_message(
 )
 
 
-# file.remove(paste0(
-#   system.file(package="lczexplore"),"/tinytest/bdtopo_2_2/Redon/Redon.zip"))
-
-# produceAnalysis(location="Redon",
-#                 outDir=paste0(
-#                   system.file(package="lczexplore"),"/tinytest"),
-#                 wf1="bdtopo_2_2",
-#                 wf2="osm",refYear1="2022",refYear2="2022",repr="standard",saveG=location)
+file.remove(paste0(
+  system.file(package="lczexplore"),"/tinytest/bdtopo_2_2/Redon/Redon.zip"))
 file.remove("Redon_bdtopo_2_2_osm_standard.png")
 file.remove("bdtopo_2_2_osm.csv")
+file.remove(paste0(
+  system.file(package="lczexplore"),"/tinytest/bdtopo_2_2/Redon/rsu_lcz.geojson"))
