### 2023-10-25 c3a882b
Check tests : some files still created and to be deleted after testing. Check references : done Missing paper orcid : added. Check author references : added ELSW ORCID
