+++ R/shinyGC/app.R
@@ -1,292 +1,0 @@
-library(shiny)
-library(shinyFiles)
-library(ggplot2)
-library(lczexplore)
-library(magrittr)
-library(osmdata)
-library(sf)
-#devtools::install_github("elipousson/getdata")
-library(getdata)
-
-
-# Define UI for app that draws a histogram ----
-ui <- fluidPage(
-  h1(" This application helps build a configuration file to feed to geoclimate"),
-  h2(" The OSM workflow allows to run GeoClimate on any given city."), 
-  h2("For BD TOPO workflows, the user has to provide a path to the input data"),
-  h2(" It is not possible to build a configuration files using bounding box coordinates yet."),
-  
-  tabsetPanel(
-  ############################################
-  ##
-  ## Tab to create the config file
-  ##
-  ############################################
-    tabPanel("Create your  GeoClimate configuration JSON file",
-
-  # Sidebar layout with input and output definitions ----
-      sidebarLayout(
-    
-    # Sidebar panel for inputs ----
-        sidebarPanel(
-          shinyDirButton(id = "outFolder", 
-                    label = "Output  folder for GeoClimates results", 
-                    title = "Select the folder where GeoClimates will output its results"),
-      
-          selectInput(inputId="wf", label="Workflow", 
-                  choices = list(OpenStreetMap="OSM","BD TOPO V2"="BDTOPO_V2","BD TOPO V3"="BDTOPO_V3"),
-                  selected=list(OpenStreetMap="OSM")),
-      
-          conditionalPanel(
-            condition='input.wf!="OSM"',
-          shinyDirButton("BDTinFolder",
-                       label = "BD_TOPO folder",
-                       title = "Choose in which folder are the BD_TOPO files"),
-            checkboxInput("forceSRID",label="Force SRID of BD TOPO inputs to 2154",value=FALSE),
-            textInput(inputId="inseeCode", label="Enter Insee code of your location (town)", value = "29162")
-          ),
-      
-          conditionalPanel(
-            condition='input.wf=="OSM"',
-            textInput(inputId="location",label="Enter your locations here",
-                  value="Allaire",placeholder="A town name or some coordinates")
-          ),
-
-        checkboxGroupInput(inputId="rsuIndics",label = "Choose the indicators to compute at RSU scale",
-             choices=c("LCZ","TEB","UTRF"),selected=c("LCZ")),
-        fluidRow(
-          column( width = 4, checkboxInput(inputId = "svfSimple", 
-                         label = "Use simplified algorithm for sky view factor",
-                         value = TRUE)),
-          column(width = 4, checkboxInput(inputId = "EstimateHeight", label = "Estimate missing building heights", value = TRUE))
-        ),
-        
-        checkboxGroupInput(inputId="gridIndics",label = "Choose the indicators to compute at grid scale",
-             choices=c("BUILDING_FRACTION",
-                       "BUILDING_HEIGHT",
-                       "WATER_FRACTION",
-                       "VEGETATION_FRACTION",
-                       "ROAD_FRACTION",
-                       "IMPERVIOUS_FRACTION",
-                       "LCZ_FRACTION"),
-             selected=c("BUILDING_FRACTION"
-                        )),
-        numericInput(inputId="xGridSize", label="Choose the x size for the grid", value = 100, min = 10, max = 1000, step = 10),
[…]
