+++ R/shinyGC/launchShinyGC.R
@@ -1,1 +1,0 @@
-shiny::runApp("/home/gousseff/Documents/2_CodesSources/R/lczexploreCleanOrbisgis/lczexplore/R/shinyGC/")
