+++ inst/tinytest/test_importQualVar.R
@@ -13,14 +13,15 @@ utrfRedonBDT<-importQualVar(dirPath=paste0(
   system.file("extdata", package = "lczexplore"), "/bdtopo_2_2/Redon"),
   file="rsu_utrf_area.geojson", column="TYPO_MAJ")
   
-showLCZ(sf=utrfRedonBDT, column="TYPO_MAJ",repr="alter")
+map1<-showLCZ(sf=utrfRedonBDT, column="TYPO_MAJ",repr="alter")
 
 utrfRedonOSM<-importQualVar(dirPath=paste0(
   system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
   file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
   
-  utrfComparison<-compareLCZ(sf1=utrfRedonBDT, column1="TYPO_MAJ", sf2=utrfRedonOSM, column2="TYPO_MAJ",wf1=" UTRF BDT", wf2 = " UTRF OSM",
-  location = " Redon",exwrite=FALSE,repr="alter")
+  utrfComparison<-compareLCZ(sf1=utrfRedonBDT, column1="TYPO_MAJ", wf1=" UTRF BDT",
+                             sf2=utrfRedonOSM, column2="TYPO_MAJ", wf2 = " UTRF OSM",
+  location = " Redon",exwrite=FALSE,repr="alter", saveG="")
   # Plot the confusion matrix of thes two classifications  
   print(utrfComparison$matConfPlot)
 
@@ -28,20 +29,24 @@ utrfRedonOSM<-importQualVar(dirPath=paste0(
 
 
 expect_silent(
-  utrfRedonBDT<-importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"), 
-                                                         file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
+  utrfRedonBDT<-
+    importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/bdtopo_2_2/Redon"), 
+                 file="rsu_utrf_area.geojson", column="TYPO_MAJ", geomID="ID_RSU", confid="UNIQUENESS_VALUE")
 )
 
-showLCZ(utrfRedonBDT,column = "TYPO_MAJ",repr="alter")
+map2<-showLCZ(utrfRedonBDT,column = "TYPO_MAJ",repr="alter")
 
 expect_silent(
-  utrfRedonOSM<-importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
-                                  file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
+  utrfRedonOSM<-
+    importQualVar(dirPath=paste0(system.file("extdata", package = "lczexplore"),"/osm/2022/Redon"),
+                  file="rsu_utrf_area.geojson", column="TYPO_MAJ",geomID="ID_RSU",confid="UNIQUENESS_VALUE")
 )
 
 #summary(st_geometry_type(utrfRedonOSM))
-utrfComparison<-compareLCZ(sf1=utrfRedonBDT, column1="TYPO_MAJ", sf2=utrfRedonOSM, column2="TYPO_MAJ",wf1=" UTRF BDT", wf2 = " UTRF OSM", 
-           location = " Redon",exwrite=FALSE,repr="alter")
+utrfComparison<-
[…]
