### 2023-10-25 #5445 @MGousseff
@matthiasdemuzere

1. I added the possibility to chose the raster band (by number or name) and checked if it allowed to perform the sensibility analysis. To have a consistent measure I divided the probability you provided by 100, to get a proba between 0 and 1, as the one GeoClimate provides.
To test it, I chose a bounding box inside the zone of the tif you provided. **Is it OK for you if I include the tif you provided in the package in order to have a working example ?

2. About the logic of the article, I get what your point about how comparing pairs of maps for any other categorical variable is an "add-on" so :

- I moved the paragraph after the sensitivity analysis, so it comes only after the whole logic workflow for LCZ maps
- I added the following sentence to provide context :
*This package focuses on LCZ maps comparison, but more often than not, people working on LCZ maps also
describe their area of interest with other categorical indicators.*

3. I think I cleaned the use of reference with and without square brackets. Thank you for pointing this out : I forgot about the two ways to cite a reference.

4. For the vignettes : I generated png for all the ggplot2 graphs that scale poorly in the vignettes. They now appear in a readable way.
