### 2023-10-25 #5445 @matthiasdemuzere
@MGousseff

> To test it, I chose a bounding box inside the zone of the tif you provided. **Is it OK for you if I include the tif you provided in the package in order to have a working example ?

Sure, I think I took it from the Generator? In any case, you can use any tiff available from the LCZ Generator. https://lcz-generator.rub.de/submissions. You could add the reference for the map you take, that is normally provided around [here](https://lcz-generator.rub.de/factsheets/561d19efc1b692888e180c9acb7e34a771f4c1f8/561d19efc1b692888e180c9acb7e34a771f4c1f8_factsheet.html#:~:text=Alternatively%2C%20you%20can%20also%20explicitly%20cite%20this%20training%20data%20by%20using%20the%20following%20reference%3A)

Thanks for taking into account my other points as well.
