### 2023-10-25 ef0450e
Added rasterMD.tif description. This file was provided by Matthias Demuzere and is used in examples for importLCZraster.
