### 2023-10-25 #5445 @MGousseff
> Thanks for taking into account my other points as well.

@matthiasdemuzere Thank you very much. The process has been a little rough, but I really am grateful for the improvements. I will improve in raster management !
