### 2023-10-30 #5445 @MGousseff
> @MGousseff could you edit Zenodo to ensure that the author list (including ORCIDs) in the archive match those in the JOSS paper? The only author there is you, others are listed as project members. They need to be authors to ensure the Zenodo-generated citation has the same authors as JOSS-generated citation. Thanks!

I am away till Monday, I will try to do it from here. Worst case scenario, it will be done by Monday. Thanks for your patience
