### 2023-10-30 #5445 @MGousseff
@martinfleis I think it is done. I just couldn't find the second affiliation of Elisabeth Le Sauce Wiederhold in Zenodo. But her researcher affiliation is Lab-sticc and is set properly. Is everything all set ? Tell me if I need to do further action, I'll try my best from here (no computer, only phone !). Thanks for the thorough process of reviewing.
