### 2023-10-30 #5445 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
