### 2023-11-06 #5445 @MGousseff
I detected a few typos ( mostly trailing spaces which were removed by code editor).
