### 2023-11-06 #5445 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
