### 2023-11-06 #5445 @oliviaguest
@MGousseff your references are a bit too bracket-y, like 3 deep is a lot! I typically ask for 2 deep to be changed.

Can you take  a look here please and change the relatevant BibTeX command?

> the raster stream processes remotely sensed information, and applies machine learning
algorithms trained using local experts’ knowledge. In this way, the WUDAPT community
(Ching et al., 2018) produced thousands of city-based LCZ maps (accessible via the
LCZ Generator ((Demuzere et al., 2021))) but also large-scale maps for Europe, the
continental United States and the whole world ((Demuzere et al., 2019), (Demuzere et
al., 2020), (Demuzere et al., 2022)).
