### 2023-11-06 #5445 @oliviaguest
oh, also here:

> A first step towards large scale
study of urban climate is to define classes based on logical division of the landscape, such as
Local Climate Zones (LCZ) defined by (Stewart & Oke, 2012).

I opened a PR here: https://github.com/orbisgis/lczexplore/pull/30
