### 2023-11-06 5575916
Update paper.md: references unbracketed.
