### 2023-11-10 #5445 @oliviaguest
@martinfleis it seems the article is 4x the size of what we consider a typical JOSS paper: https://joss.readthedocs.io/en/latest/submitting.html. Is there a reason for this I am missing?
