### 2023-11-10 #5445 @MGousseff
> @MGousseff can you do one more pass? I see line 21 should not have brackets, for example. 😊

I'm willing to comply, but I think I failed to understand to understand the logic ! You mean there should be no bracket at all when citing ? I thought that we left brackets when we cite the article and no brackets when we cite the authors, but maybe I misunderstood ? If so, there are still lots of extra brackets !

About the length of the article, I let @martinfleis answer, but the paper got a little longer when one of the reviewer asked to justify the need for comparing maps.
