### 2023-11-10 #5445 @oliviaguest
> > @MGousseff can you do one more pass? I see line 21 should not have brackets, for example. 😊
>
> I'm willing to comply, but I think I failed to understand to understand the logic ! You mean there should be no bracket at all when citing ? I thought that we left brackets when we cite the article and no brackets when we cite the authors, but maybe I misunderstood ? If so, there are still lots of extra brackets !
>

Do you mean you're confused about the difference between parenthetical versus in-text citations?
