### 2023-11-10 #5445 @oliviaguest
Here is a bit of an explanation: https://joss.readthedocs.io/en/latest/submitting.html#citations
