### 2023-11-11 #5445 @oliviaguest
> @oliviaguest The paper had originally a borderline length (I tend to minimise asking authors to cut stuff) and got longer during the review. I am personally fine with that as is. In my interpretation, the guideline is there more to limit the load on reviewers than for any other reason, but happy to reconsider that position if you think it should be shorter.

I think it should have been shortened, but now it's past that point, and it's fine to publish as it is. Thanks for clarifying. ☺️
