### 2023-11-13 db9568d
JOSS: format references and code
