+++ docs/articles/joss/paper.md
@@ -58,18 +58,18 @@ a confusion matrix to help qualify and quantify the misclassifications.
 The lczexplore package aims at comparing different LCZ classifications,
 but can be used to compare any pair of classifications on geographical units.
 
-This software is available as a free and opensource R package.
+This software is available as a free and open source R package.
 
 # Statement of need
 
 ## Comparing maps
 As stated in [@visser2006map] comparing maps is an important issue in environmental research.
-The four main reasons to compare categorical variables on geographical units are:   
+The four main reasons to compare categorical variables on geographical units are:
 
-- to assess the differences between maps generated by models under different scenarios and assumptions,   
-- to detect temporal changes,   
-- to calibrate or validate models,   
-- to perform uncertainty and sensitivity analysis.   
+- to assess the differences between maps generated by models under different scenarios and assumptions,
+- to detect temporal changes,
+- to calibrate or validate models,
+- to perform uncertainty and sensitivity analysis.
 
 
 ## Comparing specifically LCZ maps
@@ -77,26 +77,26 @@ The four main reasons to compare categorical variables on geographical units are
 Climate change is a growing concern for city planners with a special focus on Urban Heat Island phenomenons.
 The terms *rural* and *urban* lack of a clear definition and different classifications
 of urban and rural landscapes were proposed.
-@stewart2012local defined an approach that complies with criteria of geographical classification defined in [@grigg1965logic].
+@stewart2012local defined an approach that complies with criteria of geographical classification defined in @grigg1965logic.
 Their Local Climate Zones (LCZ) are based on a logical segmentation of land-use and urban canopy parameters
 and define 10 urban types and 7 land cover types.
 
 LCZ have gained popularity in the past decade as they sum up relevant information and can help, for instance,
 apprehend the intensity of the Urban Heat Island [@kotharkar2018evaluating].
 
-Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification 
-for any given area.  
-@quan2021systematic distinguishes two main streams of production of these LCZ:   
+Several methods aim to classify a territory into LCZ, but only few workflows allow an automatic classification
+for any given area.
+@quan2021systematic distinguishes two main streams of production of these LCZ:
 
 - the raster stream processes remotely sensed information, and applies machine learning
   algorithms trained using local experts' knowledge. In this way, the WUDAPT community [@chingWUDAPTUrbanWeather2018a]
   produced thousands of city-based LCZ maps, accessible via the LCZ Generator [@demuzereLCZGeneratorWeb2021],
-  but also large-scale maps for Europe, the continental United States and the whole world [@demuzere2019mapping],
-  [@demuzere2020combining], [@demuzere2022global].   
+  but also large-scale maps for Europe, the continental United States and the whole world
+  [@demuzere2019mapping; @demuzere2020combining; @demuzere2022global].
 - the vector stream uses Geographic Information System (GIS) layers that represent the main topographic features,
   defines spatial units, computes urban canopy parameters and uses them to classify spatial units into LCZ.
   For instance, the GeoClimate geospatial toolbox produces LCZ classifications
-  from OpenStreetMap or french BDTopo data [@bocher2021geoclimate].   
+  from OpenStreetMap or french BDTopo data [@bocher2021geoclimate].
 
 The existence of several methods to produce LCZ classifications, or the use of a method with different input data,
 raises the need for a tool to quickly get:
[…]
