+++ docs/articles/joss/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{monserudComparingGlobalVegetation1992,
-  title = {Comparing Global Vegetation Maps with the {{Kappa}} Statistic},
+  title = {Comparing Global Vegetation Maps with the {Kappa} Statistic},
   author = {Monserud, Robert A. and Leemans, Rik},
   year = {1992},
   month = aug,
@@ -25,7 +25,7 @@
 }
 
 @article{hammerbergImplicationsEmployingDetailed2018,
-  title = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling: A Comparison between {{WUDAPT}} and {{GIS}} Databases over {{Vienna}}, {{Austria}}},
+  title = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling: A Comparison between {WUDAPT} and {GIS} Databases over {Vienna}, {Austria},
   shorttitle = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling},
   author = {Hammerberg, Kris and Brousse, Oscar and Martilli, Alberto and Mahdavi, Ardeshir},
   year = {2018},
@@ -142,8 +142,8 @@
 }
 
 @article{chingWUDAPTUrbanWeather2018a,
-  title = {{{WUDAPT}}: {{An Urban Weather}}, {{Climate}}, and {{Environmental Modeling Infrastructure}} for the {{Anthropocene}}},
-  shorttitle = {{{WUDAPT}}},
+  title = {{WUDAPT}: {An Urban Weather}, {Climate}, and {Environmental Modeling Infrastructure} for the {Anthropocene}},
+  shorttitle = {{WUDAPT}},
   author = {Ching, J. and Mills, G. and Bechtel, B. and See, L. and Feddema, J. and Wang, X. and Ren, C. and Brousse, O. and Martilli, A. and Neophytou, M. and Mouzourides, P. and Stewart, I. and Hanna, A. and Ng, E. and Foley, M. and Alexander, P. and Aliaga, D. and Niyogi, D. and Shreevastava, A. and Bhalachandran, P. and Masson, V. and Hidalgo, J. and Fung, J. and Andrade, M. and Baklanov, A. and Dai, W. and Milcinski, G. and Demuzere, M. and Brunsell, N. and Pesaresi, M. and Miao, S. and Mu, Q. and Chen, F. and Theeuwes, N.},
   year = {2018},
   month = sep,
@@ -157,7 +157,7 @@
 }
 
 @article{cohenCoefficientAgreementNominal1960,
-  title = {A {{Coefficient}} of {{Agreement}} for {{Nominal Scales}}},
+  title = {A {Coefficient} of {Agreement} for {Nominal Scales}},
   author = {Cohen, Jacob},
   year = {1960},
   month = apr,
@@ -172,8 +172,8 @@
 }
 
 @article{demuzereLCZGeneratorWeb2021,
-  title = {{{LCZ Generator}}: {{A Web Application}} to {{Create Local Climate Zone Maps}}},
-  shorttitle = {{{LCZ Generator}}},
+  title = {{LCZ Generator}: {A Web Application} to {Create Local Climate Zone Maps}},
+  shorttitle = {LCZ Generator},
   author = {Demuzere, Matthias and Kittner, Jonas and Bechtel, Benjamin},
   year = {2021},
   month = apr,
@@ -187,30 +187,30 @@
 
 @techreport{bernardGenericAlgorithmAutomatically2023,
   type = {Preprint},
-  title = {A Generic Algorithm to Automatically Classify Urban Fabric According to the {{Local Climate Zone}} System: Implementation in {{GeoClimate}} 0.0.1 and Application to {{French}} Cities},
[…]
