+++ docs/articles/joss/paper.bib
@@ -25,7 +25,7 @@
 }
 
 @article{hammerbergImplicationsEmployingDetailed2018,
-  title = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling: A Comparison between {WUDAPT} and {GIS} Databases over {Vienna}, {Austria},
+  title = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling: A Comparison between {WUDAPT} and {GIS} Databases over {Vienna}, {Austria}},
   shorttitle = {Implications of Employing Detailed Urban Canopy Parameters for Mesoscale Climate Modelling},
   author = {Hammerberg, Kris and Brousse, Oscar and Martilli, Alberto and Mahdavi, Ardeshir},
   year = {2018},
