### 2023-11-15 #5445 @MGousseff
I AM going to party like I just published a paper. 🤘

🙏🙏
I want to thank very warmly @wcjochem and @matthiasdemuzere for their thorough process of reviewing. It is thanks to you that the package is way better than what I submitted in the first place. @matthiasdemuzere I feel very grateful for the time you spent, especially for trying to reinstall the package and running the examples when you were not convinced it could be useful.

🙏🙏
@martinfleis I really appreciated how patient you were and how you guided me in the process. I hope I will be a bit "JOSS-smarter" next time.  The last PR to help me correct bibtex and markdown syntax was a very classy move.

🙏🙏
@oliviaguest Thank you for your comprehension about article length.

🙏🙏
@ebocher Thank you for your trust, I'm so glad I have this chance to be part of this team (not just a polite b
@jeremy-b thanks for the support.

The package will be updated, as there already is a shiny interface in development to launch GeoClimate.
