### 2023-11-19 #5445 @oliviaguest
Huge thanks to the editor: @martinfleis; and the reviewers: @matthiasdemuzere, @wcjochem! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the author: @MGousseff! 🥳 🍾
